% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{apply_selection}
\alias{apply_selection}
\title{Apply a selection to a space}
\usage{
apply_selection(space_id, selection)
}
\arguments{
\item{space_id}{The space_id of the space to which to apply the selection to}

\item{selection}{The selection object of class \code{point_selection} or \code{hyperslab_selection}}
}
\description{
Apply a selection to a space
}
\details{
Calls the respective stand-alone functions for point-selection or multiple hyperslab selection. The reason for not
calling a method of an R6 object is to make it more efficient and make it useable without creating a full R6 object.
}
\author{
Holger Hoefling
}
\keyword{internal}

