% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{are_args_scalar}
\alias{are_args_scalar}
\title{Can arguments be interpreted as a scalar?}
\usage{
are_args_scalar(args)
}
\arguments{
\item{args}{The arguments to check}
}
\value{
Logical if the arguments can be interpreted as a scalar
}
\description{
Can arguments be interpreted as a scalar?
}
\details{
Check if there is only one argument and if it is either empty
of of length 1 with value 1, i.e. can be interpreted as a scalar.
}
\author{
Holger Hoefling
}
\keyword{internal}
