% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{array_counter}
\alias{array_counter}
\title{Cycle through n-dimensional array indices}
\usage{
array_counter(count, dims)
}
\arguments{
\item{count}{The counter (0-based)}

\item{dims}{The sizes of the dimension}
}
\value{
An integer vector of the same length as \code{dim}, with 0-based indices
}
\description{
Cycle through n-dimensional array indices
}
\details{
Cycles through all indices of an n-dimensional array. The first dimension
moves fastest. The counter is 0-based and the output is 0-based as well.
}
\author{
Holger Hoefling
}
\keyword{internal}
