% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{array_reorder}
\alias{array_reorder}
\title{Reorder an array}
\usage{
array_reorder(x, dims, reorder_dim, new_order, item_size)
}
\arguments{
\item{x}{The array; doesn't have to have a dim attribute; is just assumed to be a vector}

\item{dims}{The dimensionality of the array}

\item{reorder_dim}{The dimension to reorder}

\item{new_order}{The new ordering of the reorder_dim; not checked for correctness; 1-based}

\item{item_size}{The size in bytes of each array item; not discovered automatically}
}
\value{
The re-ordered array
}
\description{
Reorder an array
}
\details{
Reorders an array using a fast underlying c-function. It is implemented
for its simple generality and only intended for internal use in the package.
}
\author{
Holger Hoefling
}
\keyword{internal}
