% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_ext.R
\name{as.character.factor_ext}
\alias{as.character.factor_ext}
\title{Coerce a \code{factor_ext} to a character vector}
\usage{
\method{as.character}{factor_ext}(x, ...)
}
\arguments{
\item{x}{The \code{factor_ext} object to coerce}

\item{...}{Currently ignored}
}
\value{
Character vector with the level name for each item of an \code{factor_ext}
}
\description{
Coerce a \code{factor_ext} to a character vector
}
\details{
Coerces an extended factor to a character vector. Each item is replaced with it level name.
}
\author{
Holger Hoefling
}
