% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{as.data.frame.H5R}
\alias{as.data.frame.H5R}
\title{Coerce an \code{H5R} object to a data.frame}
\usage{
\method{as.data.frame}{H5R}(x, row.names = NULL, optional = FALSE, ...,
  nm = paste(deparse(substitute(x), width.cutoff = 500L), collapse = " "))
}
\arguments{
\item{x}{The \code{H5R} object to coerce to a data-frame}

\item{row.names}{\code{NULL} or a character vector giving the row names for the
data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and converting column
names (to syntactic names: see \code{make.names}) is optional.}

\item{...}{additional arguments to be passed to or from methods.}

\item{nm}{The column names to use}

\item{width.cutoff}{ignored}

\item{collapse}{ignored}
}
\value{
A data frame
}
\description{
Coerce an \code{H5R} object to a data.frame
}
\details{
Coerce an \code{H5R} object to a data.frame
}
\author{
Holger Hoefling
}

