% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{as.vector.H5R}
\alias{as.vector.H5R}
\title{Coerce to a vector}
\usage{
\method{as.vector}{H5R}(x, mode = "any")
}
\arguments{
\item{x}{The object to coerce into a vector}

\item{mode}{Only 'any' supported}
}
\value{
The object cloned and coerced to a vector
}
\description{
Coerce to a vector
}
\details{
Coerce an \code{H5R} object to a vector - in practice this just means setting the \code{dim} to \code{NULL}.
}
\author{
Holger Hoefling
}
