% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{as_hex}
\alias{as_hex}
\title{Convert a double or integer to hex}
\usage{
as_hex(x)
}
\arguments{
\item{x}{The integer or double vector to convert}
}
\value{
Character string with the hex value
}
\description{
Convert a double or integer to hex
}
\details{
Converts a double or integer to hex. Contrary to the builtin \code{\link{format}},
this is done without any conversion of integers in double-format to integers in integer format.
}
\author{
Holger Hoefling
}
