% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_ext.R
\name{c.factor_ext}
\alias{c.factor_ext}
\title{Concatenation of \code{factor_ext} objects}
\usage{
\method{c}{factor_ext}(...)
}
\arguments{
\item{...}{Items to concatenate}
}
\value{
A \code{factor_ext}
}
\description{
Concatenation of \code{factor_ext} objects
}
\details{
Concatenation of \code{factor_ext} objects.
}
\author{
Holger Hoefling
}
