% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{c.H5R}
\alias{c.H5R}
\title{Concatenate HDF5-Reference objects}
\usage{
\method{c}{H5R}(..., recursive = FALSE)
}
\arguments{
\item{...}{The objects to concatenate; All have to be references with respect to the same class}

\item{recursive}{Ignored here}
}
\value{
An object of the same class, with all components concatenated
}
\description{
Concatenate HDF5-Reference objects
}
\details{
Concatenates HDF5-Reference objects; also concatenates the names, replacing empty strings
if the names are \code{NULL}.
}
\author{
Holger Hoefling
}

