% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{cbind.H5R}
\alias{cbind.H5R}
\alias{rbind.H5R}
\title{cbind/rbind of HDF5-Reference classes}
\usage{
cbind.H5R(..., deparse.level = 1)

rbind.H5R(..., deparse.level = 1)
}
\arguments{
\item{...}{The objects to put together by row/column}

\item{deparse.level}{Has to be 1; otherwise ignored}
}
\value{
An HDF5-Reference object that is the result of binding the rows/columns of the arrays
}
\description{
cbind/rbind of HDF5-Reference classes
}
\details{
cbind/rbind functionality for HDF5-Reference classes
}
\author{
Holger Hoefling
}

