% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{check_arg_for_hyperslab_func}
\alias{check_arg_for_hyperslab_func}
\title{Check argument for known functions that encode a hyperslab}
\usage{
check_arg_for_hyperslab_func(x, envir)
}
\arguments{
\item{envir}{The environment in which to evaluate the argument}

\item{cur_arg}{The argument to check}
}
\value{
A vector of length 4 describing start, count, stride and block if appropriate
}
\description{
Check argument for known functions that encode a hyperslab
}
\details{
Checks for the functions \code{:}, \code{seq} and \code{seq_len}
}
\author{
Holger Hoefling
}
\keyword{internal}

