% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{clean_ls_df}
\alias{clean_ls_df}
\title{Cleaning result of internal \code{_H5ls} function}
\usage{
clean_ls_df(df)
}
\arguments{
\item{df}{The result of the C-function \code{_H5ls}}
}
\value{
A data frame with content that was not successfully gathered set to \code{NA} and \code{*_success} columns removed
}
\description{
Cleaning result of internal \code{_H5ls} function
}
\details{
For every \code{*_success} item that is \code{FALSE}, the correspoding row of the data.frame will be set to NA.
}
\author{
Holger Hoefling
}

