% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_ext.R
\name{coercible_to_factor}
\alias{coercible_to_factor}
\alias{coerce_to_factor}
\title{Coercing to a factor}
\usage{
coercible_to_factor(x)

coerce_to_factor(x)
}
\arguments{
\item{x}{Object of type \code{factor_ext}}
}
\value{
Either a \code{factor} for \code{coerce_to_factor} or a logical for \code{coercible_to_factor}.
}
\description{
Coercing to a factor
}
\details{
Function to test if a \code{factor_ext} can be coerced into a factor and a function that does the coercion.
A \code{factor_ext} is coercible if its values are consequtive and starting at 1.
}
\author{
Holger Hoefling
}
