% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{dim.H5R}
\alias{dim.H5R}
\alias{dim<-.H5R}
\title{Get/Assign dimension of an object}
\usage{
\method{dim}{H5R}(x)

\method{dim}{H5R}(x) <- value
}
\arguments{
\item{x}{The object to get/assign dimensions}

\item{value}{A vector specifiying the dimensions of the object}
}
\value{
A vector with the dimensions
}
\description{
Get/Assign dimension of an object
}
\details{
Get/Assign the dimensionality of the objects.
}
\author{
Holger Hoefling
}

