% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{dimnames.H5R}
\alias{dimnames.H5R}
\alias{dimnames<-.H5R}
\title{Get/set dimnames}
\usage{
\method{dimnames}{H5R}(x)

\method{dimnames}{H5R}(x) <- value
}
\arguments{
\item{x}{The object to get/set the dimnames of}

\item{value}{The dimnames to set. Has to be a list of character vectors of length equal to the rank of the array}
}
\value{
A list of length equal to the rank of the array, with each element being a character vector
}
\description{
Get/set dimnames
}
\details{
Get/set the dimnames of an \code{H5R} object
}
\author{
Holger Hoefling
}
