% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{do_reshuffle}
\alias{do_reshuffle}
\title{Reshuffle the input as needed - see \code{args_regularity_evaluation}}
\usage{
do_reshuffle(x, reg_eval_res)
}
\arguments{
\item{x}{The array to reshuffle}

\item{reg_eval_res}{The result of the regularity evaluation}
}
\value{
The reshuffled input
}
\description{
Reshuffle the input as needed - see \code{args_regularity_evaluation}
}
\details{
When necessary, this function performs the reshuffle as defined by \code{args_regularity_evaluation}.
}
\author{
Holger Hoefling
}
\keyword{internal}

