% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{equal_id_check}
\alias{equal_id_check}
\title{Compare the ids of objects}
\usage{
equal_id_check(...)
}
\arguments{
\item{...}{Any \code{H5RefClass} objects}
}
\value{
Logical - are all ids the same of the objects passed in
}
\description{
Compare the ids of objects
}
\details{
When several \code{H5RefClass} objects are passed in
}
\author{
Holger Hoefling
}
\keyword{internal}
