% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{expand_point_grid}
\alias{expand_point_grid}
\title{Expand list of points for each dimension into a matrix of all combinations}
\usage{
expand_point_grid(point_list)
}
\arguments{
\item{point_list}{A list of the points in each dimension to include}
}
\value{
A matrix with every combination of points for each dimension
}
\description{
Expand list of points for each dimension into a matrix of all combinations
}
\details{
The function is similar to the \code{expand.grid} function
}
\author{
Holger Hoefling
}
\keyword{internal}
