% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{extract_dim}
\alias{extract_dim}
\title{Set the correct dimension attribute for an object}
\usage{
extract_dim(space, dtype)
}
\arguments{
\item{space}{The space with the selection that was used to read the dataset}

\item{dtype}{The datatype of the dataset}

\item{x}{an object that was read from an HDF5 dataset or attribute}
}
\value{
\code{x}, but with a new dimension attribute
}
\description{
Set the correct dimension attribute for an object
}
\details{
This function uses the space and the selection in it to set the correct dimension
for the resulting object (but without dropping dimensions). Furhtermore, if the
datatype is an array, those dimensions are correctly attached as well.

Internal use only; not exported
}
\author{
Holger Hoefling
}
\keyword{internal}

