% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_ext.R
\name{factor_ext_eq}
\alias{factor_ext_eq}
\alias{==.factor_ext}
\alias{!=.factor_ext}
\title{Compare two \code{factor_ext} for equality}
\usage{
\method{==}{factor_ext}(e1, e2)

\method{!=}{factor_ext}(e1, e2)
}
\arguments{
\item{e1}{First item in the comparison}

\item{e2}{Second item in the comparison}
}
\value{
A logical
}
\description{
Compare two \code{factor_ext} for equality
}
\details{
Compares two \code{factor_ext} for equality. In order for them to be considered as equal,
they have to be equal after being coerced to character vectors. I.e. the can have different level values,
as long as the level names are the same.
}
\author{
Holger Hoefling
}
