% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_ext.R
\name{factor_ext_sub}
\alias{[.factor_ext}
\alias{[<-.factor_ext}
\alias{[[.factor_ext}
\alias{[[<-.factor_ext}
\alias{factor_ext_sub}
\title{Subset or assign into a subset of a \code{factor_ext} object}
\usage{
\method{[[}{factor_ext}(x, ...)

\method{[[}{factor_ext}(x, ...) <- value

\method{[}{factor_ext}(x, ..., drop = FALSE)

\method{[}{factor_ext}(x, ...) <- value
}
\arguments{
\item{x}{\code{factor_ext} object to subset or to assign into}

\item{...}{Currently ignored}

\item{value}{The object to assign}

\item{drop}{Should dimensions of size 1 be dropped?}
}
\value{
Returns single item of a \code{factor_ext} object
}
\description{
Subset or assign into a subset of a \code{factor_ext} object
}
\details{
Subset and assignment into an extended factor object
}
\author{
Holger Hoefling
}

