% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{flatten_df}
\alias{flatten_df}
\title{Flatten a nested data.frame}
\usage{
flatten_df(df, factor_ext_to_char = FALSE)
}
\arguments{
\item{df}{The data.frame to flatten}

\item{factor_ext_to_char}{Should extended factor variables be converted to characters (mainly for easy printing)}
}
\value{
A flattened \code{\link{data.frame}}
}
\description{
Flatten a nested data.frame
}
\details{
HDF5 Compounds allow for nesting. Correspondingly, nested data.frames are being produced.
This function flattens such a nested data.frame.

For easier printing to the screen, it also allows for coercion of \code{\link{factor_ext}} to
character variables.
}
\author{
Holger Hoefling
}

