% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{format.H5R}
\alias{format.H5R}
\title{Formatting of an H5R object}
\usage{
\method{format}{H5R}(x, ...)
}
\arguments{
\item{x}{The object to format}

\item{...}{ignored}
}
\value{
Character vector with the class names in angle-brackets
}
\description{
Formatting of an H5R object
}
\details{
Formatting of H5R objects
}
\author{
Holger Hoefling
}

