% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5D.R
\name{get_id}
\alias{get_id}
\title{Get the id of an H5RefClass}
\usage{
get_id(obj)
}
\arguments{
\item{obj}{Object to get the id from}
}
\value{
The id itself
}
\description{
Get the id of an H5RefClass
}
\details{
If it is a H5RefClass, returns the id, otherwise returns the
object itself as it assumes it is already an id.
}
\author{
Holger Hoefling
}
\keyword{internal}
