% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{get_robj_dim}
\alias{get_robj_dim}
\title{Get the size of the resulting R object}
\usage{
get_robj_dim(reg_eval_res, dtype)
}
\arguments{
\item{reg_eval_res}{The result of the regularity evaluation}

\item{dtype}{The datatype under consideration}
}
\value{
A list with the dimensions of the resulting object, pre- and post shuffle
}
\description{
Get the size of the resulting R object
}
\details{
For normal objects, just uses the size of the indices in the request, and evaluates
them bost pre- and post-shuffle. If the internal object is an array, additional dimensions
are appeded at the end.
}
\author{
Holger Hoefling
}
\keyword{internal}

