% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{guess_space}
\alias{guess_space}
\title{Guess the dataspace of an object}
\usage{
guess_space(x, dtype, chunked = TRUE)
}
\arguments{
\item{x}{The R object for which to guess the space}

\item{dtype}{Object of type \code{\link{H5T}}, that represents that datatype to use.}

\item{chunked}{Is the datatype chunked? If yes, \code{maxdims} of the space will be set to infinity,
otherwise \code{maxdims} will be set to the original extent of the space.}
}
\value{
An object of type \code{\link{H5S}}
}
\description{
Guess the dataspace of an object
}
\details{
Creates a dataspace that fits an R object so that it can be written into a dataset. This is used
for example in dataset creation based on an R-object, not a specifically defined dimensions.
}
\author{
Holger Hoefling
}
