% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5wrapper.R
\name{h5file}
\alias{h5file}
\alias{createGroup}
\alias{openLocation}
\alias{createDataSet}
\alias{readDataSet}
\alias{h5close}
\alias{h5flush}
\alias{existsGroup}
\alias{is.h5file}
\alias{list.groups}
\alias{list.datasets}
\alias{extendDataSet}
\alias{rbind.H5D}
\alias{cbind.H5D}
\alias{c.H5D}
\alias{h5unlink}
\title{h5 wrapper functions}
\usage{
h5file(...)

createGroup(object, name, ...)

openLocation(object, name)

createDataSet(object, name, ...)

readDataSet(object)

h5close(object)

h5flush(object)

existsGroup(object, name)

is.h5file(name)

list.groups(object, path = "/", recursive = TRUE, ...)

list.datasets(object, path = "/", recursive = TRUE, ...)

extendDataSet(object, dims)

\method{rbind}{H5D}(x, mat)

\method{cbind}{H5D}(x, mat)

\method{c}{H5D}(x, ...)

h5unlink(object, name)
}
\arguments{
\item{...}{Additional Parameters passed to \code{file$create_group}, \code{file$create_dataset}}

\item{object}{H5RefClass; H5 Reference Class to be used}

\item{name}{character; Group/Filename to be created}

\item{path}{character; Path named to be used for iteration.}

\item{recursive}{logical; Specify if object should be traversed recursively.}

\item{dims}{numeric; Dimension vector to which dataset should be extended.}
}
\description{
Wrapper functions to provide h5 compatible interface.
}
