% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5File.R
\name{H5_close_any}
\alias{H5_close_any}
\title{Closes any HDF5 id using the appropriate library function}
\usage{
H5_close_any(id)
}
\arguments{
\item{id}{The id to closes}
}
\value{
NULL, invisibly
}
\description{
Closes any HDF5 id using the appropriate library function
}
\details{
Internal function to help with management of open ids. It is used to close
an id that is no longer needed using the appropriate HDF5 library function.
}
\author{
Holger Hoefling
}

