% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5A.R
\docType{class}
\name{H5A}
\alias{H5A}
\alias{H5A-class}
\title{Class for representing HDF5 attributes}
\value{
Object of class \code{\link{H5A}}.
}
\description{
This class represents an HDF5 attribute. Usually it is easier to read and write attributes for
groups, datasets and committed datatypes using the functions documented in \code{\link{h5attributes}}.

Otherwise, the functionality for attributes is very similar to that of datasets (\code{\link{H5D}}),
however with the notable exception that attributes always have to be read and written as a whole.
}
\examples{
fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
h5attr(file, "attr_numeric") <- rnorm(10)
a <- file$attr_open("attr_numeric")
a$get_info()
a$attr_name()
a$get_space()
a$get_type()
a$get_storage_size()
a$read()
a$write(10:1)
a$print()
a$close()
file$close_all()
}
\author{
Holger Hoefling
}
\section{Super class}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{H5A}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5A-flush}{\code{H5A$flush()}}
    \item \href{#method-H5A-get_filename}{\code{H5A$get_filename()}}
    \item \href{#method-H5A-get_info}{\code{H5A$get_info()}}
    \item \href{#method-H5A-attr_name}{\code{H5A$attr_name()}}
    \item \href{#method-H5A-get_space}{\code{H5A$get_space()}}
    \item \href{#method-H5A-get_type}{\code{H5A$get_type()}}
    \item \href{#method-H5A-get_storage_size}{\code{H5A$get_storage_size()}}
    \item \href{#method-H5A-read_low_level}{\code{H5A$read_low_level()}}
    \item \href{#method-H5A-read}{\code{H5A$read()}}
    \item \href{#method-H5A-write_low_level}{\code{H5A$write_low_level()}}
    \item \href{#method-H5A-write}{\code{H5A$write()}}
    \item \href{#method-H5A-print}{\code{H5A$print()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="initialize"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-initialize'><code>H5RefClass$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-flush"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-flush}{}}}
\subsection{\code{H5A$flush()}}{
  This function implements the HDF5-API function H5Fflush. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$flush(scope = h5const$H5F_SCOPE_GLOBAL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{scope}}{The scope of the flush; one of the \code{H5F_SCOPE_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-get_filename"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-get_filename}{}}}
\subsection{\code{H5A$get_filename()}}{
  This function implements the HDF5-API function H5Fget_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$get_filename()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-get_info}{}}}
\subsection{\code{H5A$get_info()}}{
  This function implements the HDF5-API function H5Aget_info. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$get_info()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-attr_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-attr_name}{}}}
\subsection{\code{H5A$attr_name()}}{
  This function implements the HDF5-API function H5Aget_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$attr_name()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-get_space"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-get_space}{}}}
\subsection{\code{H5A$get_space()}}{
  This function implements the HDF5-API function H5Aget_space. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$get_space()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-get_type}{}}}
\subsection{\code{H5A$get_type()}}{
  This function implements the HDF5-API function H5Aget_type. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$get_type(native = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{native}}{Logical; should the native version of the datatype be returned.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-get_storage_size"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-get_storage_size}{}}}
\subsection{\code{H5A$get_storage_size()}}{
  This function implements the HDF5-API function H5Aget_storage_size. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$get_storage_size()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-read_low_level"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-read_low_level}{}}}
\subsection{\code{H5A$read_low_level()}}{
  Only for advanced users. See documentation for \code{read} instead. This function implements the HDF5-API function H5Aread. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$read_low_level(buffer, mem_type, duplicate_buffer = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buffer}}{The raw buffer to read into or write from.}
      \item{\code{mem_type}}{The memory datatype to use for the transfer; an object of class \code{\link{H5T}}.}
      \item{\code{duplicate_buffer}}{Logical; should the buffer be copied before it is returned.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-read"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-read}{}}}
\subsection{\code{H5A$read()}}{
  Reads the data of the attribute and returns it as an R-object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$read(flags = getOption("hdf5r.h5tor_default"), drop = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{flags}}{Conversion rules for integer values. See also \code{\link{h5const}}}
      \item{\code{drop}}{Logical. Should dimensions of length 1 be dropped (R-default for arrays)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-write_low_level"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-write_low_level}{}}}
\subsection{\code{H5A$write_low_level()}}{
  Only for advanced users. See documentation for \code{write} instead. This function implements the HDF5-API function H5Awrite. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$write_low_level(buffer, mem_type)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buffer}}{The raw buffer to read into or write from.}
      \item{\code{mem_type}}{The memory datatype to use for the transfer; an object of class \code{\link{H5T}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-write"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-write}{}}}
\subsection{\code{H5A$write()}}{
  Writes the data of \code{robj} to the attribute
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$write(robj, mem_type = NULL, flush = getOption("hdf5r.flush_on_write"))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{robj}}{The object to write into the attribute}
      \item{\code{mem_type}}{The memory data type to use when transferring from HDF5 to intermediate storage. This is an advanced development feature and may be removed in the future.}
      \item{\code{flush}}{Logical; should the file be flushed after the write.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5A-print"></a>}}
\if{latex}{\out{\hypertarget{method-H5A-print}{}}}
\subsection{\code{H5A$print()}}{
  Prints information for the dataset
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5A$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
    }
    \if{html}{\out{</div>}}
  }
}

}
