% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{H5Class_overview}
\alias{H5Class_overview}
\title{Call the HDF5 class overview}
\usage{
H5Class_overview(browser = getOption("browser"))
}
\arguments{
\item{browser}{The browser to use}
}
\value{
Invisibly, the name of the overivew html-page
}
\description{
Call the HDF5 class overview
}
\details{
Brings up a searchable table that lists the methods available in the R6 classes
}
\author{
Holger Hoefling
}

