% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{h5garbage_collect}
\alias{h5garbage_collect}
\title{Trigger the HDF5 garbage collection}
\usage{
h5garbage_collect()
}
\value{
Invisible  \code{NULL}
}
\description{
Trigger the HDF5 garbage collection
}
\details{
This function triggers the HDF5 internal garbage collection. It is independent of the
R garbage collection and currently has to be triggered by hand.
}
\author{
Holger Hoefling
}
