% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5Group.R
\name{H5GTD_factory}
\alias{H5GTD_factory}
\title{Wrap an HDF5-id in the appropriate class}
\usage{
H5GTD_factory(id)
}
\arguments{
\item{id}{The id to wrap in an R6-class object}
}
\value{
An R6-class object corresponding to the HDF5 internal class of the ID
}
\description{
Wrap an HDF5-id in the appropriate class
}
\details{
This particular factory dispatches ids that can be of type Group, Datatype of dataset (these are
id types that can be the result of opening an object. For datatypes, the \code{\link{H5T_factory}} exists that can be used
}
\author{
Holger Hoefling
}

