% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P-class}
\alias{H5P-class}
\alias{H5P}
\title{Class for HDF5 property lists.}
\value{
Object of class \code{\link{H5P}}.
}
\description{
This is the base class for all property lists, but most have a specialised class.
It inherits all functions of the
\code{\link{H5RefClass-class}}. It is also the base class for many other classes well, specifically
\describe{
\item{Dataset Creation}{\code{\link{H5P_DATASET_CREATE-class}}}
\item{Dataset Access}{\code{\link{H5P_DATASET_ACCESS-class}}}
\item{Dataset Transfer}{\code{\link{H5P_DATASET_XFER-class}}}
\item{Link Creation}{\code{\link{H5P_LINK_CREATE-class}}}
\item{Link Access}{\code{\link{H5P_LINK_ACCESS-class}}}
\item{Object Creation}{\code{\link{H5P_OBJECT_CREATE-class}}}
\item{Object Copy}{\code{\link{H5P_OBJECT_COPY-class}}}
\item{Attribute Creation}{\code{\link{H5P_ATTRIBUTE_CREATE-class}}}
}
}
\section{Methods}{

\describe{
\item{\code{new(id = NULL)}}{

Create a new property list; this function itself is unlikely to be needed by users. Users should
use the classes of the type they actually require

\strong{Parameters:}
\describe{
\item{id}{Internal use only}
}}

\item{\code{get_class()}}{

This function implements the HDF5-API function H5Pget_class.
Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-GetClass} for details.}

\item{\code{get_class_name()}}{

This function implements the HDF5-API function H5Pget_class_name.
Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-GetClassName} for details.}

\item{\code{copy()}}{

This function implements the HDF5-API function H5Pcopy.
Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-Copy} for details.}

\item{\code{equal(cmp)}}{

This function implements the HDF5-API function H5Pequal.
Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-Equal} for details.}
}}

\seealso{
H5Class_overview
}
\author{
Holger Hoefling
}
