% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P}
\alias{H5P}
\alias{H5P-class}
\title{Class for HDF5 property lists.}
\value{
Object of class \code{\link{H5P}}.
}
\description{
This is the base class for all property lists, but most have a specialized class.
It inherits all functions of the
\code{\link{H5RefClass}}. It is also the base class for many other classes, specifically
\describe{
\item{Dataset Creation}{\code{\link{H5P_DATASET_CREATE}}}
\item{Dataset Access}{\code{\link{H5P_DATASET_ACCESS}}}
\item{Dataset Transfer}{\code{\link{H5P_DATASET_XFER}}}
\item{Link Creation}{\code{\link{H5P_LINK_CREATE}}}
\item{Link Access}{\code{\link{H5P_LINK_ACCESS}}}
\item{Object Creation}{\code{\link{H5P_OBJECT_CREATE}}}
\item{Object Copy}{\code{\link{H5P_OBJECT_COPY}}}
\item{Attribute Creation}{\code{\link{H5P_ATTRIBUTE_CREATE}}}
}
The base class is unlikely to be needed by users - they should use the appropriate subclass required.
}
\examples{
fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
file[["testdataset"]] <- 1:10
p <- file[["testdataset"]]$get_create_plist()
p$get_class()
p$get_class_name()
p$copy()
p$equal(p)
file$close_all()
}
\author{
Holger Hoefling
}
\section{Super class}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{H5P}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5P-initialize}{\code{H5P$new()}}
    \item \href{#method-H5P-get_class}{\code{H5P$get_class()}}
    \item \href{#method-H5P-get_class_name}{\code{H5P$get_class_name()}}
    \item \href{#method-H5P-copy}{\code{H5P$copy()}}
    \item \href{#method-H5P-equal}{\code{H5P$equal()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5P-initialize}{}}}
\subsection{\code{H5P$new()}}{
  Create a new property list; this function itself is unlikely to be needed by users. Users should use the classes of the type they actually require
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P-get_class"></a>}}
\if{latex}{\out{\hypertarget{method-H5P-get_class}{}}}
\subsection{\code{H5P$get_class()}}{
  This function implements the HDF5-API function H5Pget_class. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P$get_class()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P-get_class_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5P-get_class_name}{}}}
\subsection{\code{H5P$get_class_name()}}{
  This function implements the HDF5-API function H5Pget_class_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P$get_class_name()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P-copy"></a>}}
\if{latex}{\out{\hypertarget{method-H5P-copy}{}}}
\subsection{\code{H5P$copy()}}{
  This function implements the HDF5-API function H5Pcopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P$copy()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P-equal"></a>}}
\if{latex}{\out{\hypertarget{method-H5P-equal}{}}}
\subsection{\code{H5P$equal()}}{
  This function implements the HDF5-API function H5Pequal. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P$equal(cmp)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{cmp}}{The object to compare against.}
    }
    \if{html}{\out{</div>}}
  }
}

}
