% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_ATTRIBUTE_CREATE}
\alias{H5P_ATTRIBUTE_CREATE}
\alias{H5P_ATTRIBUTE_CREATE-class}
\title{Class for HDF5 property list for attribute creation}
\value{
Object of class \code{\link[=H5P_ATTRIBUTE_CREATE]{H5P_ATTRIBUTE_CREATE}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5P]{H5P}} -> \code{H5P_ATTRIBUTE_CREATE}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5P_ATTRIBUTE_CREATE-initialize}{\code{H5P_ATTRIBUTE_CREATE$new()}}
    \item \href{#method-H5P_ATTRIBUTE_CREATE-set_char_encoding}{\code{H5P_ATTRIBUTE_CREATE$set_char_encoding()}}
    \item \href{#method-H5P_ATTRIBUTE_CREATE-get_char_encoding}{\code{H5P_ATTRIBUTE_CREATE$get_char_encoding()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="copy"><a href='../../hdf5r/html/H5P.html#method-H5P-copy'><code>H5P$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="equal"><a href='../../hdf5r/html/H5P.html#method-H5P-equal'><code>H5P$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class'><code>H5P$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class_name"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class_name'><code>H5P$get_class_name()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_ATTRIBUTE_CREATE-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_ATTRIBUTE_CREATE-initialize}{}}}
\subsection{\code{H5P_ATTRIBUTE_CREATE$new()}}{
  Create a new class of type \code{\link{H5P_ATTRIBUTE_CREATE}}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_ATTRIBUTE_CREATE$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_ATTRIBUTE_CREATE-set_char_encoding"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_ATTRIBUTE_CREATE-set_char_encoding}{}}}
\subsection{\code{H5P_ATTRIBUTE_CREATE$set_char_encoding()}}{
  This function implements the HDF5-API function H5Pset_char_encoding. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_ATTRIBUTE_CREATE$set_char_encoding(encoding = h5const$H5T_CSET_UTF8)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{encoding}}{The character encoding to use; one of the \code{H5T_CSET_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_ATTRIBUTE_CREATE-get_char_encoding"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_ATTRIBUTE_CREATE-get_char_encoding}{}}}
\subsection{\code{H5P_ATTRIBUTE_CREATE$get_char_encoding()}}{
  This function implements the HDF5-API function H5Pget_char_encoding. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_ATTRIBUTE_CREATE$get_char_encoding()}
    \if{html}{\out{</div>}}
  }
}

}
