% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_CLASS}
\alias{H5P_CLASS}
\alias{H5P_CLASS-class}
\title{Class for HDF5 property list classes (not HDF5 property lists)}
\value{
Object of class \code{\link[=H5P_CLASS]{H5P_CLASS}}.
}
\description{
It inherits all functions of the \code{\link{H5RefClass}}. The intent of this class is to
provide a mechanism to compare the class of HDF5 property classes. This is mainly intended for internal use
to get the class type of an HDF5 identifier that is known to be a property list, but not of which type.
}
\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
\section{Super class}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{H5P_CLASS}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5P_CLASS-equal}{\code{H5P_CLASS$equal()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="initialize"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-initialize'><code>H5RefClass$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_CLASS-equal"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_CLASS-equal}{}}}
\subsection{\code{H5P_CLASS$equal()}}{
  This function implements the HDF5-API function H5Pequal. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_CLASS$equal(cmp)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{cmp}}{The object to compare against.}
    }
    \if{html}{\out{</div>}}
  }
}

}
