% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_DATASET_CREATE}
\alias{H5P_DATASET_CREATE}
\title{Class for HDF5 property list for dataset creation}
\value{
Object of class \code{\link[=H5P_DATASET_CREATE-class]{H5P_DATASET_CREATE}}.
}
\description{
It inherits all functions of the \code{\link{H5P-class}}.
}
\seealso{
H5Class_overview, \code{\link[=H5P-class]{H5P}}
}
\author{
Holger Hoefling
}
