% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_DATASET_CREATE}
\alias{H5P_DATASET_CREATE}
\alias{H5P_DATASET_CREATE-class}
\title{Class for HDF5 property list for dataset creation}
\value{
Object of class \code{\link[=H5P_DATASET_CREATE]{H5P_DATASET_CREATE}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5P]{H5P}} -> \code{H5P_DATASET_CREATE}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5P_DATASET_CREATE-initialize}{\code{H5P_DATASET_CREATE$new()}}
    \item \href{#method-H5P_DATASET_CREATE-set_layout}{\code{H5P_DATASET_CREATE$set_layout()}}
    \item \href{#method-H5P_DATASET_CREATE-get_layout}{\code{H5P_DATASET_CREATE$get_layout()}}
    \item \href{#method-H5P_DATASET_CREATE-set_chunk}{\code{H5P_DATASET_CREATE$set_chunk()}}
    \item \href{#method-H5P_DATASET_CREATE-get_chunk}{\code{H5P_DATASET_CREATE$get_chunk()}}
    \item \href{#method-H5P_DATASET_CREATE-set_deflate}{\code{H5P_DATASET_CREATE$set_deflate()}}
    \item \href{#method-H5P_DATASET_CREATE-set_fill_value}{\code{H5P_DATASET_CREATE$set_fill_value()}}
    \item \href{#method-H5P_DATASET_CREATE-get_fill_value}{\code{H5P_DATASET_CREATE$get_fill_value()}}
    \item \href{#method-H5P_DATASET_CREATE-set_fill_time}{\code{H5P_DATASET_CREATE$set_fill_time()}}
    \item \href{#method-H5P_DATASET_CREATE-get_fill_time}{\code{H5P_DATASET_CREATE$get_fill_time()}}
    \item \href{#method-H5P_DATASET_CREATE-set_alloc_time}{\code{H5P_DATASET_CREATE$set_alloc_time()}}
    \item \href{#method-H5P_DATASET_CREATE-get_alloc_time}{\code{H5P_DATASET_CREATE$get_alloc_time()}}
    \item \href{#method-H5P_DATASET_CREATE-set_filter}{\code{H5P_DATASET_CREATE$set_filter()}}
    \item \href{#method-H5P_DATASET_CREATE-all_filters_avail}{\code{H5P_DATASET_CREATE$all_filters_avail()}}
    \item \href{#method-H5P_DATASET_CREATE-get_nfilters}{\code{H5P_DATASET_CREATE$get_nfilters()}}
    \item \href{#method-H5P_DATASET_CREATE-get_filter}{\code{H5P_DATASET_CREATE$get_filter()}}
    \item \href{#method-H5P_DATASET_CREATE-modify_filter}{\code{H5P_DATASET_CREATE$modify_filter()}}
    \item \href{#method-H5P_DATASET_CREATE-remove_filter}{\code{H5P_DATASET_CREATE$remove_filter()}}
    \item \href{#method-H5P_DATASET_CREATE-set_fletcher32}{\code{H5P_DATASET_CREATE$set_fletcher32()}}
    \item \href{#method-H5P_DATASET_CREATE-set_nbit}{\code{H5P_DATASET_CREATE$set_nbit()}}
    \item \href{#method-H5P_DATASET_CREATE-set_scaleoffset}{\code{H5P_DATASET_CREATE$set_scaleoffset()}}
    \item \href{#method-H5P_DATASET_CREATE-set_shuffle}{\code{H5P_DATASET_CREATE$set_shuffle()}}
    \item \href{#method-H5P_DATASET_CREATE-set_szip}{\code{H5P_DATASET_CREATE$set_szip()}}
    \item \href{#method-H5P_DATASET_CREATE-set_external}{\code{H5P_DATASET_CREATE$set_external()}}
    \item \href{#method-H5P_DATASET_CREATE-get_external_count}{\code{H5P_DATASET_CREATE$get_external_count()}}
    \item \href{#method-H5P_DATASET_CREATE-get_external}{\code{H5P_DATASET_CREATE$get_external()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="copy"><a href='../../hdf5r/html/H5P.html#method-H5P-copy'><code>H5P$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="equal"><a href='../../hdf5r/html/H5P.html#method-H5P-equal'><code>H5P$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class'><code>H5P$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class_name"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class_name'><code>H5P$get_class_name()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-initialize}{}}}
\subsection{\code{H5P_DATASET_CREATE$new()}}{
  Create a new class of type \code{\link{H5P_DATASET_CREATE}}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_layout"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_layout}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_layout()}}{
  This function implements the HDF5-API function H5Pset_layout. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_layout(layout = h5const$H5D_CHUNKED)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{layout}}{The dataset layout; one of the \code{H5D_*} layout values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-get_layout"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-get_layout}{}}}
\subsection{\code{H5P_DATASET_CREATE$get_layout()}}{
  This function implements the HDF5-API function H5Pget_layout. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$get_layout()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_chunk"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_chunk}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_chunk()}}{
  This function implements the HDF5-API function H5Pset_chunk. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_chunk(chunk)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{chunk}}{The chunk dimensions.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-get_chunk"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-get_chunk}{}}}
\subsection{\code{H5P_DATASET_CREATE$get_chunk()}}{
  This function implements the HDF5-API function H5Pget_chunk. If the layout is not chunked, returns NA. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$get_chunk(max_ndims)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{max_ndims}}{The maximum number of dimensions to return.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_deflate"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_deflate}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_deflate()}}{
  This function implements the HDF5-API function H5Pset_deflate. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_deflate(level)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{level}}{The compression level, between 0 and 9.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_fill_value"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_fill_value}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_fill_value()}}{
  This function implements the HDF5-API function H5Pset_fill_value. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_fill_value(dtype, value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{value}}{The value to assign.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-get_fill_value"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-get_fill_value}{}}}
\subsection{\code{H5P_DATASET_CREATE$get_fill_value()}}{
  This function implements the HDF5-API function H5Pget_fill_value. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$get_fill_value(dtype)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_fill_time"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_fill_time}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_fill_time()}}{
  This function implements the HDF5-API function H5Pset_fill_time. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_fill_time(fill_time = h5const$H5D_FILL_TIME_IFSET)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{fill_time}}{When the fill value is written; one of the \code{H5D_FILL_TIME_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-get_fill_time"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-get_fill_time}{}}}
\subsection{\code{H5P_DATASET_CREATE$get_fill_time()}}{
  This function implements the HDF5-API function H5Pget_fill_time. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$get_fill_time()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_alloc_time"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_alloc_time}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_alloc_time()}}{
  This function implements the HDF5-API function H5Pset_alloc_time. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_alloc_time(alloc_time = h5const$H5D_ALLOC_TIME_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{alloc_time}}{When storage is allocated; one of the \code{H5D_ALLOC_TIME_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-get_alloc_time"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-get_alloc_time}{}}}
\subsection{\code{H5P_DATASET_CREATE$get_alloc_time()}}{
  This function implements the HDF5-API function H5Pget_alloc_time. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$get_alloc_time()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_filter"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_filter}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_filter()}}{
  This function implements the HDF5-API function H5Pset_filter. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_filter(
  filter = h5const$H5Z_FILTER_DEFLATE,
  flags = h5const$H5Z_FLAG_OPTIONAL,
  cd_values = integer(0)
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{filter}}{The filter to use; one of the \code{H5Z_FILTER_*} values in \code{\link{h5const}}.}
      \item{\code{flags}}{Flags governing the conversion from HDF5 to R. See the \code{H5TOR_*} values in \code{\link{h5const}}.}
      \item{\code{cd_values}}{Auxiliary data passed to the filter.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-all_filters_avail"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-all_filters_avail}{}}}
\subsection{\code{H5P_DATASET_CREATE$all_filters_avail()}}{
  This function implements the HDF5-API function H5Pall_filters_avail. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$all_filters_avail()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-get_nfilters"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-get_nfilters}{}}}
\subsection{\code{H5P_DATASET_CREATE$get_nfilters()}}{
  This function implements the HDF5-API function H5Pget_nfilters. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$get_nfilters()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-get_filter"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-get_filter}{}}}
\subsection{\code{H5P_DATASET_CREATE$get_filter()}}{
  This function implements the HDF5-API function H5Pget_filter2. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$get_filter(idx)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{idx}}{The index of the item to access; zero-based.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-modify_filter"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-modify_filter}{}}}
\subsection{\code{H5P_DATASET_CREATE$modify_filter()}}{
  This function implements the HDF5-API function H5Pmodify_filter. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$modify_filter(
  filter = h5const$H5Z_FILTER_DEFLATE,
  flags = h5const$H5Z_FLAG_OPTIONAL,
  cd_values = integer(0)
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{filter}}{The filter to use; one of the \code{H5Z_FILTER_*} values in \code{\link{h5const}}.}
      \item{\code{flags}}{Flags governing the conversion from HDF5 to R. See the \code{H5TOR_*} values in \code{\link{h5const}}.}
      \item{\code{cd_values}}{Auxiliary data passed to the filter.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-remove_filter"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-remove_filter}{}}}
\subsection{\code{H5P_DATASET_CREATE$remove_filter()}}{
  This function implements the HDF5-API function H5Premove_filter. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$remove_filter(filter = h5const$H5Z_FILTER_ALL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{filter}}{The filter to use; one of the \code{H5Z_FILTER_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_fletcher32"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_fletcher32}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_fletcher32()}}{
  This function implements the HDF5-API function H5Pset_fletcher32. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_fletcher32()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_nbit"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_nbit}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_nbit()}}{
  This function implements the HDF5-API function H5Pset_nbit. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_nbit()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_scaleoffset"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_scaleoffset}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_scaleoffset()}}{
  This function implements the HDF5-API function H5Pset_scaleoffset. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_scaleoffset(
  scale_type = h5const$H5Z_SO_FLOAT_DSCALE,
  scale_factor = 0
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{scale_type}}{The scale-offset filter type; one of the \code{H5Z_SO_*} values in \code{\link{h5const}}.}
      \item{\code{scale_factor}}{The scale factor used by the scale-offset filter.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_shuffle"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_shuffle}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_shuffle()}}{
  This function implements the HDF5-API function H5Pset_shuffle. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_shuffle()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_szip"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_szip}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_szip()}}{
  This function implements the HDF5-API function H5Pset_szip. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_szip()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-set_external"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-set_external}{}}}
\subsection{\code{H5P_DATASET_CREATE$set_external()}}{
  This function implements the HDF5-API function H5Pset_external. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$set_external(filename, offset, size)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{filename}}{The name of the file.}
      \item{\code{offset}}{The offset to apply; see the corresponding HDF5 function for how it is interpreted.}
      \item{\code{size}}{The size of the object in bytes; use \code{Inf} for variable-length strings.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-get_external_count"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-get_external_count}{}}}
\subsection{\code{H5P_DATASET_CREATE$get_external_count()}}{
  This function implements the HDF5-API function H5Pget_external_count. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$get_external_count()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_CREATE-get_external"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_CREATE-get_external}{}}}
\subsection{\code{H5P_DATASET_CREATE$get_external()}}{
  This function implements the HDF5-API function H5Pget_external. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_CREATE$get_external(idx)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{idx}}{The index of the item to access; zero-based.}
    }
    \if{html}{\out{</div>}}
  }
}

}
