% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_DATASET_XFER}
\alias{H5P_DATASET_XFER}
\alias{H5P_DATASET_XFER-class}
\title{Class for HDF5 property list for dataset transfer}
\value{
Object of class \code{\link[=H5P_DATASET_XFER]{H5P_DATASET_XFER}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5P]{H5P}} -> \code{H5P_DATASET_XFER}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5P_DATASET_XFER-initialize}{\code{H5P_DATASET_XFER$new()}}
    \item \href{#method-H5P_DATASET_XFER-set_buffer}{\code{H5P_DATASET_XFER$set_buffer()}}
    \item \href{#method-H5P_DATASET_XFER-set_edc_check}{\code{H5P_DATASET_XFER$set_edc_check()}}
    \item \href{#method-H5P_DATASET_XFER-get_edc_check}{\code{H5P_DATASET_XFER$get_edc_check()}}
    \item \href{#method-H5P_DATASET_XFER-set_hyper_vector_size}{\code{H5P_DATASET_XFER$set_hyper_vector_size()}}
    \item \href{#method-H5P_DATASET_XFER-get_hyper_vector_size}{\code{H5P_DATASET_XFER$get_hyper_vector_size()}}
    \item \href{#method-H5P_DATASET_XFER-set_btree_ratios}{\code{H5P_DATASET_XFER$set_btree_ratios()}}
    \item \href{#method-H5P_DATASET_XFER-get_btree_ratios}{\code{H5P_DATASET_XFER$get_btree_ratios()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="copy"><a href='../../hdf5r/html/H5P.html#method-H5P-copy'><code>H5P$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="equal"><a href='../../hdf5r/html/H5P.html#method-H5P-equal'><code>H5P$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class'><code>H5P$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class_name"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class_name'><code>H5P$get_class_name()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_XFER-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_XFER-initialize}{}}}
\subsection{\code{H5P_DATASET_XFER$new()}}{
  Create a new class of type \code{\link{H5P_DATASET_XFER}}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_XFER$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_XFER-set_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_XFER-set_buffer}{}}}
\subsection{\code{H5P_DATASET_XFER$set_buffer()}}{
  This function implements the HDF5-API function H5Pset_buffer. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_XFER$set_buffer(size = 2^20)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{size}}{The size of the object in bytes; use \code{Inf} for variable-length strings.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_XFER-set_edc_check"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_XFER-set_edc_check}{}}}
\subsection{\code{H5P_DATASET_XFER$set_edc_check()}}{
  This function implements the HDF5-API function H5Pset_edc_check. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_XFER$set_edc_check(check = h5const$H5Z_ENABLE_EDC)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{check}}{Logical; should the arguments be checked for validity.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_XFER-get_edc_check"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_XFER-get_edc_check}{}}}
\subsection{\code{H5P_DATASET_XFER$get_edc_check()}}{
  This function implements the HDF5-API function H5Pget_edc_check. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_XFER$get_edc_check()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_XFER-set_hyper_vector_size"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_XFER-set_hyper_vector_size}{}}}
\subsection{\code{H5P_DATASET_XFER$set_hyper_vector_size()}}{
  This function implements the HDF5-API function H5Pset_hyper_vector_size. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_XFER$set_hyper_vector_size(size = 2^10)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{size}}{The size of the object in bytes; use \code{Inf} for variable-length strings.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_XFER-get_hyper_vector_size"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_XFER-get_hyper_vector_size}{}}}
\subsection{\code{H5P_DATASET_XFER$get_hyper_vector_size()}}{
  This function implements the HDF5-API function H5Pget_hyper_vector_size. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_XFER$get_hyper_vector_size()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_XFER-set_btree_ratios"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_XFER-set_btree_ratios}{}}}
\subsection{\code{H5P_DATASET_XFER$set_btree_ratios()}}{
  This function implements the HDF5-API function H5Pset_btree_ratios. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_XFER$set_btree_ratios(left, middle, right)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{left}}{The B-tree split ratio for the left-most node.}
      \item{\code{middle}}{The B-tree split ratio for interior nodes.}
      \item{\code{right}}{The B-tree split ratio for the right-most node.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DATASET_XFER-get_btree_ratios"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DATASET_XFER-get_btree_ratios}{}}}
\subsection{\code{H5P_DATASET_XFER$get_btree_ratios()}}{
  This function implements the HDF5-API function H5Pget_btree_ratios. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DATASET_XFER$get_btree_ratios()}
    \if{html}{\out{</div>}}
  }
}

}
