% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_DEFAULT-class}
\alias{H5P_DEFAULT-class}
\alias{H5P_DEFAULT}
\title{Class for default values for HDF5 property lists.}
\value{
Object of class \code{\link{H5P}}.
}
\description{
This class represents default values for H5P property lists. As with \code{\link{H5S_ALL}}, the current choice is not optimal
for the same reasons and likely to be changed
}
\section{Methods}{

\describe{
\item{\code{print(...)}}{

Just prints that it is the default class

\strong{Parameters}
\describe{
\item{...}{ignored}
}}
}}

\author{
Holger Hoefling
}
\keyword{internal}
