% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_DEFAULT}
\alias{H5P_DEFAULT}
\alias{H5P_DEFAULT-class}
\title{Class for default values for HDF5 property lists.}
\value{
Object of class \code{\link{H5P}}.
}
\description{
This class represents default values for H5P property lists. As with \code{\link{H5S_ALL}}, the current choice is not optimal
for the same reasons and likely to be changed
}
\author{
Holger Hoefling
}
\keyword{internal}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5P]{H5P}} -> \code{H5P_DEFAULT}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5P_DEFAULT-initialize}{\code{H5P_DEFAULT$new()}}
    \item \href{#method-H5P_DEFAULT-print}{\code{H5P_DEFAULT$print()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="copy"><a href='../../hdf5r/html/H5P.html#method-H5P-copy'><code>H5P$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="equal"><a href='../../hdf5r/html/H5P.html#method-H5P-equal'><code>H5P$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class'><code>H5P$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class_name"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class_name'><code>H5P$get_class_name()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DEFAULT-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DEFAULT-initialize}{}}}
\subsection{\code{H5P_DEFAULT$new()}}{
  Wraps the default property list. For internal use only.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DEFAULT$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{An HDF5 id; for internal use only.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_DEFAULT-print"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_DEFAULT-print}{}}}
\subsection{\code{H5P_DEFAULT$print()}}{
  Just prints that it is the default class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_DEFAULT$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
    }
    \if{html}{\out{</div>}}
  }
}

}
