% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\name{H5P_factory}
\alias{H5P_factory}
\title{Create an \\code{\\link{H5P-class}} out of an id}
\usage{
H5P_factory(id)
}
\arguments{
\item{id}{The id to wrap inside an object}
}
\value{
An object of class \\code{\\link{H5P}}
}
\description{
Create an \\code{\\link{H5P-class}} out of an id
}
\details{
Function that determines the property list type of an id and creates
the appropriate class for it.
}
\author{
Holger Hoefling
}
