% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_FILE_ACCESS-class}
\alias{H5P_FILE_ACCESS}
\alias{H5P_FILE_ACCESS-class}
\title{Class for HDF5 property list for file creation}
\value{
Object of class \code{\link[=H5P_FILE_ACCESS-class]{H5P_FILE_ACCESS}}.
}
\description{
It inherits all functions of the \code{\link{H5P-class}}.
}
\section{Methods}{

\describe{
\item{\code{new}}{

\strong{Usage:}
\preformatted{new(id = NULL)}
Create a new class of type \code{\link{H5P_FILE_ACCESS}}

\strong{Parameters:}
\describe{
\item{id}{Internal use only}
}}

\item{\code{set_cache}}{

\strong{Usage:}
\preformatted{set_cache(rdcc_nslots = 521, rdcc_nbytes = 2^20, rdcc_w0 = 0.75)}
This function implements the HDF5-API function H5Pset_cache.
Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-SetCache} for details.}

\item{\code{get_cache}}{

\strong{Usage:}
\preformatted{get_cache()}
This function implements the HDF5-API function H5Pget_cache.
Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-GetCache} for details.}
}}
\author{
Holger Hoefling
}
\seealso{
H5Class_overview, \code{\link[=H5P-class]{H5P}}
}

