% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_FILE_ACCESS}
\alias{H5P_FILE_ACCESS}
\alias{H5P_FILE_ACCESS-class}
\title{Class for HDF5 property list for file creation}
\value{
Object of class \code{\link[=H5P_FILE_ACCESS]{H5P_FILE_ACCESS}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5P]{H5P}} -> \code{H5P_FILE_ACCESS}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5P_FILE_ACCESS-initialize}{\code{H5P_FILE_ACCESS$new()}}
    \item \href{#method-H5P_FILE_ACCESS-set_cache}{\code{H5P_FILE_ACCESS$set_cache()}}
    \item \href{#method-H5P_FILE_ACCESS-get_cache}{\code{H5P_FILE_ACCESS$get_cache()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="copy"><a href='../../hdf5r/html/H5P.html#method-H5P-copy'><code>H5P$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="equal"><a href='../../hdf5r/html/H5P.html#method-H5P-equal'><code>H5P$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class'><code>H5P$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class_name"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class_name'><code>H5P$get_class_name()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_ACCESS-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_ACCESS-initialize}{}}}
\subsection{\code{H5P_FILE_ACCESS$new()}}{
  Create a new class of type \code{\link{H5P_FILE_ACCESS}}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_ACCESS$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_ACCESS-set_cache"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_ACCESS-set_cache}{}}}
\subsection{\code{H5P_FILE_ACCESS$set_cache()}}{
  This function implements the HDF5-API function H5Pset_cache. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_ACCESS$set_cache(
  rdcc_nslots = 521,
  rdcc_nbytes = 2^20,
  rdcc_w0 = 0.75
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rdcc_nslots}}{The number of slots in the raw data chunk cache. Use -1 for the library default.}
      \item{\code{rdcc_nbytes}}{The size of the raw data chunk cache in bytes. Use -1 for the library default.}
      \item{\code{rdcc_w0}}{The chunk preemption policy, between 0 and 1. Use -1 for the library default.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_ACCESS-get_cache"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_ACCESS-get_cache}{}}}
\subsection{\code{H5P_FILE_ACCESS$get_cache()}}{
  This function implements the HDF5-API function H5Pget_cache. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_ACCESS$get_cache()}
    \if{html}{\out{</div>}}
  }
}

}
