% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_FILE_CREATE}
\alias{H5P_FILE_CREATE}
\alias{H5P_FILE_CREATE-class}
\title{Class for HDF5 property list for file creation}
\value{
Object of class \code{\link[=H5P_FILE_CREATE]{H5P_FILE_CREATE}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5P]{H5P}} -> \code{H5P_FILE_CREATE}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5P_FILE_CREATE-initialize}{\code{H5P_FILE_CREATE$new()}}
    \item \href{#method-H5P_FILE_CREATE-set_userblock}{\code{H5P_FILE_CREATE$set_userblock()}}
    \item \href{#method-H5P_FILE_CREATE-get_userblock}{\code{H5P_FILE_CREATE$get_userblock()}}
    \item \href{#method-H5P_FILE_CREATE-set_sizes}{\code{H5P_FILE_CREATE$set_sizes()}}
    \item \href{#method-H5P_FILE_CREATE-get_sizes}{\code{H5P_FILE_CREATE$get_sizes()}}
    \item \href{#method-H5P_FILE_CREATE-set_sym_k}{\code{H5P_FILE_CREATE$set_sym_k()}}
    \item \href{#method-H5P_FILE_CREATE-get_sym_k}{\code{H5P_FILE_CREATE$get_sym_k()}}
    \item \href{#method-H5P_FILE_CREATE-set_istore_k}{\code{H5P_FILE_CREATE$set_istore_k()}}
    \item \href{#method-H5P_FILE_CREATE-get_istore_k}{\code{H5P_FILE_CREATE$get_istore_k()}}
    \item \href{#method-H5P_FILE_CREATE-set_file_space}{\code{H5P_FILE_CREATE$set_file_space()}}
    \item \href{#method-H5P_FILE_CREATE-get_file_space}{\code{H5P_FILE_CREATE$get_file_space()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="copy"><a href='../../hdf5r/html/H5P.html#method-H5P-copy'><code>H5P$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="equal"><a href='../../hdf5r/html/H5P.html#method-H5P-equal'><code>H5P$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class'><code>H5P$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class_name"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class_name'><code>H5P$get_class_name()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-initialize}{}}}
\subsection{\code{H5P_FILE_CREATE$new()}}{
  Create a new class of type \code{\link{H5P_FILE_CREATE}}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-set_userblock"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-set_userblock}{}}}
\subsection{\code{H5P_FILE_CREATE$set_userblock()}}{
  This function implements the HDF5-API function H5Pset_userblock. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$set_userblock(size)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{size}}{The size of the object in bytes; use \code{Inf} for variable-length strings.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-get_userblock"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-get_userblock}{}}}
\subsection{\code{H5P_FILE_CREATE$get_userblock()}}{
  This function implements the HDF5-API function H5Pget_userblock. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$get_userblock()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-set_sizes"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-set_sizes}{}}}
\subsection{\code{H5P_FILE_CREATE$set_sizes()}}{
  This function implements the HDF5-API function H5Pset_sizes. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$set_sizes(sizeof_addr, sizeof_size)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{sizeof_addr}}{The size of an address in the file, in bytes.}
      \item{\code{sizeof_size}}{The size of a size in the file, in bytes.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-get_sizes"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-get_sizes}{}}}
\subsection{\code{H5P_FILE_CREATE$get_sizes()}}{
  This function implements the HDF5-API function H5Pget_sizes. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$get_sizes()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-set_sym_k"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-set_sym_k}{}}}
\subsection{\code{H5P_FILE_CREATE$set_sym_k()}}{
  This function implements the HDF5-API function H5Pset_sym_k. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$set_sym_k(ik, lk)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{ik}}{The 1/2 rank of the indexed storage or symbol table B-tree.}
      \item{\code{lk}}{The 1/2 rank of the symbol table leaf nodes.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-get_sym_k"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-get_sym_k}{}}}
\subsection{\code{H5P_FILE_CREATE$get_sym_k()}}{
  This function implements the HDF5-API function H5Pget_sym_k. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$get_sym_k()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-set_istore_k"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-set_istore_k}{}}}
\subsection{\code{H5P_FILE_CREATE$set_istore_k()}}{
  This function implements the HDF5-API function H5Pset_istore_k. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$set_istore_k(ik)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{ik}}{The 1/2 rank of the indexed storage or symbol table B-tree.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-get_istore_k"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-get_istore_k}{}}}
\subsection{\code{H5P_FILE_CREATE$get_istore_k()}}{
  This function implements the HDF5-API function H5Pget_istore_k. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$get_istore_k()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-set_file_space"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-set_file_space}{}}}
\subsection{\code{H5P_FILE_CREATE$set_file_space()}}{
  This function implements the HDF5-API function H5Pset_file_space. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$set_file_space(strategy, threshold)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{strategy}}{The file space handling strategy; one of the \code{H5F_FSPACE_STRATEGY_*} values in \code{\link{h5const}}.}
      \item{\code{threshold}}{The smallest free-space section to track.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_FILE_CREATE-get_file_space"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_FILE_CREATE-get_file_space}{}}}
\subsection{\code{H5P_FILE_CREATE$get_file_space()}}{
  This function implements the HDF5-API function H5Pget_file_space. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_FILE_CREATE$get_file_space()}
    \if{html}{\out{</div>}}
  }
}

}
