% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_OBJECT_CREATE}
\alias{H5P_OBJECT_CREATE}
\alias{H5P_OBJECT_CREATE-class}
\title{Class for HDF5 property list for object creation}
\value{
Object of class \code{\link[=H5P_OBJECT_CREATE]{H5P_OBJECT_CREATE}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5P]{H5P}} -> \code{H5P_OBJECT_CREATE}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5P_OBJECT_CREATE-initialize}{\code{H5P_OBJECT_CREATE$new()}}
    \item \href{#method-H5P_OBJECT_CREATE-set_obj_track_times}{\code{H5P_OBJECT_CREATE$set_obj_track_times()}}
    \item \href{#method-H5P_OBJECT_CREATE-get_obj_track_times}{\code{H5P_OBJECT_CREATE$get_obj_track_times()}}
    \item \href{#method-H5P_OBJECT_CREATE-set_attr_phase_change}{\code{H5P_OBJECT_CREATE$set_attr_phase_change()}}
    \item \href{#method-H5P_OBJECT_CREATE-get_attr_phase_change}{\code{H5P_OBJECT_CREATE$get_attr_phase_change()}}
    \item \href{#method-H5P_OBJECT_CREATE-set_attr_creation_order}{\code{H5P_OBJECT_CREATE$set_attr_creation_order()}}
    \item \href{#method-H5P_OBJECT_CREATE-get_attr_creation_order}{\code{H5P_OBJECT_CREATE$get_attr_creation_order()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="copy"><a href='../../hdf5r/html/H5P.html#method-H5P-copy'><code>H5P$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="equal"><a href='../../hdf5r/html/H5P.html#method-H5P-equal'><code>H5P$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class'><code>H5P$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5P" data-id="get_class_name"><a href='../../hdf5r/html/H5P.html#method-H5P-get_class_name'><code>H5P$get_class_name()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_OBJECT_CREATE-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_OBJECT_CREATE-initialize}{}}}
\subsection{\code{H5P_OBJECT_CREATE$new()}}{
  Create a new class of type \code{\link{H5P_OBJECT_CREATE}}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_OBJECT_CREATE$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_OBJECT_CREATE-set_obj_track_times"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_OBJECT_CREATE-set_obj_track_times}{}}}
\subsection{\code{H5P_OBJECT_CREATE$set_obj_track_times()}}{
  This function implements the HDF5-API function H5Pset_obj_track_times. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_OBJECT_CREATE$set_obj_track_times(track_times = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{track_times}}{Logical; should object times be recorded in the file.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_OBJECT_CREATE-get_obj_track_times"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_OBJECT_CREATE-get_obj_track_times}{}}}
\subsection{\code{H5P_OBJECT_CREATE$get_obj_track_times()}}{
  This function implements the HDF5-API function H5Pget_obj_track_times. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_OBJECT_CREATE$get_obj_track_times()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_OBJECT_CREATE-set_attr_phase_change"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_OBJECT_CREATE-set_attr_phase_change}{}}}
\subsection{\code{H5P_OBJECT_CREATE$set_attr_phase_change()}}{
  This function implements the HDF5-API function H5Pset_attr_phase_change. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_OBJECT_CREATE$set_attr_phase_change(max_compact, min_dense)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{max_compact}}{The maximum number of links to store compactly in a group.}
      \item{\code{min_dense}}{The minimum number of links before a group switches to dense storage.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_OBJECT_CREATE-get_attr_phase_change"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_OBJECT_CREATE-get_attr_phase_change}{}}}
\subsection{\code{H5P_OBJECT_CREATE$get_attr_phase_change()}}{
  This function implements the HDF5-API function H5Pget_attr_phase_change. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_OBJECT_CREATE$get_attr_phase_change()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_OBJECT_CREATE-set_attr_creation_order"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_OBJECT_CREATE-set_attr_creation_order}{}}}
\subsection{\code{H5P_OBJECT_CREATE$set_attr_creation_order()}}{
  This function implements the HDF5-API function H5Pset_attr_creation_order. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_OBJECT_CREATE$set_attr_creation_order(crt_order_flags = 0)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{crt_order_flags}}{Flags governing whether link creation order is tracked and indexed.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5P_OBJECT_CREATE-get_attr_creation_order"></a>}}
\if{latex}{\out{\hypertarget{method-H5P_OBJECT_CREATE-get_attr_creation_order}{}}}
\subsection{\code{H5P_OBJECT_CREATE$get_attr_creation_order()}}{
  This function implements the HDF5-API function H5Pget_attr_creation_order. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5P_OBJECT_CREATE$get_attr_creation_order()}
    \if{html}{\out{</div>}}
  }
}

}
