% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\docType{class}
\name{H5R}
\alias{H5R}
\alias{H5R_OBJECT}
\alias{H5R_DATASET_REGION}
\title{Class for HDF5 Reference datatypes.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
H5R_OBJECT

H5R_DATASET_REGION
}
\value{
Object of class \code{\link[=H5R-class]{H5R}}.
}
\description{
\code{H5R} is only the common, class, and is never used. User should not create objects of this
class by themselves and instead use the \code{create_reference} methods of \code{\link{H5D}}, \code{\link{H5Group}}, \code{\link{H5File}} classes.
}
\seealso{
H5Class_overview
}
\author{
Holger Hoefling
}
\keyword{datasets}
