% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\docType{class}
\name{H5R}
\alias{H5R}
\alias{H5R-class}
\title{Class for HDF5 Reference datatypes.}
\value{
Object of class \code{\link[=H5R]{H5R}}.
}
\description{
\code{H5R} is only the common base class and is never used. User should not create objects of this
class by themselves and instead use the \code{create_reference} methods of \code{H5D}, \code{H5Group} or \code{H5File} classes.
Sub-classes are \code{\link{H5R_OBJECT}} and \code{\link{H5R_DATASET_REGION}}
}
\examples{
fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
file[["testset"]] <- matrix(rnorm(9), nrow = 3)
dset <- file[["testset"]]
r <- file$create_reference("testset")
file$close_all()
}
\author{
Holger Hoefling
}
\section{Super class}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{H5R}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{length}}{Get the length of the object}

    \item{\code{ref}}{Get or assign the internal raw-vector representation of the data. Usually, user's shouldn't have to use this.}

    \item{\code{dim}}{Get or assign the dimensionality of the object}

    \item{\code{dimnames}}{Get or assign the dimnames of the object}

    \item{\code{names}}{Get or assign the names of the object}

    \item{\code{rank}}{Get the rank of the object}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5R-subset_read}{\code{H5R$subset_read()}}
    \item \href{#method-H5R-subset2_read}{\code{H5R$subset2_read()}}
    \item \href{#method-H5R-subset_assign}{\code{H5R$subset_assign()}}
    \item \href{#method-H5R-subset2_assign}{\code{H5R$subset2_assign()}}
    \item \href{#method-H5R-t}{\code{H5R$t()}}
    \item \href{#method-H5R-clone}{\code{H5R$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="initialize"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-initialize'><code>H5RefClass$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5R-subset_read"></a>}}
\if{latex}{\out{\hypertarget{method-H5R-subset_read}{}}}
\subsection{\code{H5R$subset_read()}}{
  Method that returns a subset of the data in the H5R-object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5R$subset_read(dim_index, drop = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dim_index}}{A list of dimension indices as usually pasted into \code{[}}
      \item{\code{drop}}{Logical. Should dimensions of size 1 be dropped.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5R-subset2_read"></a>}}
\if{latex}{\out{\hypertarget{method-H5R-subset2_read}{}}}
\subsection{\code{H5R$subset2_read()}}{
  Method to read a single item
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5R$subset2_read(i, exact = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{i}}{The single item to read}
      \item{\code{exact}}{Is the item name exact or should partial matching be allowed?}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5R-subset_assign"></a>}}
\if{latex}{\out{\hypertarget{method-H5R-subset_assign}{}}}
\subsection{\code{H5R$subset_assign()}}{
  Assign values into a subset of the H5R-vector
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5R$subset_assign(dim_index, value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dim_index}}{A list of dimension indices as usually passed into \code{[}}
      \item{\code{value}}{The value to assign}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5R-subset2_assign"></a>}}
\if{latex}{\out{\hypertarget{method-H5R-subset2_assign}{}}}
\subsection{\code{H5R$subset2_assign()}}{
  Assign a value to a single value in the array
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5R$subset2_assign(i, exact = TRUE, value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{i}}{the index where to assign the value}
      \item{\code{exact}}{Logical; does the subscript have to match exactly.}
      \item{\code{value}}{The value to assign}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5R-t"></a>}}
\if{latex}{\out{\hypertarget{method-H5R-t}{}}}
\subsection{\code{H5R$t()}}{
  Transpose the object if it is a matrix (i.e. has rank 2
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5R$t()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5R-clone"></a>}}
\if{latex}{\out{\hypertarget{method-H5R-clone}{}}}
\subsection{\code{H5R$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5R$clone()}
    \if{html}{\out{</div>}}
  }
}

}
