% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\docType{class}
\name{H5R_OBJECT}
\alias{H5R_OBJECT}
\alias{H5R_OBJECT-class}
\title{Class for HDF5 Object-references.}
\value{
Object of class \code{\link[=H5R_OBJECT]{H5R_OBJECT}}.
}
\description{
\code{H5R_OBJECT} is the reference class for objects. Users should not create this class by themselves, but use the appropriate 
and instead use the \code{create_reference} methods of \code{H5D}, \code{H5Group} or \code{H5File} classes.
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5R]{H5R}} -> \code{H5R_OBJECT}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5R_OBJECT-initialize}{\code{H5R_OBJECT$new()}}
    \item \href{#method-H5R_OBJECT-dereference}{\code{H5R_OBJECT$dereference()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="print"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-print'><code>H5RefClass$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5R" data-id="subset2_assign"><a href='../../hdf5r/html/H5R.html#method-H5R-subset2_assign'><code>H5R$subset2_assign()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5R" data-id="subset2_read"><a href='../../hdf5r/html/H5R.html#method-H5R-subset2_read'><code>H5R$subset2_read()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5R" data-id="subset_assign"><a href='../../hdf5r/html/H5R.html#method-H5R-subset_assign'><code>H5R$subset_assign()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5R" data-id="subset_read"><a href='../../hdf5r/html/H5R.html#method-H5R-subset_read'><code>H5R$subset_read()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5R" data-id="t"><a href='../../hdf5r/html/H5R.html#method-H5R-t'><code>H5R$t()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5R_OBJECT-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5R_OBJECT-initialize}{}}}
\subsection{\code{H5R_OBJECT$new()}}{
  Create a new reference for object; Usually, users shouldn't have to call this, but use the \code{create_reference} method of a dataset, group of committed datatype
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5R_OBJECT$new(num = 0, id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{num}}{The number of items.}
      \item{\code{id}}{An HDF5 id; for internal use only.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5R_OBJECT-dereference"></a>}}
\if{latex}{\out{\hypertarget{method-H5R_OBJECT-dereference}{}}}
\subsection{\code{H5R_OBJECT$dereference()}}{
  Dereference an H5R reference. The file the reference is pointing to is assigned automatically This function implements the HDF5-API function H5Rdereference. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_r.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5R_OBJECT$dereference(object_access_pl = h5const$H5P_DEFAULT, obj = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{object_access_pl}}{The object-access property list. Currently always the default}
      \item{\code{obj}}{Overriding the default file the reference is referring to}
    }
    \if{html}{\out{</div>}}
  }
}

}
