% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes.R
\docType{class}
\name{H5RefClass}
\alias{H5RefClass}
\title{Base class that tracks the ids and allows for closing an id}
\value{
Object of reference class \code{\link{H5RefClass}}.
}
\description{
This class is not indended for creating objects, but as a base class for all other
H5-derived classes to provide common functionality for id tracking
}
\section{Fields}{

\describe{
\item{\code{id}}{Returns the id of the object as an integer}
}}

\author{
Holger Hoefling
}
