% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes.R
\docType{class}
\name{H5RefClass}
\alias{H5RefClass}
\alias{H5RefClass-class}
\title{Base class that tracks the ids and allows for closing an id}
\value{
Object of reference class \code{\link{H5RefClass}}.
}
\description{
This class is not intended for creating objects, but as a base class for all other
H5-derived classes to provide common functionality for id tracking
}
\author{
Holger Hoefling
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{id}}{Returns the id of the object}

    \item{\code{id_env}}{The environment that holds the id, used internally for reference counting.}

    \item{\code{is_valid}}{This function implements the HDF5-API function H5Iis_valid. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details. Additionally, the R-object representing the HDF5-id can be invalidated as well. In this case, the class id is set to \code{NA} and \code{is_valid} returns \code{FALSE}.}

    \item{\code{message}}{Legacy function; currently not used; may be removed}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5RefClass-get_file_id}{\code{H5RefClass$get_file_id()}}
    \item \href{#method-H5RefClass-get_obj_type}{\code{H5RefClass$get_obj_type()}}
    \item \href{#method-H5RefClass-get_ref}{\code{H5RefClass$get_ref()}}
    \item \href{#method-H5RefClass-inc_ref}{\code{H5RefClass$inc_ref()}}
    \item \href{#method-H5RefClass-dec_ref}{\code{H5RefClass$dec_ref()}}
    \item \href{#method-H5RefClass-initialize}{\code{H5RefClass$new()}}
    \item \href{#method-H5RefClass-close}{\code{H5RefClass$close()}}
    \item \href{#method-H5RefClass-print}{\code{H5RefClass$print()}}
    \item \href{#method-H5RefClass-methods}{\code{H5RefClass$methods()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5RefClass-get_file_id"></a>}}
\if{latex}{\out{\hypertarget{method-H5RefClass-get_file_id}{}}}
\subsection{\code{H5RefClass$get_file_id()}}{
  This function implements the HDF5-API function H5Iget_file_id. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5RefClass$get_file_id()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5RefClass-get_obj_type"></a>}}
\if{latex}{\out{\hypertarget{method-H5RefClass-get_obj_type}{}}}
\subsection{\code{H5RefClass$get_obj_type()}}{
  This function implements the HDF5-API function H5Iget_type. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5RefClass$get_obj_type()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5RefClass-get_ref"></a>}}
\if{latex}{\out{\hypertarget{method-H5RefClass-get_ref}{}}}
\subsection{\code{H5RefClass$get_ref()}}{
  This function implements the HDF5-API function H5Iget_ref. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5RefClass$get_ref()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5RefClass-inc_ref"></a>}}
\if{latex}{\out{\hypertarget{method-H5RefClass-inc_ref}{}}}
\subsection{\code{H5RefClass$inc_ref()}}{
  This function implements the HDF5-API function H5Iinc_ref. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5RefClass$inc_ref()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5RefClass-dec_ref"></a>}}
\if{latex}{\out{\hypertarget{method-H5RefClass-dec_ref}{}}}
\subsection{\code{H5RefClass$dec_ref()}}{
  This function implements the HDF5-API function H5Idec_ref. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5RefClass$dec_ref()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5RefClass-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5RefClass-initialize}{}}}
\subsection{\code{H5RefClass$new()}}{
  Constructor for the basic class for hdf5 objects. Takes an id and stores it appropriately, including the necessary counting of object references that the package implements. This reference counting is included in addition to R's internal method in order to allow for the invalidation of objects in R itself when all open handles in an R-file are being closed.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5RefClass$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{An HDF5 id; for internal use only.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5RefClass-close"></a>}}
\if{latex}{\out{\hypertarget{method-H5RefClass-close}{}}}
\subsection{\code{H5RefClass$close()}}{
  Closes an object and calls the appropriate HDF5 function for the type of object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5RefClass$close()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5RefClass-print"></a>}}
\if{latex}{\out{\hypertarget{method-H5RefClass-print}{}}}
\subsection{\code{H5RefClass$print()}}{
  Prints the class of the object and the id
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5RefClass$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Further arguments; currently ignored.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5RefClass-methods"></a>}}
\if{latex}{\out{\hypertarget{method-H5RefClass-methods}{}}}
\subsection{\code{H5RefClass$methods()}}{
  Prints available methods on the screen
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5RefClass$methods()}
    \if{html}{\out{</div>}}
  }
}

}
