% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5S.R
\docType{class}
\name{H5S}
\alias{H5S}
\title{Class for representing HDF5 spaces}
\value{
Object of class \code{\link{H5S}}.
}
\description{
This class represents an HDF5 space-id. It inherits all functions of the
\code{\link{H5RefClass-class}}.
}
\seealso{
H5Class_overview
}
\author{
Holger Hoefling
}
