% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5S.R
\docType{class}
\name{H5S}
\alias{H5S}
\alias{H5S-class}
\title{Class for representing HDF5 spaces}
\value{
Object of class \code{\link{H5S}}.
}
\description{
This class represents \code{Spaces} in HDF5. These are mostly useful to define the
dimensions of a dataset as well as the maximum dimensions to which it can grow. By default, the
maximum dimension is equal to the initial dimension. If you want the array to be able to grow arbitrarily
large in one dimension, set the maximum dimension for this index to \code{Inf}. See the examples below
for code how to do this.
}
\examples{
h5s_fixed <- H5S$new("simple", dims=c(5, 2))
h5s_fixed

h5s_variable <- H5S$new("simple", dims=c(5,2), maxdims=c(Inf,2))
h5s_variable
h5s_variable$set_extent_simple(c(10,2), c(Inf, 2))
h5s_variable

# now select a subset of points
# argument evaluation has a heuristic; here it chooses point selection
h5s_variable[c(1, 3, 8), 1]
h5s_variable$get_select_type()
h5s_variable$get_select_elem_pointlist()

# and a hyperslab (chosen by the argument heuristic)
h5s_variable[2:7, 1:2]
h5s_variable$get_select_type()
h5s_variable$get_select_hyper_blocklist()
}
\author{
Holger Hoefling
}
\section{Super class}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{H5S}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{dims}}{Get the dimensions of the space. Return NULL if the space is not simple (i.e. NULL-space) or a length-0 integer if it is a scalar}

    \item{\code{maxdims}}{Get the maximal dimensions of the space.Return NULL if the space is not simple (i.e. NULL-space) or a length-0 integer if it is a scalar}

    \item{\code{rank}}{The rank, i.e. the number of dimensions, of the space.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5S-initialize}{\code{H5S$new()}}
    \item \href{#method-H5S-copy}{\code{H5S$copy()}}
    \item \href{#method-H5S-encode}{\code{H5S$encode()}}
    \item \href{#method-H5S-is_simple}{\code{H5S$is_simple()}}
    \item \href{#method-H5S-get_simple_extent_ndims}{\code{H5S$get_simple_extent_ndims()}}
    \item \href{#method-H5S-offset_simple}{\code{H5S$offset_simple()}}
    \item \href{#method-H5S-get_simple_extent_dims}{\code{H5S$get_simple_extent_dims()}}
    \item \href{#method-H5S-get_simple_extent_npoints}{\code{H5S$get_simple_extent_npoints()}}
    \item \href{#method-H5S-get_simple_extent_type}{\code{H5S$get_simple_extent_type()}}
    \item \href{#method-H5S-extent_copy}{\code{H5S$extent_copy()}}
    \item \href{#method-H5S-extent_equal}{\code{H5S$extent_equal()}}
    \item \href{#method-H5S-set_extent_simple}{\code{H5S$set_extent_simple()}}
    \item \href{#method-H5S-set_extent_none}{\code{H5S$set_extent_none()}}
    \item \href{#method-H5S-get_select_type}{\code{H5S$get_select_type()}}
    \item \href{#method-H5S-get_select_npoints}{\code{H5S$get_select_npoints()}}
    \item \href{#method-H5S-get_select_hyper_nblocks}{\code{H5S$get_select_hyper_nblocks()}}
    \item \href{#method-H5S-get_select_hyper_blocklist}{\code{H5S$get_select_hyper_blocklist()}}
    \item \href{#method-H5S-get_select_elem_npoints}{\code{H5S$get_select_elem_npoints()}}
    \item \href{#method-H5S-get_select_elem_pointlist}{\code{H5S$get_select_elem_pointlist()}}
    \item \href{#method-H5S-get_select_bounds}{\code{H5S$get_select_bounds()}}
    \item \href{#method-H5S-select_all}{\code{H5S$select_all()}}
    \item \href{#method-H5S-select_none}{\code{H5S$select_none()}}
    \item \href{#method-H5S-select_valid}{\code{H5S$select_valid()}}
    \item \href{#method-H5S-select_elements}{\code{H5S$select_elements()}}
    \item \href{#method-H5S-select_hyperslab}{\code{H5S$select_hyperslab()}}
    \item \href{#method-H5S-subset}{\code{H5S$subset()}}
    \item \href{#method-H5S-print}{\code{H5S$print()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-initialize}{}}}
\subsection{\code{H5S$new()}}{
  Create a new HDF5-space. This can be done by either specifying a space with appropriate dimensions or by decoding a character string that represents an encoded space
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$new(
  type = c("simple", "scalar", "null"),
  dims = NULL,
  maxdims = dims,
  decode_buf = NULL,
  id = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{Either a \code{simple} space, for which \code{dims} and \code{maxdims} have to be given or a \code{scalar} or \code{null} space. See the HDF5 user guide on spaces to explain the differences.}
      \item{\code{dims}}{The dimension of the space in case it is of type \code{simple}}
      \item{\code{maxdims}}{The maximal dimensions of the space}
      \item{\code{decode_buf}}{The character string that holds the encoded representation of a space}
      \item{\code{id}}{An existing HDF5 id; internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-copy"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-copy}{}}}
\subsection{\code{H5S$copy()}}{
  This function implements the HDF5-API function H5Scopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$copy()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-encode"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-encode}{}}}
\subsection{\code{H5S$encode()}}{
  This function implements the HDF5-API function H5Sencode. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$encode()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-is_simple"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-is_simple}{}}}
\subsection{\code{H5S$is_simple()}}{
  This function implements the HDF5-API function H5Sis_simple. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$is_simple()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_simple_extent_ndims"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_simple_extent_ndims}{}}}
\subsection{\code{H5S$get_simple_extent_ndims()}}{
  This function implements the HDF5-API function H5Sget_simple_extent_ndims. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_simple_extent_ndims()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-offset_simple"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-offset_simple}{}}}
\subsection{\code{H5S$offset_simple()}}{
  This function implements the HDF5-API function H5Soffset_simple. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$offset_simple(offset)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{offset}}{The offset to apply; see the corresponding HDF5 function for how it is interpreted.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_simple_extent_dims"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_simple_extent_dims}{}}}
\subsection{\code{H5S$get_simple_extent_dims()}}{
  This function implements the HDF5-API function H5Sget_simple_extent_dims. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_simple_extent_dims()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_simple_extent_npoints"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_simple_extent_npoints}{}}}
\subsection{\code{H5S$get_simple_extent_npoints()}}{
  This function implements the HDF5-API function H5Sget_simple_extent_npoints. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_simple_extent_npoints()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_simple_extent_type"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_simple_extent_type}{}}}
\subsection{\code{H5S$get_simple_extent_type()}}{
  This function implements the HDF5-API function H5Sget_simple_extent_type. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_simple_extent_type()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-extent_copy"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-extent_copy}{}}}
\subsection{\code{H5S$extent_copy()}}{
  This function implements the HDF5-API function H5Sextent_copy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$extent_copy(h5s_source)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{h5s_source}}{The dataspace to copy the extent from; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-extent_equal"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-extent_equal}{}}}
\subsection{\code{H5S$extent_equal()}}{
  This function implements the HDF5-API function H5Sextent_equal. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$extent_equal(h5s_cmp)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{h5s_cmp}}{The dataspace to compare the extent against; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-set_extent_simple"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-set_extent_simple}{}}}
\subsection{\code{H5S$set_extent_simple()}}{
  This function implements the HDF5-API function H5Sset_extent_simple. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$set_extent_simple(dims, maxdims)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dims}}{The dimensions of the object.}
      \item{\code{maxdims}}{The maximum dimensions of the dataspace; use \code{Inf} for unlimited.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-set_extent_none"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-set_extent_none}{}}}
\subsection{\code{H5S$set_extent_none()}}{
  This function implements the HDF5-API function H5Sset_extent_none. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$set_extent_none()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_select_type"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_select_type}{}}}
\subsection{\code{H5S$get_select_type()}}{
  This function implements the HDF5-API function H5Sget_select_type. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_select_type()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_select_npoints"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_select_npoints}{}}}
\subsection{\code{H5S$get_select_npoints()}}{
  This function implements the HDF5-API function H5Sget_select_npoints. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_select_npoints()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_select_hyper_nblocks"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_select_hyper_nblocks}{}}}
\subsection{\code{H5S$get_select_hyper_nblocks()}}{
  This function implements the HDF5-API function H5Sget_select_hyper_nblocks. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_select_hyper_nblocks()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_select_hyper_blocklist"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_select_hyper_blocklist}{}}}
\subsection{\code{H5S$get_select_hyper_blocklist()}}{
  This function implements the HDF5-API function H5Sget_select_hyper_blocklist. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_select_hyper_blocklist(
  startblock = 0,
  numblocks = (self$get_select_hyper_nblocks() - startblock)
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{startblock}}{The first block to return; zero-based.}
      \item{\code{numblocks}}{The number of blocks to return.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_select_elem_npoints"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_select_elem_npoints}{}}}
\subsection{\code{H5S$get_select_elem_npoints()}}{
  This function implements the HDF5-API function H5Sget_select_elem_npoints. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_select_elem_npoints()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_select_elem_pointlist"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_select_elem_pointlist}{}}}
\subsection{\code{H5S$get_select_elem_pointlist()}}{
  This function implements the HDF5-API function H5Sget_select_elem_pointlist. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_select_elem_pointlist(
  startpoint = 0,
  numpoints = (self$get_select_elem_npoints() - startpoint)
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{startpoint}}{The first point to return; zero-based.}
      \item{\code{numpoints}}{The number of points to return.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-get_select_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-get_select_bounds}{}}}
\subsection{\code{H5S$get_select_bounds()}}{
  This function implements the HDF5-API function H5Sget_select_bounds. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$get_select_bounds()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-select_all"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-select_all}{}}}
\subsection{\code{H5S$select_all()}}{
  This function implements the HDF5-API function H5Sselect_all. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$select_all()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-select_none"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-select_none}{}}}
\subsection{\code{H5S$select_none()}}{
  This function implements the HDF5-API function H5Sselect_none. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$select_none()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-select_valid"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-select_valid}{}}}
\subsection{\code{H5S$select_valid()}}{
  This function implements the HDF5-API function H5Sselect_valid. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$select_valid()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-select_elements"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-select_elements}{}}}
\subsection{\code{H5S$select_elements()}}{
  This function implements the HDF5-API function H5Sselect_elements. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$select_elements(coord, op = h5const$H5S_SELECT_SET, byrow = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{coord}}{The coordinates of the elements to select.}
      \item{\code{op}}{The selection operator to apply; one of the \code{H5S_SELECT_*} values in \code{\link{h5const}}.}
      \item{\code{byrow}}{Logical; are the coordinates given row-wise.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-select_hyperslab"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-select_hyperslab}{}}}
\subsection{\code{H5S$select_hyperslab()}}{
  This function implements the HDF5-API function H5Sselect_hyperslab. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$select_hyperslab(
  start,
  count,
  stride = NULL,
  block = NULL,
  op = h5const$H5S_SELECT_SET
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start}}{The offset at which the hyperslab starts, one value per dimension.}
      \item{\code{count}}{The number of blocks to select, one value per dimension.}
      \item{\code{stride}}{The distance between successive blocks, one value per dimension.}
      \item{\code{block}}{The size of each block, one value per dimension.}
      \item{\code{op}}{The selection operator to apply; one of the \code{H5S_SELECT_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-subset"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-subset}{}}}
\subsection{\code{H5S$subset()}}{
  Subsetting the space. This is mainly intended as a helper function for the '[' function, but can also be used on its own.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$subset(args, op = h5const$H5S_SELECT_SET, envir = parent.frame())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{args}}{The indices for each dimension to subset given as a list. This makes this easier to use as a programmatic API. For interactive use we recommend the use of the \code{[} operator.}
      \item{\code{op}}{The operator to use. Same as for the other HDF5 space selection functions. One of the elements shown in \code{h5const$H5S_seloper_t}}
      \item{\code{envir}}{The environment in which to evaluate \code{args}}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S-print"></a>}}
\if{latex}{\out{\hypertarget{method-H5S-print}{}}}
\subsection{\code{H5S$print()}}{
  Prints information for the group
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
    }
    \if{html}{\out{</div>}}
  }
}

}
