% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5S.R
\docType{class}
\name{H5S_ALL-class}
\alias{H5S_ALL-class}
\alias{H5S_ALL}
\title{Class for HDF5 default space}
\value{
Object of class \code{\link[=H5S_ALL]{H5S_ALL}}.
}
\description{
It inherits all functions of the \code{\link{H5S}}. As this is implemented, closing the id is overriden as this is a special id.
}
\section{Methods}{

\describe{
\item{\code{print(...)}}{

Just prints that it is the default class

\strong{Parameters}
\describe{
\item{...}{ignored}
}}
}}

\seealso{
\code{\link[=H5S]{H5S}}
}
\author{
Holger Hoefling
}
\keyword{internal}
