% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5S.R
\docType{class}
\name{H5S_ALL}
\alias{H5S_ALL}
\alias{H5S_ALL-class}
\title{Class for HDF5 default space}
\value{
Object of class \code{\link[=H5S_ALL]{H5S_ALL}}.
}
\description{
It inherits all functions of the \code{\link{H5S}}. As this is implemented, closing the id is overridden as this is a special id.
}
\seealso{
\code{\link[=H5S]{H5S}}
}
\author{
Holger Hoefling
}
\keyword{internal}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5S]{H5S}} -> \code{H5S_ALL}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5S_ALL-initialize}{\code{H5S_ALL$new()}}
    \item \href{#method-H5S_ALL-print}{\code{H5S_ALL$print()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="copy"><a href='../../hdf5r/html/H5S.html#method-H5S-copy'><code>H5S$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="encode"><a href='../../hdf5r/html/H5S.html#method-H5S-encode'><code>H5S$encode()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="extent_copy"><a href='../../hdf5r/html/H5S.html#method-H5S-extent_copy'><code>H5S$extent_copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="extent_equal"><a href='../../hdf5r/html/H5S.html#method-H5S-extent_equal'><code>H5S$extent_equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_select_bounds"><a href='../../hdf5r/html/H5S.html#method-H5S-get_select_bounds'><code>H5S$get_select_bounds()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_select_elem_npoints"><a href='../../hdf5r/html/H5S.html#method-H5S-get_select_elem_npoints'><code>H5S$get_select_elem_npoints()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_select_elem_pointlist"><a href='../../hdf5r/html/H5S.html#method-H5S-get_select_elem_pointlist'><code>H5S$get_select_elem_pointlist()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_select_hyper_blocklist"><a href='../../hdf5r/html/H5S.html#method-H5S-get_select_hyper_blocklist'><code>H5S$get_select_hyper_blocklist()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_select_hyper_nblocks"><a href='../../hdf5r/html/H5S.html#method-H5S-get_select_hyper_nblocks'><code>H5S$get_select_hyper_nblocks()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_select_npoints"><a href='../../hdf5r/html/H5S.html#method-H5S-get_select_npoints'><code>H5S$get_select_npoints()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_select_type"><a href='../../hdf5r/html/H5S.html#method-H5S-get_select_type'><code>H5S$get_select_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_simple_extent_dims"><a href='../../hdf5r/html/H5S.html#method-H5S-get_simple_extent_dims'><code>H5S$get_simple_extent_dims()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_simple_extent_ndims"><a href='../../hdf5r/html/H5S.html#method-H5S-get_simple_extent_ndims'><code>H5S$get_simple_extent_ndims()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_simple_extent_npoints"><a href='../../hdf5r/html/H5S.html#method-H5S-get_simple_extent_npoints'><code>H5S$get_simple_extent_npoints()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="get_simple_extent_type"><a href='../../hdf5r/html/H5S.html#method-H5S-get_simple_extent_type'><code>H5S$get_simple_extent_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="is_simple"><a href='../../hdf5r/html/H5S.html#method-H5S-is_simple'><code>H5S$is_simple()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="offset_simple"><a href='../../hdf5r/html/H5S.html#method-H5S-offset_simple'><code>H5S$offset_simple()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="select_all"><a href='../../hdf5r/html/H5S.html#method-H5S-select_all'><code>H5S$select_all()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="select_elements"><a href='../../hdf5r/html/H5S.html#method-H5S-select_elements'><code>H5S$select_elements()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="select_hyperslab"><a href='../../hdf5r/html/H5S.html#method-H5S-select_hyperslab'><code>H5S$select_hyperslab()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="select_none"><a href='../../hdf5r/html/H5S.html#method-H5S-select_none'><code>H5S$select_none()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="select_valid"><a href='../../hdf5r/html/H5S.html#method-H5S-select_valid'><code>H5S$select_valid()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="set_extent_none"><a href='../../hdf5r/html/H5S.html#method-H5S-set_extent_none'><code>H5S$set_extent_none()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="set_extent_simple"><a href='../../hdf5r/html/H5S.html#method-H5S-set_extent_simple'><code>H5S$set_extent_simple()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5S" data-id="subset"><a href='../../hdf5r/html/H5S.html#method-H5S-subset'><code>H5S$subset()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S_ALL-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5S_ALL-initialize}{}}}
\subsection{\code{H5S_ALL$new()}}{
  Wraps the \code{H5S_ALL} dataspace. For internal use only.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S_ALL$new(id)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{An HDF5 id; for internal use only.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5S_ALL-print"></a>}}
\if{latex}{\out{\hypertarget{method-H5S_ALL-print}{}}}
\subsection{\code{H5S_ALL$print()}}{
  Just prints that it is the default class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5S_ALL$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
    }
    \if{html}{\out{</div>}}
  }
}

}
