% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\docType{class}
\name{H5T}
\alias{H5T}
\title{Class for HDF5 datatypes.}
\value{
Object of class \code{\link{H5T}}.
}
\description{
This is the base class for all datatypes, but most have a specialised class.
This class represents an HDF5 datatype. It inherits all functions of the
\code{\link{H5RefClass-class}}. It is also the base class for many other classes well, specifically
\describe{
\item{Integer}{\code{\link{H5T_INTEGER-class}}}
\item{Bitfield}{\code{\link{H5T_BITFIELD-class}} (currently identical to the integer class)}
\item{Float}{\code{\link{H5T_FLOAT-class}}}
\item{Enum}{\code{\link{H5T_ENUM-class}}}
\item{Compound}{\code{\link{H5T_COMPOUND-class}}}
\item{String}{\code{\link{H5T_STRING-class}}}
\item{Complex}{\code{\link{H5T_COMPLEX-class}}}
\item{Array}{\code{\link{H5T_ARRAY-class}}}
\item{Variable Length}{\code{\link{H5T_VLEN-class}}}
}
}
\seealso{
H5Class_overview
}
\author{
Holger Hoefling
}
