% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\docType{class}
\name{H5T_ARRAY}
\alias{H5T_ARRAY}
\alias{H5T_ARRAY-class}
\title{Class for HDF5 array datatypes.}
\value{
Object of class \code{\link[=H5T_ARRAY]{H5T_ARRAY}}.
}
\description{
Inherits from class \code{\link[=H5T]{H5T}}. This class represents an array. As datasets in HDF5
are itself already arrays, this datatype is not needed there. It is mostly useful when a column in a
\code{H5T_COMPUND} object is intended to be an array. This however makes it difficult to work
with such objects in R - as a column of the corresponding \code{data.frame} has to be an array. So please
use with care.
}
\seealso{
\code{\link[=H5T]{H5T}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5T]{H5T}} -> \code{H5T_ARRAY}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5T_ARRAY-initialize}{\code{H5T_ARRAY$new()}}
    \item \href{#method-H5T_ARRAY-get_array_ndims}{\code{H5T_ARRAY$get_array_ndims()}}
    \item \href{#method-H5T_ARRAY-get_array_dims}{\code{H5T_ARRAY$get_array_dims()}}
    \item \href{#method-H5T_ARRAY-get_super}{\code{H5T_ARRAY$get_super()}}
    \item \href{#method-H5T_ARRAY-describe}{\code{H5T_ARRAY$describe()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete'><code>H5T$attr_delete()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete_by_idx'><code>H5T$attr_delete_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete_by_name'><code>H5T$attr_delete_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_exists"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_exists'><code>H5T$attr_exists()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_exists_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_exists_by_name'><code>H5T$attr_exists_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_get_number"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_get_number'><code>H5T$attr_get_number()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_info_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_info_by_idx'><code>H5T$attr_info_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_info_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_info_by_name'><code>H5T$attr_info_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_name_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_name_by_idx'><code>H5T$attr_name_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open'><code>H5T$attr_open()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open_by_idx'><code>H5T$attr_open_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open_by_name'><code>H5T$attr_open_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_rename"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_rename'><code>H5T$attr_rename()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_rename_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_rename_by_name'><code>H5T$attr_rename_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="copy"><a href='../../hdf5r/html/H5T.html#method-H5T-copy'><code>H5T$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_attr"><a href='../../hdf5r/html/H5T.html#method-H5T-create_attr'><code>H5T$create_attr()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_attr_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-create_attr_by_name'><code>H5T$create_attr_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_reference"><a href='../../hdf5r/html/H5T.html#method-H5T-create_reference'><code>H5T$create_reference()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="detect_class"><a href='../../hdf5r/html/H5T.html#method-H5T-detect_class'><code>H5T$detect_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="equal"><a href='../../hdf5r/html/H5T.html#method-H5T-equal'><code>H5T$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_class"><a href='../../hdf5r/html/H5T.html#method-H5T-get_class'><code>H5T$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_create_plist"><a href='../../hdf5r/html/H5T.html#method-H5T-get_create_plist'><code>H5T$get_create_plist()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_native_type"><a href='../../hdf5r/html/H5T.html#method-H5T-get_native_type'><code>H5T$get_native_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_obj_name"><a href='../../hdf5r/html/H5T.html#method-H5T-get_obj_name'><code>H5T$get_obj_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_offset"><a href='../../hdf5r/html/H5T.html#method-H5T-get_offset'><code>H5T$get_offset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_order"><a href='../../hdf5r/html/H5T.html#method-H5T-get_order'><code>H5T$get_order()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_pad"><a href='../../hdf5r/html/H5T.html#method-H5T-get_pad'><code>H5T$get_pad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_precision"><a href='../../hdf5r/html/H5T.html#method-H5T-get_precision'><code>H5T$get_precision()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_size"><a href='../../hdf5r/html/H5T.html#method-H5T-get_size'><code>H5T$get_size()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="is_committed"><a href='../../hdf5r/html/H5T.html#method-H5T-is_committed'><code>H5T$is_committed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="is_vlen"><a href='../../hdf5r/html/H5T.html#method-H5T-is_vlen'><code>H5T$is_vlen()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="obj_info"><a href='../../hdf5r/html/H5T.html#method-H5T-obj_info'><code>H5T$obj_info()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="print"><a href='../../hdf5r/html/H5T.html#method-H5T-print'><code>H5T$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_offset"><a href='../../hdf5r/html/H5T.html#method-H5T-set_offset'><code>H5T$set_offset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_order"><a href='../../hdf5r/html/H5T.html#method-H5T-set_order'><code>H5T$set_order()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_pad"><a href='../../hdf5r/html/H5T.html#method-H5T-set_pad'><code>H5T$set_pad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_precision"><a href='../../hdf5r/html/H5T.html#method-H5T-set_precision'><code>H5T$set_precision()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_size"><a href='../../hdf5r/html/H5T.html#method-H5T-set_size'><code>H5T$set_size()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="to_text"><a href='../../hdf5r/html/H5T.html#method-H5T-to_text'><code>H5T$to_text()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_ARRAY-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_ARRAY-initialize}{}}}
\subsection{\code{H5T_ARRAY$new()}}{
  Create an array datatype.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_ARRAY$new(dims, dtype_base, id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dims}}{The dimension of the datatype}
      \item{\code{dtype_base}}{The datatype that makes up the elements of the array}
      \item{\code{id}}{internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_ARRAY-get_array_ndims"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_ARRAY-get_array_ndims}{}}}
\subsection{\code{H5T_ARRAY$get_array_ndims()}}{
  This function implements the HDF5-API function H5Tget_array_ndims. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_ARRAY$get_array_ndims()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_ARRAY-get_array_dims"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_ARRAY-get_array_dims}{}}}
\subsection{\code{H5T_ARRAY$get_array_dims()}}{
  This function implements the HDF5-API function H5Tget_array_dims2. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_ARRAY$get_array_dims()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_ARRAY-get_super"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_ARRAY-get_super}{}}}
\subsection{\code{H5T_ARRAY$get_super()}}{
  This function implements the HDF5-API function H5Tget_super. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_ARRAY$get_super()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_ARRAY-describe"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_ARRAY-describe}{}}}
\subsection{\code{H5T_ARRAY$describe()}}{
  Print a detailed description of the datatype; this is experimental
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_ARRAY$describe()}
    \if{html}{\out{</div>}}
  }
}

}
