% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\docType{class}
\name{H5T_COMPLEX-class}
\alias{H5T_COMPLEX}
\alias{H5T_COMPLEX-class}
\title{Class for HDF5 complex datatypes}
\value{
Object of class \code{\link[=H5T_COMPLEX-class]{H5T_COMPLEX}}.
}
\description{
In HDF5, complex numbers don't actually exist. They are represented as H5T_COMPOUND with two columns named \code{Real} and
\code{Imaginary}.
Inherits from class \code{\link[=H5T_COMPOUND-class]{H5T_COMPOUND}}.
}
\section{Methods}{

\describe{
\item{\code{new}}{

\strong{Usage:}
\preformatted{new(id = NULL)}
Create a new complex datatype

\strong{Parameters:}
\describe{
\item{id}{Internal use only}
}}
}}
\author{
Holger Hoefling
}
\seealso{
H5Class_overview, \code{\link[=H5T-class]{H5T}}, \code{\link[=H5T_COMPOUND-class]{H5T_COMPOUND}}
}

