% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\docType{class}
\name{H5T_COMPOUND}
\alias{H5T_COMPOUND}
\alias{H5T_COMPOUND-class}
\title{Class for HDF5 compound datatypes.}
\value{
Object of class \code{\link[=H5T_COMPOUND]{H5T_COMPOUND}}.
}
\description{
Inherits from class \code{\link[=H5T]{H5T}}.
}
\examples{
# create a H5T_COMPOUND  corresponding to a data-frame
my_cpd <- H5T_COMPOUND$new(c("name", "age", "salary"),
    dtypes=list(H5T_STRING$new(size=200), h5types$H5T_NATIVE_INT, h5types$H5T_NATIVE_DOUBLE))
my_cpd
}
\seealso{
\code{\link[=H5T]{H5T}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5T]{H5T}} -> \code{H5T_COMPOUND}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5T_COMPOUND-initialize}{\code{H5T_COMPOUND$new()}}
    \item \href{#method-H5T_COMPOUND-pack}{\code{H5T_COMPOUND$pack()}}
    \item \href{#method-H5T_COMPOUND-get_cpd_types}{\code{H5T_COMPOUND$get_cpd_types()}}
    \item \href{#method-H5T_COMPOUND-get_cpd_labels}{\code{H5T_COMPOUND$get_cpd_labels()}}
    \item \href{#method-H5T_COMPOUND-get_cpd_classes}{\code{H5T_COMPOUND$get_cpd_classes()}}
    \item \href{#method-H5T_COMPOUND-get_cpd_offsets}{\code{H5T_COMPOUND$get_cpd_offsets()}}
    \item \href{#method-H5T_COMPOUND-describe}{\code{H5T_COMPOUND$describe()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete'><code>H5T$attr_delete()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete_by_idx'><code>H5T$attr_delete_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete_by_name'><code>H5T$attr_delete_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_exists"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_exists'><code>H5T$attr_exists()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_exists_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_exists_by_name'><code>H5T$attr_exists_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_get_number"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_get_number'><code>H5T$attr_get_number()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_info_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_info_by_idx'><code>H5T$attr_info_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_info_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_info_by_name'><code>H5T$attr_info_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_name_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_name_by_idx'><code>H5T$attr_name_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open'><code>H5T$attr_open()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open_by_idx'><code>H5T$attr_open_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open_by_name'><code>H5T$attr_open_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_rename"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_rename'><code>H5T$attr_rename()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_rename_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_rename_by_name'><code>H5T$attr_rename_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="copy"><a href='../../hdf5r/html/H5T.html#method-H5T-copy'><code>H5T$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_attr"><a href='../../hdf5r/html/H5T.html#method-H5T-create_attr'><code>H5T$create_attr()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_attr_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-create_attr_by_name'><code>H5T$create_attr_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_reference"><a href='../../hdf5r/html/H5T.html#method-H5T-create_reference'><code>H5T$create_reference()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="detect_class"><a href='../../hdf5r/html/H5T.html#method-H5T-detect_class'><code>H5T$detect_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="equal"><a href='../../hdf5r/html/H5T.html#method-H5T-equal'><code>H5T$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_class"><a href='../../hdf5r/html/H5T.html#method-H5T-get_class'><code>H5T$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_create_plist"><a href='../../hdf5r/html/H5T.html#method-H5T-get_create_plist'><code>H5T$get_create_plist()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_native_type"><a href='../../hdf5r/html/H5T.html#method-H5T-get_native_type'><code>H5T$get_native_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_obj_name"><a href='../../hdf5r/html/H5T.html#method-H5T-get_obj_name'><code>H5T$get_obj_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_offset"><a href='../../hdf5r/html/H5T.html#method-H5T-get_offset'><code>H5T$get_offset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_order"><a href='../../hdf5r/html/H5T.html#method-H5T-get_order'><code>H5T$get_order()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_pad"><a href='../../hdf5r/html/H5T.html#method-H5T-get_pad'><code>H5T$get_pad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_precision"><a href='../../hdf5r/html/H5T.html#method-H5T-get_precision'><code>H5T$get_precision()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_size"><a href='../../hdf5r/html/H5T.html#method-H5T-get_size'><code>H5T$get_size()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="is_committed"><a href='../../hdf5r/html/H5T.html#method-H5T-is_committed'><code>H5T$is_committed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="is_vlen"><a href='../../hdf5r/html/H5T.html#method-H5T-is_vlen'><code>H5T$is_vlen()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="obj_info"><a href='../../hdf5r/html/H5T.html#method-H5T-obj_info'><code>H5T$obj_info()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="print"><a href='../../hdf5r/html/H5T.html#method-H5T-print'><code>H5T$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_offset"><a href='../../hdf5r/html/H5T.html#method-H5T-set_offset'><code>H5T$set_offset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_order"><a href='../../hdf5r/html/H5T.html#method-H5T-set_order'><code>H5T$set_order()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_pad"><a href='../../hdf5r/html/H5T.html#method-H5T-set_pad'><code>H5T$set_pad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_precision"><a href='../../hdf5r/html/H5T.html#method-H5T-set_precision'><code>H5T$set_precision()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_size"><a href='../../hdf5r/html/H5T.html#method-H5T-set_size'><code>H5T$set_size()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="to_text"><a href='../../hdf5r/html/H5T.html#method-H5T-to_text'><code>H5T$to_text()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_COMPOUND-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_COMPOUND-initialize}{}}}
\subsection{\code{H5T_COMPOUND$new()}}{
  Create at compound type that is the HDF5 equivalent of a table
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_COMPOUND$new(labels, dtypes, size = NULL, offset = NULL, id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{labels}}{The labels of the columns of the compound object}
      \item{\code{dtypes}}{The datatypes of the columns of the object; this is usually a list of objects of class \code{\link{H5T}}}
      \item{\code{size}}{The size of each datatype; if \code{NULL}, automatically inferred}
      \item{\code{offset}}{The offset where each datatype starts; can be different from the sum of the individual sizes so that datatypes are aligned with memory addresses. If \code{NULL}, inferred automatically}
      \item{\code{id}}{Internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_COMPOUND-pack"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_COMPOUND-pack}{}}}
\subsection{\code{H5T_COMPOUND$pack()}}{
  This function implements the HDF5-API function H5Tpack. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_COMPOUND$pack()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_COMPOUND-get_cpd_types"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_COMPOUND-get_cpd_types}{}}}
\subsection{\code{H5T_COMPOUND$get_cpd_types()}}{
  Return \code{\link{H5T}} objects that represent the datatypes of the columns of the compound object. Returned as a list if more than 1
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_COMPOUND$get_cpd_types()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_COMPOUND-get_cpd_labels"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_COMPOUND-get_cpd_labels}{}}}
\subsection{\code{H5T_COMPOUND$get_cpd_labels()}}{
  Return the labels of the columns as a character vector
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_COMPOUND$get_cpd_labels()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_COMPOUND-get_cpd_classes"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_COMPOUND-get_cpd_classes}{}}}
\subsection{\code{H5T_COMPOUND$get_cpd_classes()}}{
  Return the classes of the columns as an object of type \code{\link{factor_ext}}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_COMPOUND$get_cpd_classes()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_COMPOUND-get_cpd_offsets"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_COMPOUND-get_cpd_offsets}{}}}
\subsection{\code{H5T_COMPOUND$get_cpd_offsets()}}{
  Return the offsets of the datatypes
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_COMPOUND$get_cpd_offsets()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_COMPOUND-describe"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_COMPOUND-describe}{}}}
\subsection{\code{H5T_COMPOUND$describe()}}{
  Print a detailed description of the datatype; this is experimental
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_COMPOUND$describe()}
    \if{html}{\out{</div>}}
  }
}

}
