% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\name{H5T_extractID}
\alias{H5T_extractID}
\title{Extract HDF5-ids and return as a vector}
\usage{
H5T_extractID(dtype_ids)
}
\arguments{
\item{dtype_ids}{Ids itself or class objects to extract the ids from.}
}
\value{
The ids as a integer64-vector
}
\description{
Extract HDF5-ids and return as a vector
}
\details{
Internal helper function. Given a list of id-type-objects, it extracts the
ids itself and returns them as a vector
}
\author{
Holger Hoefling
}

