% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\name{H5T_factory}
\alias{H5T_factory}
\title{Turn ids into  \code{\link{H5T-class}} objects}
\usage{
H5T_factory(ids, do_copy = FALSE)
}
\arguments{
\item{ids}{The ids to return as objects}

\item{do_copy}{Should the ids be copied using HDF5s internal copy mechanism}
}
\value{
An object of type R6 class \code{\link{H5T-class}}
}
\description{
Turn ids into  \code{\link{H5T-class}} objects
}
\details{
Gets the HDF5 internal class for an id and dispactches them so that the correct
R6-classes are being created.
}
\author{
Holger Hoefling
}
\keyword{internal}
