% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\docType{class}
\name{H5T_INTEGER}
\alias{H5T_INTEGER}
\alias{H5T_BITFIELD}
\alias{H5T_BITFIELD-class}
\title{Class for HDF5 integer-datatypes.}
\value{
Object of class \code{\link[=H5T_INTEGER-class]{H5T_INTEGER}}.
}
\description{
Inherits from class \code{\link[=H5T-class]{H5T}}.
}
\seealso{
H5Class_overview, \code{\link[=H5T-class]{H5T}}
}
\author{
Holger Hoefling
}
