% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\docType{class}
\name{H5T_LOGICAL}
\alias{H5T_LOGICAL}
\title{Class for HDF5 logical datatypes. This is an enum with the 3 values FALSE, TRUE and NA mapped on values 0, 1 and 2.
Is transparently mapped onto a logical variable}
\value{
Object of class \code{\link[=H5T_LOGICAL-class]{H5T_LOGICAL}}.
}
\description{
Inherits from class \code{\link[=H5T-class]{H5T}}.
}
\seealso{
H5Class_overview, \code{\link[=H5T-class]{H5T}}, \code{\link[=H5T_ENUM-class]{H5T_ENUM}}
}
\author{
Holger Hoefling
}
