% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\docType{class}
\name{H5T_LOGICAL}
\alias{H5T_LOGICAL}
\alias{H5T_LOGICAL-class}
\title{Class for HDF5 logical datatypes. This is an enum with the 3 values FALSE, TRUE and NA mapped on values 0, 1 and 2.
Is transparently mapped onto a logical variable}
\value{
Object of class \code{\link[=H5T_LOGICAL]{H5T_LOGICAL}}.
}
\description{
Inherits from class \code{\link[=H5T]{H5T}}.
}
\seealso{
\code{\link[=H5T]{H5T}}, \code{\link[=H5T_ENUM]{H5T_ENUM}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5T]{H5T}} -> \code{\link[hdf5r:H5T_INTEGER]{H5T_INTEGER}} -> \code{\link[hdf5r:H5T_ENUM]{H5T_ENUM}} -> \code{H5T_LOGICAL}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5T_LOGICAL-initialize}{\code{H5T_LOGICAL$new()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete'><code>H5T$attr_delete()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete_by_idx'><code>H5T$attr_delete_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete_by_name'><code>H5T$attr_delete_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_exists"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_exists'><code>H5T$attr_exists()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_exists_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_exists_by_name'><code>H5T$attr_exists_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_get_number"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_get_number'><code>H5T$attr_get_number()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_info_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_info_by_idx'><code>H5T$attr_info_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_info_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_info_by_name'><code>H5T$attr_info_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_name_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_name_by_idx'><code>H5T$attr_name_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open'><code>H5T$attr_open()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open_by_idx'><code>H5T$attr_open_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open_by_name'><code>H5T$attr_open_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_rename"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_rename'><code>H5T$attr_rename()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_rename_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_rename_by_name'><code>H5T$attr_rename_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="copy"><a href='../../hdf5r/html/H5T.html#method-H5T-copy'><code>H5T$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_attr"><a href='../../hdf5r/html/H5T.html#method-H5T-create_attr'><code>H5T$create_attr()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_attr_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-create_attr_by_name'><code>H5T$create_attr_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_reference"><a href='../../hdf5r/html/H5T.html#method-H5T-create_reference'><code>H5T$create_reference()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="detect_class"><a href='../../hdf5r/html/H5T.html#method-H5T-detect_class'><code>H5T$detect_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="equal"><a href='../../hdf5r/html/H5T.html#method-H5T-equal'><code>H5T$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_class"><a href='../../hdf5r/html/H5T.html#method-H5T-get_class'><code>H5T$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_create_plist"><a href='../../hdf5r/html/H5T.html#method-H5T-get_create_plist'><code>H5T$get_create_plist()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_native_type"><a href='../../hdf5r/html/H5T.html#method-H5T-get_native_type'><code>H5T$get_native_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_obj_name"><a href='../../hdf5r/html/H5T.html#method-H5T-get_obj_name'><code>H5T$get_obj_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_offset"><a href='../../hdf5r/html/H5T.html#method-H5T-get_offset'><code>H5T$get_offset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_order"><a href='../../hdf5r/html/H5T.html#method-H5T-get_order'><code>H5T$get_order()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_pad"><a href='../../hdf5r/html/H5T.html#method-H5T-get_pad'><code>H5T$get_pad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_precision"><a href='../../hdf5r/html/H5T.html#method-H5T-get_precision'><code>H5T$get_precision()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_size"><a href='../../hdf5r/html/H5T.html#method-H5T-get_size'><code>H5T$get_size()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="is_committed"><a href='../../hdf5r/html/H5T.html#method-H5T-is_committed'><code>H5T$is_committed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="is_vlen"><a href='../../hdf5r/html/H5T.html#method-H5T-is_vlen'><code>H5T$is_vlen()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="obj_info"><a href='../../hdf5r/html/H5T.html#method-H5T-obj_info'><code>H5T$obj_info()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="print"><a href='../../hdf5r/html/H5T.html#method-H5T-print'><code>H5T$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_offset"><a href='../../hdf5r/html/H5T.html#method-H5T-set_offset'><code>H5T$set_offset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_order"><a href='../../hdf5r/html/H5T.html#method-H5T-set_order'><code>H5T$set_order()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_pad"><a href='../../hdf5r/html/H5T.html#method-H5T-set_pad'><code>H5T$set_pad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_precision"><a href='../../hdf5r/html/H5T.html#method-H5T-set_precision'><code>H5T$set_precision()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="to_text"><a href='../../hdf5r/html/H5T.html#method-H5T-to_text'><code>H5T$to_text()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T_INTEGER" data-id="get_sign"><a href='../../hdf5r/html/H5T_INTEGER.html#method-H5T_INTEGER-get_sign'><code>H5T_INTEGER$get_sign()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T_INTEGER" data-id="set_sign"><a href='../../hdf5r/html/H5T_INTEGER.html#method-H5T_INTEGER-set_sign'><code>H5T_INTEGER$set_sign()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T_ENUM" data-id="describe"><a href='../../hdf5r/html/H5T_ENUM.html#method-H5T_ENUM-describe'><code>H5T_ENUM$describe()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T_ENUM" data-id="get_labels"><a href='../../hdf5r/html/H5T_ENUM.html#method-H5T_ENUM-get_labels'><code>H5T_ENUM$get_labels()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T_ENUM" data-id="get_super"><a href='../../hdf5r/html/H5T_ENUM.html#method-H5T_ENUM-get_super'><code>H5T_ENUM$get_super()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T_ENUM" data-id="get_values"><a href='../../hdf5r/html/H5T_ENUM.html#method-H5T_ENUM-get_values'><code>H5T_ENUM$get_values()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T_ENUM" data-id="set_size"><a href='../../hdf5r/html/H5T_ENUM.html#method-H5T_ENUM-set_size'><code>H5T_ENUM$set_size()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_LOGICAL-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_LOGICAL-initialize}{}}}
\subsection{\code{H5T_LOGICAL$new()}}{
  Create a logical datatype. This is internally represented by an ENUM-type
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_LOGICAL$new(include_NA = TRUE, id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{include_NA}}{Logical; should the enumeration include a level for \code{NA}.}
      \item{\code{id}}{Internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

}
