% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\docType{class}
\name{H5T_STRING}
\alias{H5T_STRING}
\alias{H5T_STRING-class}
\title{Class for HDF5 string datatypes.}
\value{
Object of class \code{\link[=H5T_STRING]{H5T_STRING}}.
}
\description{
Inherits from class \code{\link[=H5T]{H5T}}.
}
\examples{
# fixed width string
str_flen <- H5T_STRING$new(size=100)
str_flen$is_vlen()
str_flen

# variable length string
str_vlen <- H5T_STRING$new(size=Inf)
str_vlen$is_vlen()
str_vlen
}
\seealso{
\code{\link[=H5T]{H5T}}
}
\author{
Holger Hoefling
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{\link[hdf5r:H5T]{H5T}} -> \code{H5T_STRING}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5T_STRING-initialize}{\code{H5T_STRING$new()}}
    \item \href{#method-H5T_STRING-get_size}{\code{H5T_STRING$get_size()}}
    \item \href{#method-H5T_STRING-get_cset}{\code{H5T_STRING$get_cset()}}
    \item \href{#method-H5T_STRING-set_cset}{\code{H5T_STRING$set_cset()}}
    \item \href{#method-H5T_STRING-set_strpad}{\code{H5T_STRING$set_strpad()}}
    \item \href{#method-H5T_STRING-get_strpad}{\code{H5T_STRING$get_strpad()}}
    \item \href{#method-H5T_STRING-describe}{\code{H5T_STRING$describe()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete'><code>H5T$attr_delete()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete_by_idx'><code>H5T$attr_delete_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_delete_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_delete_by_name'><code>H5T$attr_delete_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_exists"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_exists'><code>H5T$attr_exists()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_exists_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_exists_by_name'><code>H5T$attr_exists_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_get_number"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_get_number'><code>H5T$attr_get_number()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_info_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_info_by_idx'><code>H5T$attr_info_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_info_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_info_by_name'><code>H5T$attr_info_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_name_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_name_by_idx'><code>H5T$attr_name_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open'><code>H5T$attr_open()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open_by_idx"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open_by_idx'><code>H5T$attr_open_by_idx()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_open_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_open_by_name'><code>H5T$attr_open_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_rename"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_rename'><code>H5T$attr_rename()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="attr_rename_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-attr_rename_by_name'><code>H5T$attr_rename_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="copy"><a href='../../hdf5r/html/H5T.html#method-H5T-copy'><code>H5T$copy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_attr"><a href='../../hdf5r/html/H5T.html#method-H5T-create_attr'><code>H5T$create_attr()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_attr_by_name"><a href='../../hdf5r/html/H5T.html#method-H5T-create_attr_by_name'><code>H5T$create_attr_by_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="create_reference"><a href='../../hdf5r/html/H5T.html#method-H5T-create_reference'><code>H5T$create_reference()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="detect_class"><a href='../../hdf5r/html/H5T.html#method-H5T-detect_class'><code>H5T$detect_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="equal"><a href='../../hdf5r/html/H5T.html#method-H5T-equal'><code>H5T$equal()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_class"><a href='../../hdf5r/html/H5T.html#method-H5T-get_class'><code>H5T$get_class()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_create_plist"><a href='../../hdf5r/html/H5T.html#method-H5T-get_create_plist'><code>H5T$get_create_plist()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_native_type"><a href='../../hdf5r/html/H5T.html#method-H5T-get_native_type'><code>H5T$get_native_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_obj_name"><a href='../../hdf5r/html/H5T.html#method-H5T-get_obj_name'><code>H5T$get_obj_name()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_offset"><a href='../../hdf5r/html/H5T.html#method-H5T-get_offset'><code>H5T$get_offset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_order"><a href='../../hdf5r/html/H5T.html#method-H5T-get_order'><code>H5T$get_order()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_pad"><a href='../../hdf5r/html/H5T.html#method-H5T-get_pad'><code>H5T$get_pad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="get_precision"><a href='../../hdf5r/html/H5T.html#method-H5T-get_precision'><code>H5T$get_precision()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="is_committed"><a href='../../hdf5r/html/H5T.html#method-H5T-is_committed'><code>H5T$is_committed()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="is_vlen"><a href='../../hdf5r/html/H5T.html#method-H5T-is_vlen'><code>H5T$is_vlen()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="obj_info"><a href='../../hdf5r/html/H5T.html#method-H5T-obj_info'><code>H5T$obj_info()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="print"><a href='../../hdf5r/html/H5T.html#method-H5T-print'><code>H5T$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_offset"><a href='../../hdf5r/html/H5T.html#method-H5T-set_offset'><code>H5T$set_offset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_order"><a href='../../hdf5r/html/H5T.html#method-H5T-set_order'><code>H5T$set_order()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_pad"><a href='../../hdf5r/html/H5T.html#method-H5T-set_pad'><code>H5T$set_pad()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_precision"><a href='../../hdf5r/html/H5T.html#method-H5T-set_precision'><code>H5T$set_precision()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="set_size"><a href='../../hdf5r/html/H5T.html#method-H5T-set_size'><code>H5T$set_size()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5T" data-id="to_text"><a href='../../hdf5r/html/H5T.html#method-H5T-to_text'><code>H5T$to_text()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_STRING-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_STRING-initialize}{}}}
\subsection{\code{H5T_STRING$new()}}{
  Create a string datatype
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_STRING$new(type = c("c", "fortran"), size = 1, id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{The type to use. For a dataspace one of \code{"simple"}, \code{"scalar"} or \code{"null"}; otherwise an object of class \code{\link{H5T}}.}
      \item{\code{size}}{Size of the string object. Set to \code{Inf} for variable size strings}
      \item{\code{id}}{internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_STRING-get_size"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_STRING-get_size}{}}}
\subsection{\code{H5T_STRING$get_size()}}{
  Retrieves the length of the string, setting it to \code{Inf} it is of variable length. This function implements the HDF5-API function H5Tis_variable_str. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_STRING$get_size(variable_as_inf = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{variable_as_inf}}{Logical; should variable-length strings report their size as \code{Inf}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_STRING-get_cset"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_STRING-get_cset}{}}}
\subsection{\code{H5T_STRING$get_cset()}}{
  This function implements the HDF5-API function H5Tget_cset. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_STRING$get_cset()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_STRING-set_cset"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_STRING-set_cset}{}}}
\subsection{\code{H5T_STRING$set_cset()}}{
  This function implements the HDF5-API function H5Tset_cset. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_STRING$set_cset(cset = c("unknown", "UTF-8"))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{cset}}{The character set to use; one of the \code{H5T_CSET_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_STRING-set_strpad"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_STRING-set_strpad}{}}}
\subsection{\code{H5T_STRING$set_strpad()}}{
  This function implements the HDF5-API function H5Tset_strpad. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_STRING$set_strpad(strpad)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{strpad}}{The string padding to use; one of the \code{H5T_STR_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_STRING-get_strpad"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_STRING-get_strpad}{}}}
\subsection{\code{H5T_STRING$get_strpad()}}{
  This function implements the HDF5-API function H5Tget_strpad. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_STRING$get_strpad()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T_STRING-describe"></a>}}
\if{latex}{\out{\hypertarget{method-H5T_STRING-describe}{}}}
\subsection{\code{H5T_STRING$describe()}}{
  Print a detailed description of the datatype; this is experimental
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T_STRING$describe()}
    \if{html}{\out{</div>}}
  }
}

}
