% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5constants.R
\name{h5types}
\alias{h5types}
\title{These are all types that are used in HDF5}
\description{
HDF5 provides many native datatypes. These are all stored in the \code{h5types}
environment. An overview of all available types can be seen using \code{h5types$overview}.
Any specific type can be accessed using the \code{$}-operator. See also the examples below.
}
\examples{
h5types$overview
h5types$H5T_NATIVE_INT
h5types$H5T_NATIVE_DOUBLE
}
\author{
Holger Hoefling
}
