% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{h5version}
\alias{h5version}
\title{Return the version of the HDF5-API}
\usage{
h5version(verbose = TRUE)
}
\arguments{
\item{verbose}{Should the information be printed to the screen as well}
}
\value{
Version of the underlying HDF5 api as a string
}
\description{
Return the version of the HDF5-API
}
\details{
Return the version of the HDF5-API and print it to the screen if requested
}
\author{
Holger Hoefling
}
