% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5r.R
\docType{package}
\name{hdf5r-package}
\alias{hdf5r}
\alias{hdf5r-package}
\title{hdf5r: A package to provide an interface to hdf5 from R}
\description{
A package that allows to interface with the HDF5 C-library. Provides access to most of its
functionality from inside R using R6 classes. For more details please see the README at
the github page \url{https://github.com/hhoeflin/hdf5r}.
}
\examples{
test_file <- tempfile(fileext=".h5")
file.h5 <- H5File$new(test_file, mode="w")

data(cars)
file.h5$create_group("test")
file.h5[["test/cars"]] <- cars
cars_ds <- file.h5[["test/cars"]]
h5attr(cars_ds, "rownames") <- rownames(cars)

# Close the file at the end
# the 'close' method closes only the file-id, but leaves object inside the file open
# This may prevent re-opening of the file. 'close_all' closes the file and all objects in it
file.h5$close_all()
# now re-open it 
file.h5 <- H5File$new(test_file, mode="r+")

# lets look at the content
file.h5$ls(recursive=TRUE)

cars_ds <- file.h5[["test/cars"]]
# note that for now tables in HDF5 are 1-dimensional, not 2-dimensional
mycars <- cars_ds[]
h5attr_names(cars_ds)
h5attr(cars_ds, "rownames")

file.h5$close_all()
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hhoeflin.github.io/hdf5r/}
  \item \url{https://github.com/hhoeflin/hdf5r/}
  \item Report bugs at \url{https://github.com/hhoeflin/hdf5r/issues}
}

}
\author{
\strong{Maintainer}: Holger Hoefling \email{hhoeflin@gmail.com}

Authors:
\itemize{
  \item Holger Hoefling \email{hhoeflin@gmail.com}
  \item Mario Annau \email{mario.annau@gmail.com}
}

Other contributors:
\itemize{
  \item Novartis Institute for BioMedical Research (NIBR) [copyright holder]
}

}
\keyword{internal}
