% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5r.R
\docType{package}
\name{hdf5r}
\alias{hdf5r}
\alias{hdf5r-package}
\title{hdf5r: A package to provide an interface to hdf5 from R}
\description{
A package that allows to interface with the HDF5 C-library. Proivdes acces to most of its
functionality from inside R using R6 classes.
}

