% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{hyperslab}
\alias{hyperslab}
\title{Create a hyperslab object for simpler internal processing}
\usage{
hyperslab(start, count, stride, block)
}
\arguments{
\item{start}{The start value for an HDF5 hyperslab}

\item{count}{The count value for an HDF5 hyperslab}

\item{stride}{The stride value for an HDF5 hyperslab}

\item{block}{The block value for an HDF5 hyperslab}
}
\value{
A structure of class \code{hyperslab} that wraps a data.frame with columns  \code{start}, \code{count},
\code{stride} and \code{block}.
}
\description{
Create a hyperslab object for simpler internal processing
}
\details{
Simple wrapper around a data.frame with a specific pattern.
}
\author{
Holger Hoefling
}
\keyword{internal}

