% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{hyperslab_to_points}
\alias{hyperslab_to_points}
\title{Single hyperslab dimension to explicit vector}
\usage{
hyperslab_to_points(hyperslab)
}
\arguments{
\item{hyperslab}{a length 4 vector describing the start, count, stride and block component of a single dimension of a hyperslab}
}
\value{
An explicit vector describing the points in the hyperslab dimension
}
\description{
Single hyperslab dimension to explicit vector
}
\details{
Uses the information of a hyperslab and turns it into an explicit vector.
}
\author{
Holger Hoefling
}
\keyword{internal}
