% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_ext.R
\name{is.factor_ext}
\alias{is.factor_ext}
\title{Check if it is an extended factor}
\usage{
is.factor_ext(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical of length 1
}
\description{
Check if it is an extended factor
}
\details{
Check if it is an object of class \code{factor_ext}
}
\author{
Holger Hoefling
}

