% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{is.H5R}
\alias{is.H5R}
\alias{is.H5R_DATASET_REGION}
\alias{is.H5R_OBJECT}
\title{Checks if object is of a HDF5-Reference class}
\usage{
is.H5R(x)

is.H5R_OBJECT(x)

is.H5R_DATASET_REGION(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
Logical
}
\description{
Checks if object is of a HDF5-Reference class
}
\details{
Simple check if it inherits from \code{H5R}, \code{H5R_OBJECT} or \code{H5R_DATSET_REGION}.
}
\author{
Holger Hoefling
}

