% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5File.R
\name{is_hdf5}
\alias{is_hdf5}
\title{Check if a file is an HDF5 file}
\usage{
is_hdf5(name)
}
\arguments{
\item{name}{The name of the file to check}
}
\value{
Logical, TRUE if file is of type HDF5
}
\description{
Check if a file is an HDF5 file
}
\details{
Uses the HDF5 function \code{H5Fis_hdf5} to check if a file is of type HDF5.
}
\author{
Holger Hoefling
}
