% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{length.H5R}
\alias{length.H5R}
\title{Retrieve length of an object}
\usage{
\method{length}{H5R}(x)
}
\arguments{
\item{x}{The object to get the length of}
}
\value{
Length of the object
}
\description{
Retrieve length of an object
}
\details{
Returns the slot that describes the length of the object
}
\author{
Holger Hoefling
}

