% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{match.call.withDef}
\alias{match.call.withDef}
\title{Match arguments in a call to function and add default values}
\usage{
match.call.withDef(definition, call)
}
\arguments{
\item{definition}{Definition of the function to match against}

\item{call}{The call that should be matched}
}
\value{
A call with named arguments and default values
}
\description{
Match arguments in a call to function and add default values
}
\details{
Given the definition of a function and a call, it matches the arguments
so that they are named and inserts any default argument values where
those are missing
}
\author{
Holger Hoefling
}
\keyword{internal}

