% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{names.H5R}
\alias{names.H5R}
\title{Return the name-vector of an object}
\usage{
\method{names}{H5R}(x)
}
\arguments{
\item{x}{The object to get the names of}
}
\value{
A character vector
}
\description{
Return the name-vector of an object
}
\details{
Returns the name vector of the object
}
\author{
Holger Hoefling
}
