% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{print.data.frame_ext}
\alias{print.data.frame_ext}
\title{Print a data frame with extended factor objects}
\usage{
\method{print}{data.frame_ext}(x, ...)
}
\arguments{
\item{x}{The \code{data.frame_ext} object to print; Is returned by ls from \code{\link{H5File-class}}
and \code{\link{H5Group-class}} and this function allows for petter printing of \code{\link{factor_ext}}
so that the label instead of the value is printed.}

\item{...}{Parameters to be passed on directly to \code{\link{print.data.frame}}}
}
\value{
The object to print itself, invisibly
}
\description{
Print a data frame that includes extended factor objects
}
\details{
The regular print function for data-frames has special methods built-in for factors so that
the label is printed instead of the constant. This function is intended to provide the same functionality
for data frames with extended factors, by adding the class \code{data.frame_ext} to the class vector.
}
\author{
Holger Hoefling
}
