% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{print.H5R}
\alias{print.H5R}
\title{Printing of an object of class \code{h5R}}
\usage{
\method{print}{H5R}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Ignored}
}
\value{
The printed text as an invisible character vector
}
\description{
Printing of an object of class \code{h5R}
}
\details{
Rudimentary printing of information of the object, mainly its length
}
\author{
Holger Hoefling
}
