% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{print_attributes}
\alias{print_attributes}
\title{Print attributes}
\usage{
print_attributes(obj, max_to_print)
}
\arguments{
\item{obj}{The obj for which to print the attributes}

\item{max_to_print}{Maximum number of attributes to print}
}
\value{
Invisible NULL
}
\description{
Print attributes
}
\details{
Prints the names of the attributes up to a given maximum number
}
\author{
Holger Hoefling
}
\keyword{internal}
