% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{print_class_id}
\alias{print_class_id}
\title{Print the class and ID}
\usage{
print_class_id(obj, is_valid)
}
\arguments{
\item{obj}{The object for which to print the class and id}

\item{is_valid}{is the object valid}
}
\value{
invisible NULL
}
\description{
Print the class and ID
}
\details{
Used by the print-methods
}
\author{
Holger Hoefling
}
\keyword{internal}
