% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{print_listing}
\alias{print_listing}
\title{Print listing}
\usage{
print_listing(obj, max_to_print)
}
\arguments{
\item{obj}{Object for which to print the listing}

\item{max_to_print}{Maximum number of listing items to print}
}
\value{
Invisible NULL
}
\description{
Print listing
}
\details{
Prints a smaller part of the \code{ls} output of an object, up to a maximum number
}
\author{
Holger Hoefling
}
\keyword{internal}
