% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_UI.R
\name{regularity_eval_to_selection}
\alias{regularity_eval_to_selection}
\title{Turn regulation evaluation into a selection for a space object}
\usage{
regularity_eval_to_selection(reg_eval_res)
}
\arguments{
\item{reg_eval_res}{The result of the \code{args_regularity_evaluation} function}
}
\value{
Returns an object with either the point-matrix or the hyperslab-selection array. The resulting object is
of class \code{point_selection} or \code{hyperslab_selection}.
}
\description{
Turn regulation evaluation into a selection for a space object
}
\details{
Analyzes the results of the regularity evaluation of each dimension and checks if
it needs to be into a hyperslab-selection or a point-selection. A hyperslab selection will be chosen
whenever there are significantly less of it than the number of selected points. The ratio is determined
by the option \code{hdf5r.point_to_hyperslab_ratio}. If this is 1, then always hyperslabs will be used.
}
\author{
Holger Hoefling
}
\keyword{internal}

