% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5D.R
\name{standalone_H5D_get_type}
\alias{standalone_H5D_get_type}
\title{Get the id of a type of the dataset}
\usage{
standalone_H5D_get_type(h5d_id, native = TRUE)
}
\arguments{
\item{h5d_id}{The id of the dataset to get the type from}

\item{native}{Should it be ensured that it is a native type}
}
\value{
An id; the user has to ensure that the id is eventually closed
}
\description{
Get the id of a type of the dataset
}
\details{
A function that just returns an id; it is written standalone so that
one can use it to avoid the creation of R6 classes that be a considerable overhead in
certain circumstances
}
\author{
Holger Hoefling
}
\keyword{internal}
