% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5S.R
\name{standalone_H5S_select_multiple_hyperslab}
\alias{standalone_H5S_select_multiple_hyperslab}
\title{Select multiple hyperslabs in a space}
\usage{
standalone_H5S_select_multiple_hyperslab(id, hyperslab_array)
}
\arguments{
\item{id}{The id of the space}

\item{hyperslab_array}{The array with the hyperslabs. Is of dimension num_dim x num_hyperslabs x 4. With teh elements
being start, count, stride and block}
}
\value{
\code{NULL}. The space has been manipulated as a side effect
}
\description{
Select multiple hyperslabs in a space
}
\details{
Selects multiple hyperslabs in a space. Before the selection, the space selection will be cleared.
}
\author{
Holger Hoefling
}
\keyword{internal}
