% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{[.H5R}
\alias{[.H5R}
\alias{[<-.H5R}
\title{Subset or assign into a vector/array}
\usage{
\method{[}{H5R}(x, i, j, ..., drop = TRUE)

\method{[}{H5R}(x, i, ...) <- value
}
\arguments{
\item{x}{The object to subset}

\item{i}{First dimension}

\item{j}{Second dimension}

\item{...}{Any other dimensions}

\item{drop}{Should dimensions of size 1 be dropped; LOGICAL}

\item{value}{The value in an assignment}
}
\value{
A corresponding object that is the subset of the vector
}
\description{
Subset or assign into a vector/array
}
\details{
The standard array/vector subsetting functions
}
\author{
Holger Hoefling
}

