% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5R.R
\name{t.H5R}
\alias{t.H5R}
\title{Transpose a HDF5-Reference matrix}
\usage{
\method{t}{H5R}(x)
}
\arguments{
\item{x}{The matrix to transpose; will be cloned}
}
\value{
A transposed version of the HDF5-Reference matrix; a clone
}
\description{
Transpose a HDF5-Reference matrix
}
\details{
Clones the matrix and returns a transpose of the clone
}
\author{
Holger Hoefling
}
