% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\name{text_to_dtype}
\alias{text_to_dtype}
\title{Convert a text description to a datatype}
\usage{
text_to_dtype(text, lang_type = h5const$H5LT_DDL)
}
\arguments{
\item{text}{The text to convert to the datatype}

\item{lang_type}{The type of language to use; currently only \code{H5LT_DDL} is supported.}
}
\value{
A datatype corresponding to the text with the appropriate class inheriting from \code{\link{H5T}}.
}
\description{
Convert a text description to a datatype
}
\details{
Converts a text to a datatype using the HDF5 function H5LT_text_to_dtype. Documentation can be found at 
\url{https://portal.hdfgroup.org/display/HDF5/H5LT_TEXT_TO_DTYPE}.
}
\author{
Holger Hoefling
}
