% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5constants.R
\name{$.types_env}
\alias{$.types_env}
\alias{[[.types_env}
\title{Retrieving a copy of a type}
\usage{
\method{$}{types_env}(x, name)

\method{[[}{types_env}(x, name)
}
\arguments{
\item{x}{The environment to request it from}

\item{name}{The name of the type that is requested}
}
\value{
Returns an object that is a copy of a type that was requested
}
\description{
Retrieving a copy of a type
}
\details{
The types are stored in the environment \code{\link{h5types}}. These types should not be accessed
directly. Therefor, the \code{$}-operator is overloaded to ensure that every type that is accessed is a copy of the
original type
}
\author{
Holger Hoefling
}
