% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_ext.R
\name{values}
\alias{values}
\alias{values.factor_ext}
\title{Extract values from an object}
\usage{
values(x, ...)

\method{values}{factor_ext}(x, ...)
}
\arguments{
\item{x}{The object for which to extract the factors. Currently only implemented for \code{factor_ext}}

\item{...}{Currently ignored}
}
\value{
Return the integer values used for the factor
}
\description{
Extract values from an object
}
\details{
A \code{factor_ext} variable does not use consequitive value. This is a generic function and it is currently only
overloaded for \code{factor_ext} classes. It extract the values used in the \code{factor_ext} object.
}
\author{
Holger Hoefling
}
