% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertRoundTrip}
\alias{convertRoundTrip}
\title{Round-trip of converting data to HDF5 and back to R}
\usage{
convertRoundTrip(Robj, dtype, nelem = length(Robj),
  flags = h5const$H5TOR_CONV_INT64_NOLOSS)
}
\arguments{
\item{Robj}{The object to convert}

\item{dtype}{The datatype to convert it into}

\item{nelem}{The number of elements in the object}

\item{flags}{conversion flags from HDF5 to R}
}
\value{
A list with input, number of elements, raw vector for intermediate storage and output
}
\description{
Round-trip of converting data to HDF5 and back to R
}
\details{
Take an R-object, convert it to HDF5, convert it back and return input, output and intermediate steps
This is mainly intended for use in tests.
}
\author{
Holger Hoefling
}
\keyword{internal}

