% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{extract_dim}
\alias{extract_dim}
\title{Extract the dimension of a space and datatype}
\usage{
extract_dim(space, dtype)
}
\arguments{
\item{space}{The space with the selection that was used to read the dataset}

\item{dtype}{The datatype of the dataset}
}
\value{
\code{x}, but with a new dimension attribute
}
\description{
Get the correct dimensions for a space and datatype
}
\details{
This function uses the space and the selection in it to get the correct dimension
for the resulting object (but without dropping dimensions). Furthermore, if the
datatype is an array, those dimensions are correctly determined as well.

Internal use only; currently unused
}
\author{
Holger Hoefling
}
\keyword{internal}
