% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5wrapper.R
\name{h5-wrapper}
\alias{h5-wrapper}
\alias{h5file}
\alias{createGroup}
\alias{openLocation}
\alias{createDataSet}
\alias{readDataSet}
\alias{h5close}
\alias{h5flush}
\alias{existsGroup}
\alias{is.h5file}
\alias{extendDataSet}
\alias{rbind.H5D}
\alias{cbind.H5D}
\alias{c.H5D}
\alias{h5unlink}
\alias{list.attributes}
\title{Wrapper functions to provide a \strong{h5} compatible interface.}
\usage{
h5file(...)

createGroup(object, name, ...)

openLocation(object, name)

createDataSet(object, name, ...)

readDataSet(object)

h5close(object)

h5flush(object)

existsGroup(object, name)

is.h5file(name)

extendDataSet(object, dims)

\method{rbind}{H5D}(x, mat, ..., deparse.level = 1)

\method{cbind}{H5D}(x, mat, ..., deparse.level = 1)

\method{c}{H5D}(x, ...)

h5unlink(object, name)

list.attributes(object)
}
\arguments{
\item{...}{Additional parameters passed to \code{create_group} or \code{h5file}.}

\item{object}{\code{CommonFG}; Object implementing the CommonFG Interface (e.g. \code{\link{H5File}}, \code{\link{H5Group}}).}

\item{name}{Name of the group to create.}

\item{dims}{numeric; Dimension vector to which dataset should be extended.}

\item{x}{An object of class H5D; the dataset to add rows or columns to; Needs to be a matrix}

\item{mat}{The matrix to add to x}

\item{deparse.level}{Set to 1; ignored otherwise; only present as required by generic}
}
\description{
Since it has been decided to join forces on the hdf5r package and deprecate
h5 we wanted to make the transition for users as smooth as possible.
Additionally, we could transfer almost all testcases to hdf5r and improve
test coverage even more.
}
\details{
Below you can find a list of all \strong{h5} function including \strong{hdf5r} mappings.
}
