% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5wrapper.R
\name{h5file}
\alias{h5file}
\alias{createGroup}
\alias{openLocation}
\alias{createDataSet}
\alias{readDataSet}
\alias{h5close}
\alias{h5flush}
\alias{existsGroup}
\alias{is.h5file}
\alias{list.groups}
\alias{list.datasets}
\alias{extendDataSet}
\alias{rbind.H5D}
\alias{cbind.H5D}
\alias{c.H5D}
\alias{h5unlink}
\alias{list.attributes}
\title{h5 wrapper functions}
\usage{
h5file(...)

createGroup(object, name, ...)

openLocation(object, name)

createDataSet(object, name, ...)

readDataSet(object)

h5close(object)

h5flush(object)

existsGroup(object, name)

is.h5file(name)

list.groups(object, path = "/", recursive = TRUE, ...)

list.datasets(object, path = "/", recursive = TRUE, ...)

extendDataSet(object, dims)

\method{rbind}{H5D}(x, mat, ..., deparse.level = 1)

\method{cbind}{H5D}(x, mat, ..., deparse.level = 1)

\method{c}{H5D}(x, ...)

h5unlink(object, name)

list.attributes(object)
}
\arguments{
\item{...}{Additional Parameters passed to \code{file$create_group}, \code{file$create_dataset} or ignored for \code{cbind} and \code{rbind}.}

\item{object}{H5RefClass; H5 Reference Class to be used}

\item{name}{character; Group/Filename to be created}

\item{path}{character; Path named to be used for iteration.}

\item{recursive}{logical; Specify if object should be traversed recursively.}

\item{dims}{numeric; Dimension vector to which dataset should be extended.}

\item{x}{An object of class H5D; the dataset to add rows or columns to; Needs to be a matrix}

\item{mat}{The matrix to add to x}

\item{deparse.level}{Set to 1; ignored otherwise; only present as required by generic}
}
\description{
Wrapper functions to provide h5 compatible interface.
}
