% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5D.R
\docType{class}
\name{H5D}
\alias{H5D}
\alias{H5D-class}
\title{Class for representing HDF5 datasets}
\value{
Object of class \code{\link{H5D}}.
}
\description{
In HDF5, datasets can be located in a group (see \code{\link{H5Group}}) or at the
root of a file (see \code{\link{H5File}}). They can be created either with a pre-existing R-object
(arrays as well as data.frames are supported, but not lists or other complex objects), or by specifying
an explicit datatype (for available datatypes see \code{h5types$overview} as well as the dimension.
In addition, other features are supported such as transparent compression (for which a chunk-size can be selected).

In order to create a dataset, the \code{create_dataset} methods of either \code{\link{H5Group}} or
\code{\link{H5File}} should be used. Please see the documentation there for how to create them.

The most important parts of a dataset are the 
\describe{
  \item{Space}{The space of the dataset. It describes the dimension of the dataset as well as the maximum dimensions.
               Can be obtained using the \code{get_space} of the \code{\link{H5S}} object.}
  \item{Datatype}{The datatypes that is being used in the dataset. Can be obtained using the \code{get_type} method.
                  See \code{\link{H5T}} to get more information about using datatypes.}
}

In order to read and write datasets, the \code{read} and \code{write} methods are available. In addition, the standard way of using
\code{[} to access arrays is supported as well (see \code{\link{H5S_H5D_subset_assign}} for more help).

Other information/action of possible interest are
\describe{
  \item{Storage size}{The size of the dataset can be extracted using \code{get_storage_size}}
  \item{Size change}{The size of the dataset can be changed using the \code{set_extent} method}
}

Please also note the active methods
\describe{
  \item{dims}{Dimension of the dataset}
  \item{maxdims}{Maximum dimensions of the dataset}
  \item{chunk_dims}{Dimension of the chunks}
  \item{key_info}{Returns the space, type, property-list and dimensions}
}
}
\examples{
# First create a file to create datasets in it
fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")

# Show the 3 different ways how to create a dataset
file[["directly"]] <- matrix(1:10, ncol=2)
file$create_dataset("from_robj", matrix(1:10, ncol=2))
dset <- file$create_dataset("basic", dtype=h5types$H5T_NATIVE_INT,
             space=H5S$new("simple", dims=c(5, 2), maxdims=c(10,2)), chunk_dims=c(5,2))

# Different ways of reading the dataset
dset$read(args=list(1:5, 1))
dset$read(args=list(1:5, quote(expr=)))
dset$read(args=list(1:5, NULL))
dset[1:5, 1]
dset[1:5, ]
dset[1:5, NULL]

# Writing to the dataset
dset$write(args=list(1:3, 1:2), value=11:16)
dset[4:5, 1:2] <- -(1:4)
dset[,]

# Extract key information
dset$dims
dset$maxdims
dset$chunk_dims
dset$key_info
dset

file$close_all()
file.remove(fname)
}
\author{
Holger Hoefling
}
\section{Super class}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{H5D}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{dims}}{Get the dimension of the dataset}

    \item{\code{maxdims}}{Get the maximal dimension of the dataset}

    \item{\code{chunk_dims}}{Return the dimension of the chunks. NA if the dataset is not chunked}

    \item{\code{key_info}}{Returns the key types as a list, consisting of type, space, dataset_create_pl, type_size_raw, type_size_variable, dims and chunk_dims. type_size_raw versus variable differs for variable length types, which return \code{Inf} for type_size_variable and the underlying size for type_size_raw}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5D-get_obj_name}{\code{H5D$get_obj_name()}}
    \item \href{#method-H5D-create_attr}{\code{H5D$create_attr()}}
    \item \href{#method-H5D-attr_open}{\code{H5D$attr_open()}}
    \item \href{#method-H5D-create_attr_by_name}{\code{H5D$create_attr_by_name()}}
    \item \href{#method-H5D-attr_open_by_name}{\code{H5D$attr_open_by_name()}}
    \item \href{#method-H5D-attr_open_by_idx}{\code{H5D$attr_open_by_idx()}}
    \item \href{#method-H5D-attr_exists_by_name}{\code{H5D$attr_exists_by_name()}}
    \item \href{#method-H5D-attr_exists}{\code{H5D$attr_exists()}}
    \item \href{#method-H5D-attr_rename_by_name}{\code{H5D$attr_rename_by_name()}}
    \item \href{#method-H5D-attr_rename}{\code{H5D$attr_rename()}}
    \item \href{#method-H5D-attr_delete}{\code{H5D$attr_delete()}}
    \item \href{#method-H5D-attr_delete_by_name}{\code{H5D$attr_delete_by_name()}}
    \item \href{#method-H5D-attr_delete_by_idx}{\code{H5D$attr_delete_by_idx()}}
    \item \href{#method-H5D-attr_info_by_name}{\code{H5D$attr_info_by_name()}}
    \item \href{#method-H5D-attr_info_by_idx}{\code{H5D$attr_info_by_idx()}}
    \item \href{#method-H5D-attr_name_by_idx}{\code{H5D$attr_name_by_idx()}}
    \item \href{#method-H5D-attr_get_number}{\code{H5D$attr_get_number()}}
    \item \href{#method-H5D-flush}{\code{H5D$flush()}}
    \item \href{#method-H5D-get_filename}{\code{H5D$get_filename()}}
    \item \href{#method-H5D-initialize}{\code{H5D$new()}}
    \item \href{#method-H5D-get_space}{\code{H5D$get_space()}}
    \item \href{#method-H5D-get_space_status}{\code{H5D$get_space_status()}}
    \item \href{#method-H5D-get_type}{\code{H5D$get_type()}}
    \item \href{#method-H5D-get_create_plist}{\code{H5D$get_create_plist()}}
    \item \href{#method-H5D-get_access_plist}{\code{H5D$get_access_plist()}}
    \item \href{#method-H5D-get_offset}{\code{H5D$get_offset()}}
    \item \href{#method-H5D-get_storage_size}{\code{H5D$get_storage_size()}}
    \item \href{#method-H5D-vlen_get_buf_size}{\code{H5D$vlen_get_buf_size()}}
    \item \href{#method-H5D-vlen_reclaim}{\code{H5D$vlen_reclaim()}}
    \item \href{#method-H5D-read_low_level}{\code{H5D$read_low_level()}}
    \item \href{#method-H5D-read}{\code{H5D$read()}}
    \item \href{#method-H5D-write_low_level}{\code{H5D$write_low_level()}}
    \item \href{#method-H5D-write}{\code{H5D$write()}}
    \item \href{#method-H5D-set_extent}{\code{H5D$set_extent()}}
    \item \href{#method-H5D-get_fill_value}{\code{H5D$get_fill_value()}}
    \item \href{#method-H5D-create_reference}{\code{H5D$create_reference()}}
    \item \href{#method-H5D-print}{\code{H5D$print()}}
    \item \href{#method-H5D-obj_info}{\code{H5D$obj_info()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_obj_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_obj_name}{}}}
\subsection{\code{H5D$get_obj_name()}}{
  This function implements the HDF5-API function H5Iget_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_obj_name()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-create_attr"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-create_attr}{}}}
\subsection{\code{H5D$create_attr()}}{
  This function implements the HDF5-API function H5Acreate2. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$create_attr(attr_name, robj = NULL, dtype = NULL, space = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{robj}}{An R object to use as a template for the datatype and dataspace.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_open"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_open}{}}}
\subsection{\code{H5D$attr_open()}}{
  This function implements the HDF5-API function H5Aopen. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_open(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-create_attr_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-create_attr_by_name}{}}}
\subsection{\code{H5D$create_attr_by_name()}}{
  This function implements the HDF5-API function H5Acreate_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$create_attr_by_name(
  attr_name,
  obj_name,
  robj = NULL,
  dtype = NULL,
  space = NULL,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{robj}}{An R object to use as a template for the datatype and dataspace.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_open_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_open_by_name}{}}}
\subsection{\code{H5D$attr_open_by_name()}}{
  This function implements the HDF5-API function H5Aopen_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_open_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_open_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_open_by_idx}{}}}
\subsection{\code{H5D$attr_open_by_idx()}}{
  This function implements the HDF5-API function H5Aopen_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_open_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_exists_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_exists_by_name}{}}}
\subsection{\code{H5D$attr_exists_by_name()}}{
  This function implements the HDF5-API function H5Aexists_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_exists_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_exists"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_exists}{}}}
\subsection{\code{H5D$attr_exists()}}{
  This function implements the HDF5-API function H5Aexists. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_exists(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_rename_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_rename_by_name}{}}}
\subsection{\code{H5D$attr_rename_by_name()}}{
  This function implements the HDF5-API function H5Arename_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_rename_by_name(
  old_attr_name,
  new_attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{old_attr_name}}{The current name of the attribute.}
      \item{\code{new_attr_name}}{The new name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_rename"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_rename}{}}}
\subsection{\code{H5D$attr_rename()}}{
  This function implements the HDF5-API function H5Arename. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_rename(old_attr_name, new_attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{old_attr_name}}{The current name of the attribute.}
      \item{\code{new_attr_name}}{The new name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_delete"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_delete}{}}}
\subsection{\code{H5D$attr_delete()}}{
  This function implements the HDF5-API function H5Adelete. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_delete(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_delete_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_delete_by_name}{}}}
\subsection{\code{H5D$attr_delete_by_name()}}{
  This function implements the HDF5-API function H5Adelete_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_delete_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_delete_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_delete_by_idx}{}}}
\subsection{\code{H5D$attr_delete_by_idx()}}{
  This function implements the HDF5-API function H5Adelete_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_delete_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_info_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_info_by_name}{}}}
\subsection{\code{H5D$attr_info_by_name()}}{
  This function implements the HDF5-API function H5Aget_info_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_info_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_info_by_idx}{}}}
\subsection{\code{H5D$attr_info_by_idx()}}{
  This function implements the HDF5-API function H5Aget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_info_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_name_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_name_by_idx}{}}}
\subsection{\code{H5D$attr_name_by_idx()}}{
  This function implements the HDF5-API function H5Aget_name_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_name_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-attr_get_number"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-attr_get_number}{}}}
\subsection{\code{H5D$attr_get_number()}}{
  This function implements the HDF5-API function H5Aget_num_attrs. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$attr_get_number()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-flush"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-flush}{}}}
\subsection{\code{H5D$flush()}}{
  This function implements the HDF5-API function H5Fflush. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$flush(scope = h5const$H5F_SCOPE_GLOBAL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{scope}}{The scope of the flush; one of the \code{H5F_SCOPE_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_filename"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_filename}{}}}
\subsection{\code{H5D$get_filename()}}{
  This function implements the HDF5-API function H5Fget_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_filename()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-initialize}{}}}
\subsection{\code{H5D$new()}}{
  Initializes a new dataset-object. Only for internal use. Use the \code{create_dataset} function for \code{\link{H5Group}} and \code{\link{H5File}} objects
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$new(id = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{For internal use only}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_space"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_space}{}}}
\subsection{\code{H5D$get_space()}}{
  This function implements the HDF5-API function H5Dget_space. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_space()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_space_status"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_space_status}{}}}
\subsection{\code{H5D$get_space_status()}}{
  This function implements the HDF5-API function H5Dget_space_status. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_space_status()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_type}{}}}
\subsection{\code{H5D$get_type()}}{
  This function implements the HDF5-API function H5Dget_type. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_type(native = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{native}}{Logical; should the native version of the datatype be returned.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_create_plist"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_create_plist}{}}}
\subsection{\code{H5D$get_create_plist()}}{
  This function implements the HDF5-API function H5Dget_create_plist. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_create_plist()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_access_plist"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_access_plist}{}}}
\subsection{\code{H5D$get_access_plist()}}{
  This function implements the HDF5-API function H5Dget_access_plist. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_access_plist()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_offset"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_offset}{}}}
\subsection{\code{H5D$get_offset()}}{
  This function implements the HDF5-API function H5Dget_offset. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_offset()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_storage_size"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_storage_size}{}}}
\subsection{\code{H5D$get_storage_size()}}{
  This function implements the HDF5-API function H5Dget_storage_size. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_storage_size()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-vlen_get_buf_size"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-vlen_get_buf_size}{}}}
\subsection{\code{H5D$vlen_get_buf_size()}}{
  This function implements the HDF5-API function H5Dvlen_get_buf_size. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$vlen_get_buf_size(type, space)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{The type to use. For a dataspace one of \code{"simple"}, \code{"scalar"} or \code{"null"}; otherwise an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-vlen_reclaim"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-vlen_reclaim}{}}}
\subsection{\code{H5D$vlen_reclaim()}}{
  This function implements the HDF5-API function H5Dvlen_reclaim. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$vlen_reclaim(buffer, type, space, dataset_xfer_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{buffer}}{The raw buffer to read into or write from.}
      \item{\code{type}}{The type to use. For a dataspace one of \code{"simple"}, \code{"scalar"} or \code{"null"}; otherwise an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
      \item{\code{dataset_xfer_pl}}{The dataset transfer property list. See \code{\link{H5P_DATASET_XFER}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-read_low_level"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-read_low_level}{}}}
\subsection{\code{H5D$read_low_level()}}{
  This function is for advanced users. It is recommended to use \code{read} instead or the \code{[} interface. This function implements the HDF5-API function H5Dread, with minor changes to the API to accommodate R. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details. It reads the data in the dataset as specified by \code{mem_space} and return it as an R-obj
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$read_low_level(
  file_space = h5const$H5S_ALL,
  mem_space = NULL,
  mem_type = NULL,
  dataset_xfer_pl = h5const$H5P_DEFAULT,
  flags = getOption("hdf5r.h5tor_default"),
  set_dim = FALSE,
  dim_to_set = NULL,
  drop = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{file_space}}{An HDF5-space, represented as class \code{\link{H5S}} that determines which part of the dataset is being read. Can also be given as an id}
      \item{\code{mem_space}}{The space as it is represented in memory; advanced feature; may be removed in the future. Can also be given as an id.}
      \item{\code{mem_type}}{Memory type; extracted from the dataset if null (can be passed in for efficiency reasons Can also be given as an id.}
      \item{\code{dataset_xfer_pl}}{Dataset transfer property list. See \code{\link{H5P_DATASET_XFER}}}
      \item{\code{flags}}{Conversion rules for integer values. See also \code{\link{h5const}}}
      \item{\code{set_dim}}{If \code{TRUE}, the dimension attribute is set in the return value. How it is set is determined by \code{dim_to_set}.}
      \item{\code{dim_to_set}}{The dimension to set; Has to be numeric and needs to be specified if \code{set_dim} is \code{TRUE}. If the result is a data.frame, i.e. the data-type is a compound, then the dimension is ignored as the correct dimension is already set.}
      \item{\code{drop}}{Logical. Should dimensions of length 1 be dropped (R-default for arrays)}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-read"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-read}{}}}
\subsection{\code{H5D$read()}}{
  Main interface for reading data from the dataset. It is the function that is used by \code{[}, where all indices are being passed in the parameter \code{args}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$read(
  args = NULL,
  dataset_xfer_pl = h5const$H5P_DEFAULT,
  flags = getOption("hdf5r.h5tor_default"),
  drop = TRUE,
  envir = parent.frame()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{args}}{The indices for each dimension to subset given as a list. This makes this easier to use as a programmatic API. For interactive use we recommend the use of the \code{[} operator. If set to \code{NULL}, the entire dataset will be read.}
      \item{\code{dataset_xfer_pl}}{An object of class \code{\link{H5P_DATASET_XFER}}.}
      \item{\code{flags}}{Some flags governing edge cases of conversion from HDF5 to R. This is related to how integers are being treated and the issue of R not being able to natively represent 64bit integers and not at all being able to represent unsigned 64bit integers (even using add-on packages). The constants governing this are part of \code{\link{h5const}}. The relevant ones start with the term \code{H5TOR} and are documented there. The default set here returns a regular 32bit integer if it doesn't lead to an overflow and returns a 64bit integer from the \code{bit64} package otherwise. For 64bit unsigned int that are larger than 64bit signed int, it return a \code{double}. This looses precision, however.}
      \item{\code{drop}}{Logical. When reading data, should dimensions of size 1 be dropped.}
      \item{\code{envir}}{The environment in which to evaluate \code{args}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The data that was read as an R object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-write_low_level"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-write_low_level}{}}}
\subsection{\code{H5D$write_low_level()}}{
  This function is for advanced users. It is recommended to use \code{read} instead or the \code{[<-} interface as used for arrays. This function implements the HDF5-API function H5Dwrite, with some changes to accommodate R. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details. It writes that data from the \code{robj} into the dataset.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$write_low_level(
  robj,
  file_space = h5const$H5S_ALL,
  mem_space = NULL,
  mem_type = NULL,
  dataset_xfer_pl = h5const$H5P_DEFAULT,
  flush = getOption("hdf5r.flush_on_write")
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{robj}}{The object to write into the dataset}
      \item{\code{file_space}}{An HDF5-space, represented as class \code{\link{H5S}} that determines which part of the dataset is being written.}
      \item{\code{mem_space}}{The space as it is represented in memory; advanced feature; may be removed in the future}
      \item{\code{mem_type}}{Memory type; extracted from the dataset if null (can be passed in for efficiency reasons}
      \item{\code{dataset_xfer_pl}}{Dataset transfer property list. See \code{\link{H5P_DATASET_XFER}}}
      \item{\code{flush}}{Should a flush be done after the write}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-write"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-write}{}}}
\subsection{\code{H5D$write()}}{
  Main interface for writing data to the dataset. It is the function that is used by \code{[<-}, where all indices are being passed in the parameter \code{args}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$write(
  args,
  value,
  dataset_xfer_pl = h5const$H5P_DEFAULT,
  envir = parent.frame()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{args}}{The indices for each dimension to subset given as a list. This makes this easier to use as a programmatic API. For interactive use we recommend the use of the \code{[} operator. If set to \code{NULL}, the entire dataset will be read.}
      \item{\code{value}}{The data to write to the dataset}
      \item{\code{dataset_xfer_pl}}{An object of class \code{\link{H5P_DATASET_XFER}}.}
      \item{\code{envir}}{The environment in which to evaluate \code{args}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The HDF5 dataset object, returned invisibly
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-set_extent"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-set_extent}{}}}
\subsection{\code{H5D$set_extent()}}{
  This function implements the HDF5-API function H5Dset_extent. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$set_extent(dims)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dims}}{The dimensions of the object.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-get_fill_value"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-get_fill_value}{}}}
\subsection{\code{H5D$get_fill_value()}}{
  This function implements the HDF5-API function H5Pget_fill_value, automatically supplying the datatype of the dataset for convenience. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$get_fill_value()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-create_reference"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-create_reference}{}}}
\subsection{\code{H5D$create_reference()}}{
  This function implements the HDF5-API function H5Rcreate. The parameters are interpreted as in '['. The function always create \code{H5R_DATASET_REGION} references Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_r.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$create_reference(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Further arguments; currently ignored.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-print"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-print}{}}}
\subsection{\code{H5D$print()}}{
  Prints information for the dataset
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$print(..., max.attributes = 10)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
      \item{\code{max.attributes}}{Maximum number of attribute names to print}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5D-obj_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5D-obj_info}{}}}
\subsection{\code{H5D$obj_info()}}{
  This function implements the HDF5-API function H5Oget_info. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5D$obj_info(remove_internal_use_only = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{remove_internal_use_only}}{Logical; should fields that are only of internal use be dropped from the result.}
    }
    \if{html}{\out{</div>}}
  }
}

}
