% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5File.R
\docType{class}
\name{H5File}
\alias{H5File}
\alias{H5File-class}
\title{Class for interacting with HDF5 files.}
\value{
Object of class \code{\link{H5File}}.
}
\description{
\code{H5File} objects are are the main entry point to access HDF5 data from binary 
files. This class represents an open HDF5 File-id. It inherits all functions 
of the \code{\link{H5RefClass}}.

HDF5 files can be opened or generated using the \code{H5File$new()} function and
a specified file access mode. \code{H5File$new()} returns a \code{H5File} object
which can be used to access \code{\link{H5Group}}s and Datasets (see \code{\link{H5D}})
using subsetting parameters or according class methods.

HDF5 files which have been created or opened through \code{H5File$new()} need 
to be closed afterwards using \code{$close_all()}. \code{$close_all()} not only closes the file itself,
but also all objects that are still open inside it (such as groups or datasets). \code{$flush()} can be used 
to flush unwritten data to an HDF5 file. 

HDF5 Files typically contain the following objects:
\describe{
  \item{Groups}{Similar to a file system folder, used to organize HDF5 objects in a hierarchical way, see also \code{\link{H5Group}}}
  \item{Datasets}{Objects to store actual data, see also \code{\link{H5D}}}
  \item{Attributes}{Meta data objects to store extra information about Files, Groups and Datasets, see also \code{\link{H5A}}}
}
}
\section{Extract/List File Contents}{

The following functions are defined to extract HDF5 file contents:
\describe{
  \item{\code{\link{list.groups}}}{List HDF5 groups in file.}
  \item{\code{\link{list.datasets}}}{List HDF5 datasets in file.}
  \item{\code{\link{names}}}{List all items in a file or group (applicable for \code{\link{H5File}} and \code{\link{H5Group}})}
  \item{\code{\link{list.attributes}}}{List Attributes of HDF5 object (file, group or dataset).}
  \item{\code{\link{h5attr_names}}}{Attribute names of an HDF5 object; similar to list.attributes}
}
}

\examples{
# The following examples generates a HDF5 file with the different HDF5 
# Objects and shows its contents:
fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
file[["testdataset"]] <- 1:10
h5attr(file, "testattrib") <- LETTERS[1:10]
file$create_group("testgroup")
file[["testgroup/testdataset2"]] <- 1:10
# Show contents of file
file
# Close file and delete
file$close_all()

# The following example shows hdf5 file contents and how to use them to iterate over HDF5 elements:
file <- h5file(fname, mode = "a")
sapply(c("testgroup1", "testgroup2", "testgroup3"), file$create_group)
file[["testgroup1/testset1"]] <- 1:10
file[["testgroup2/testset2"]] <- 11:20
file[["testgroup3/testset3"]] <- 21:30

# Extract first 3 elements from each dataset and combine result to matrix
sapply(list.datasets(file, recursive = TRUE), function(x) file[[x]][1:3])
# Close file
file$close_all()
file.remove(fname)

}
\seealso{
\code{\link{h5file}}
}
\author{
Holger Hoefling, Mario Annau
}
\section{Super class}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{H5File}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{mode}}{The mode the file was opened with.}

    \item{\code{filename}}{The name of the file on disk.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{names}}{Returns the names of the items in the group or at the root of the file}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5File-obj_info_by_idx}{\code{H5File$obj_info_by_idx()}}
    \item \href{#method-H5File-obj_info_by_name}{\code{H5File$obj_info_by_name()}}
    \item \href{#method-H5File-obj_info}{\code{H5File$obj_info()}}
    \item \href{#method-H5File-open}{\code{H5File$open()}}
    \item \href{#method-H5File-open_by_idx}{\code{H5File$open_by_idx()}}
    \item \href{#method-H5File-ls}{\code{H5File$ls()}}
    \item \href{#method-H5File-exists}{\code{H5File$exists()}}
    \item \href{#method-H5File-path_valid}{\code{H5File$path_valid()}}
    \item \href{#method-H5File-link}{\code{H5File$link()}}
    \item \href{#method-H5File-obj_copy_to}{\code{H5File$obj_copy_to()}}
    \item \href{#method-H5File-obj_copy_from}{\code{H5File$obj_copy_from()}}
    \item \href{#method-H5File-group_info}{\code{H5File$group_info()}}
    \item \href{#method-H5File-group_info_by_name}{\code{H5File$group_info_by_name()}}
    \item \href{#method-H5File-group_info_by_idx}{\code{H5File$group_info_by_idx()}}
    \item \href{#method-H5File-create_group}{\code{H5File$create_group()}}
    \item \href{#method-H5File-create_dataset}{\code{H5File$create_dataset()}}
    \item \href{#method-H5File-commit}{\code{H5File$commit()}}
    \item \href{#method-H5File-link_create_hard}{\code{H5File$link_create_hard()}}
    \item \href{#method-H5File-link_create_soft}{\code{H5File$link_create_soft()}}
    \item \href{#method-H5File-link_create_external}{\code{H5File$link_create_external()}}
    \item \href{#method-H5File-link_exists}{\code{H5File$link_exists()}}
    \item \href{#method-H5File-link_move_from}{\code{H5File$link_move_from()}}
    \item \href{#method-H5File-link_move_to}{\code{H5File$link_move_to()}}
    \item \href{#method-H5File-link_copy_from}{\code{H5File$link_copy_from()}}
    \item \href{#method-H5File-link_copy_to}{\code{H5File$link_copy_to()}}
    \item \href{#method-H5File-link_delete}{\code{H5File$link_delete()}}
    \item \href{#method-H5File-link_delete_by_idx}{\code{H5File$link_delete_by_idx()}}
    \item \href{#method-H5File-link_info}{\code{H5File$link_info()}}
    \item \href{#method-H5File-link_info_by_idx}{\code{H5File$link_info_by_idx()}}
    \item \href{#method-H5File-link_value}{\code{H5File$link_value()}}
    \item \href{#method-H5File-link_value_by_idx}{\code{H5File$link_value_by_idx()}}
    \item \href{#method-H5File-link_name_by_idx}{\code{H5File$link_name_by_idx()}}
    \item \href{#method-H5File-mount}{\code{H5File$mount()}}
    \item \href{#method-H5File-unmount}{\code{H5File$unmount()}}
    \item \href{#method-H5File-get_obj_name}{\code{H5File$get_obj_name()}}
    \item \href{#method-H5File-create_attr}{\code{H5File$create_attr()}}
    \item \href{#method-H5File-attr_open}{\code{H5File$attr_open()}}
    \item \href{#method-H5File-create_attr_by_name}{\code{H5File$create_attr_by_name()}}
    \item \href{#method-H5File-attr_open_by_name}{\code{H5File$attr_open_by_name()}}
    \item \href{#method-H5File-attr_open_by_idx}{\code{H5File$attr_open_by_idx()}}
    \item \href{#method-H5File-attr_exists_by_name}{\code{H5File$attr_exists_by_name()}}
    \item \href{#method-H5File-attr_exists}{\code{H5File$attr_exists()}}
    \item \href{#method-H5File-attr_rename_by_name}{\code{H5File$attr_rename_by_name()}}
    \item \href{#method-H5File-attr_rename}{\code{H5File$attr_rename()}}
    \item \href{#method-H5File-attr_delete}{\code{H5File$attr_delete()}}
    \item \href{#method-H5File-attr_delete_by_name}{\code{H5File$attr_delete_by_name()}}
    \item \href{#method-H5File-attr_delete_by_idx}{\code{H5File$attr_delete_by_idx()}}
    \item \href{#method-H5File-attr_info_by_name}{\code{H5File$attr_info_by_name()}}
    \item \href{#method-H5File-attr_info_by_idx}{\code{H5File$attr_info_by_idx()}}
    \item \href{#method-H5File-attr_name_by_idx}{\code{H5File$attr_name_by_idx()}}
    \item \href{#method-H5File-attr_get_number}{\code{H5File$attr_get_number()}}
    \item \href{#method-H5File-create_reference}{\code{H5File$create_reference()}}
    \item \href{#method-H5File-flush}{\code{H5File$flush()}}
    \item \href{#method-H5File-get_filename}{\code{H5File$get_filename()}}
    \item \href{#method-H5File-initialize}{\code{H5File$new()}}
    \item \href{#method-H5File-get_obj_count}{\code{H5File$get_obj_count()}}
    \item \href{#method-H5File-get_obj_ids}{\code{H5File$get_obj_ids()}}
    \item \href{#method-H5File-get_filesize}{\code{H5File$get_filesize()}}
    \item \href{#method-H5File-file_info}{\code{H5File$file_info()}}
    \item \href{#method-H5File-get_intent}{\code{H5File$get_intent()}}
    \item \href{#method-H5File-close_all}{\code{H5File$close_all()}}
    \item \href{#method-H5File-print}{\code{H5File$print()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-obj_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-obj_info_by_idx}{}}}
\subsection{\code{H5File$obj_info_by_idx()}}{
  This function implements the HDF5-API function H5Oget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$obj_info_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  remove_internal_use_only = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{remove_internal_use_only}}{Logical; should fields that are only of internal use be dropped from the result.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-obj_info_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-obj_info_by_name}{}}}
\subsection{\code{H5File$obj_info_by_name()}}{
  This function implements the HDF5-API function H5Oget_info_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$obj_info_by_name(object_name, remove_internal_use_only = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{object_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{remove_internal_use_only}}{Logical; should fields that are only of internal use be dropped from the result.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-obj_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-obj_info}{}}}
\subsection{\code{H5File$obj_info()}}{
  This function implements the HDF5-API function H5Oget_info. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$obj_info(remove_internal_use_only = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{remove_internal_use_only}}{Logical; should fields that are only of internal use be dropped from the result.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-open"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-open}{}}}
\subsection{\code{H5File$open()}}{
  Opens groups, datasets or types using the appropriate HDF5-API functions. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for datasets, \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for types and \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for general objects.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$open(
  name,
  link_access_pl = h5const$H5P_DEFAULT,
  dataset_access_pl = h5const$H5P_DEFAULT,
  type_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
      \item{\code{dataset_access_pl}}{The dataset access property list. See \code{\link{H5P_DATASET_ACCESS}}.}
      \item{\code{type_access_pl}}{The datatype access property list.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-open_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-open_by_idx}{}}}
\subsection{\code{H5File$open_by_idx()}}{
  This function implements the HDF5-API function H5Oopen_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$open_by_idx(
  n,
  group_name = ".",
  index_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-ls"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-ls}{}}}
\subsection{\code{H5File$ls()}}{
  Returns the contents of a file or group as a data.frame.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$ls(
  recursive = FALSE,
  detailed = FALSE,
  index_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_INC,
  link_access_pl = h5const$H5P_DEFAULT,
  dataset_access_pl = h5const$H5P_DEFAULT,
  type_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{recursive}}{Logical; should the listing descend into subgroups.}
      \item{\code{detailed}}{Logical; should a more detailed listing be returned.}
      \item{\code{index_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
      \item{\code{dataset_access_pl}}{The dataset access property list. See \code{\link{H5P_DATASET_ACCESS}}.}
      \item{\code{type_access_pl}}{The datatype access property list.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-exists"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-exists}{}}}
\subsection{\code{H5File$exists()}}{
  This function implements the HDF5-API function H5Lexists. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$exists(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-path_valid"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-path_valid}{}}}
\subsection{\code{H5File$path_valid()}}{
  This function implements the HDF5-API function H5LTpath_valid. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$path_valid(path, check_object_valid = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{path}}{The path of the object inside the file.}
      \item{\code{check_object_valid}}{Logical; should the object be checked for validity first.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link}{}}}
\subsection{\code{H5File$link()}}{
  This function implements the HDF5-API function H5Olink. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link(
  obj,
  new_link_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{obj}}{The HDF5 object to use.}
      \item{\code{new_link_name}}{The new name of the link.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-obj_copy_to"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-obj_copy_to}{}}}
\subsection{\code{H5File$obj_copy_to()}}{
  This function implements the HDF5-API function H5Ocopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$obj_copy_to(
  dst_loc,
  dst_name,
  src_name,
  object_copy_pl = h5const$H5P_DEFAULT,
  link_create_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dst_loc}}{The destination location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{object_copy_pl}}{The object copy property list. See \code{\link{H5P_OBJECT_COPY}}.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-obj_copy_from"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-obj_copy_from}{}}}
\subsection{\code{H5File$obj_copy_from()}}{
  This function implements the HDF5-API function H5Ocopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$obj_copy_from(
  src_loc,
  src_name,
  dst_name,
  object_copy_pl = h5const$H5P_DEFAULT,
  link_create_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{src_loc}}{The source location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{object_copy_pl}}{The object copy property list. See \code{\link{H5P_OBJECT_COPY}}.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-group_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-group_info}{}}}
\subsection{\code{H5File$group_info()}}{
  This function implements the HDF5-API function H5Gget_info. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$group_info()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-group_info_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-group_info_by_name}{}}}
\subsection{\code{H5File$group_info_by_name()}}{
  This function implements the HDF5-API function H5Gget_info_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$group_info_by_name(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-group_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-group_info_by_idx}{}}}
\subsection{\code{H5File$group_info_by_idx()}}{
  This function implements the HDF5-API function H5Gget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$group_info_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-create_group"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-create_group}{}}}
\subsection{\code{H5File$create_group()}}{
  This function implements the HDF5-API function H5Gcreate2 and H5Gcreate_anon (if name is NULL). Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for regular groups and \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for anonymous groups for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$create_group(
  name,
  link_create_pl = h5const$H5P_DEFAULT,
  group_create_pl = h5const$H5P_DEFAULT,
  group_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{group_create_pl}}{The group creation property list.}
      \item{\code{group_access_pl}}{The group access property list.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-create_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-create_dataset}{}}}
\subsection{\code{H5File$create_dataset()}}{
  This function is the main interface to create a new dataset. Its parameters allow for customization of the default behavior, i.e. in order to get a specific datatype, a certain chunk size or dataset dimensionality. Also note that this function implements the HDF5-API function H5Dcreate2 and H5Dcreate_anon (if name is NULL). Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for regular groups and \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for anonymous groups for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$create_dataset(
  name,
  robj = NULL,
  dtype = NULL,
  space = NULL,
  dims = NULL,
  chunk_dims = "auto",
  gzip_level = 4,
  link_create_pl = h5const$H5P_DEFAULT,
  dataset_create_pl = h5const$H5P_DEFAULT,
  dataset_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the new dataset. If missing, an anonymous dataset is created}
      \item{\code{robj}}{An R-object to take as a template for creating the dataset. Either \code{robj} or both \code{dtype} and \code{space} have to be provided}
      \item{\code{dtype}}{The datatype to use for the creation of the object. Can be null if \code{robj} is given.}
      \item{\code{space}}{The space to use for the object creation. Can be null if \code{robj} is given. Otherwise an object of type \code{H5S} which specifies the dimensions of the dataset.}
      \item{\code{dims}}{Dimension of the new dataset; used if \code{space} is \code{NULL}. overwrite the dimension guessed from \code{robj} if \code{robj} is given.}
      \item{\code{chunk_dims}}{Size of the chunk. Has to have the same length as the dataset dimension. If \code{"auto"} then the size of each chunk is estimated so that each chunk is roughly as large in bytes as the value in the \code{hdf5r.chunk_size} option. See also \code{\link{guess_chunks}} for a more detailed explanation. If set to \code{NULL}, then no chunking is used, unless explicitly set in \code{dataset_create_pl}.}
      \item{\code{gzip_level}}{Only if \code{chunk_dims} is not null. If given, then the \code{dataset_create_pl} is set to enable zipping at the level given here. If set to NULL, then gzip is not set (but could be set otherwise in \code{dataset_create_pl}}
      \item{\code{link_create_pl}}{Link creation property list. See \code{\link{H5P_LINK_CREATE}}}
      \item{\code{dataset_create_pl}}{Dataset creation property list. See \code{\link{H5P_DATASET_CREATE}}}
      \item{\code{dataset_access_pl}}{Dataset access property list. See \code{\link{H5P_DATASET_ACCESS}}}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-commit"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-commit}{}}}
\subsection{\code{H5File$commit()}}{
  This function implements the HDF5-API function H5Tcommit2. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$commit(
  name,
  dtype,
  link_create_pl = h5const$H5P_DEFAULT,
  type_create_pl = h5const$H5P_DEFAULT,
  type_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{type_create_pl}}{The datatype creation property list.}
      \item{\code{type_access_pl}}{The datatype access property list.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_create_hard"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_create_hard}{}}}
\subsection{\code{H5File$link_create_hard()}}{
  This function implements the HDF5-API function H5Lcreate_hard. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_create_hard(
  obj_loc,
  obj_name,
  link_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{obj_loc}}{The location of the object; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_name}}{The name of the link.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_create_soft"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_create_soft}{}}}
\subsection{\code{H5File$link_create_soft()}}{
  This function implements the HDF5-API function H5Lcreate_soft. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_create_soft(
  target_path,
  link_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_path}}{The path of the target object.}
      \item{\code{link_name}}{The name of the link.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_create_external"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_create_external}{}}}
\subsection{\code{H5File$link_create_external()}}{
  This function implements the HDF5-API function H5Lcreate_external. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_create_external(
  target_filename,
  target_obj_name,
  link_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_filename}}{The name of the file that holds the target object.}
      \item{\code{target_obj_name}}{The name of the target object.}
      \item{\code{link_name}}{The name of the link.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_exists"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_exists}{}}}
\subsection{\code{H5File$link_exists()}}{
  This function implements the HDF5-API function H5Lexists. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_exists(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_move_from"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_move_from}{}}}
\subsection{\code{H5File$link_move_from()}}{
  This function implements the HDF5-API function H5Lmove. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_move_from(
  src_loc,
  src_name,
  dst_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{src_loc}}{The source location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_move_to"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_move_to}{}}}
\subsection{\code{H5File$link_move_to()}}{
  This function implements the HDF5-API function H5Lmove. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_move_to(
  dst_loc,
  dst_name,
  src_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dst_loc}}{The destination location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_copy_from"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_copy_from}{}}}
\subsection{\code{H5File$link_copy_from()}}{
  This function implements the HDF5-API function H5Lcopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_copy_from(
  src_loc,
  src_name,
  dst_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{src_loc}}{The source location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_copy_to"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_copy_to}{}}}
\subsection{\code{H5File$link_copy_to()}}{
  This function implements the HDF5-API function H5Lcopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_copy_to(
  dst_loc,
  dst_name,
  src_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dst_loc}}{The destination location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_delete"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_delete}{}}}
\subsection{\code{H5File$link_delete()}}{
  This function implements the HDF5-API function H5Ldelete. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_delete(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_delete_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_delete_by_idx}{}}}
\subsection{\code{H5File$link_delete_by_idx()}}{
  This function implements the HDF5-API function H5Ldelete_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_delete_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_info}{}}}
\subsection{\code{H5File$link_info()}}{
  This function implements the HDF5-API function H5Lget_info. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_info(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_info_by_idx}{}}}
\subsection{\code{H5File$link_info_by_idx()}}{
  This function implements the HDF5-API function H5Lget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_info_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_value"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_value}{}}}
\subsection{\code{H5File$link_value()}}{
  This function implements the HDF5-API function H5Lget_val. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_value(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_value_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_value_by_idx}{}}}
\subsection{\code{H5File$link_value_by_idx()}}{
  This function implements the HDF5-API function H5Lget_val_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_value_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-link_name_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-link_name_by_idx}{}}}
\subsection{\code{H5File$link_name_by_idx()}}{
  This function implements the HDF5-API function H5Lget_name_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$link_name_by_idx(
  n,
  group_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-mount"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-mount}{}}}
\subsection{\code{H5File$mount()}}{
  This function implements the HDF5-API function H5Fmount. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$mount(name, child)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{child}}{The object to mount; an object of class \code{\link{H5File}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-unmount"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-unmount}{}}}
\subsection{\code{H5File$unmount()}}{
  This function implements the HDF5-API function H5Funmount. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$unmount(name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-get_obj_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-get_obj_name}{}}}
\subsection{\code{H5File$get_obj_name()}}{
  This function implements the HDF5-API function H5Iget_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$get_obj_name()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-create_attr"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-create_attr}{}}}
\subsection{\code{H5File$create_attr()}}{
  This function implements the HDF5-API function H5Acreate2. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$create_attr(attr_name, robj = NULL, dtype = NULL, space = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{robj}}{An R object to use as a template for the datatype and dataspace.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_open"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_open}{}}}
\subsection{\code{H5File$attr_open()}}{
  This function implements the HDF5-API function H5Aopen. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_open(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-create_attr_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-create_attr_by_name}{}}}
\subsection{\code{H5File$create_attr_by_name()}}{
  This function implements the HDF5-API function H5Acreate_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$create_attr_by_name(
  attr_name,
  obj_name,
  robj = NULL,
  dtype = NULL,
  space = NULL,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{robj}}{An R object to use as a template for the datatype and dataspace.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_open_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_open_by_name}{}}}
\subsection{\code{H5File$attr_open_by_name()}}{
  This function implements the HDF5-API function H5Aopen_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_open_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_open_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_open_by_idx}{}}}
\subsection{\code{H5File$attr_open_by_idx()}}{
  This function implements the HDF5-API function H5Aopen_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_open_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_exists_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_exists_by_name}{}}}
\subsection{\code{H5File$attr_exists_by_name()}}{
  This function implements the HDF5-API function H5Aexists_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_exists_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_exists"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_exists}{}}}
\subsection{\code{H5File$attr_exists()}}{
  This function implements the HDF5-API function H5Aexists. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_exists(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_rename_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_rename_by_name}{}}}
\subsection{\code{H5File$attr_rename_by_name()}}{
  This function implements the HDF5-API function H5Arename_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_rename_by_name(
  old_attr_name,
  new_attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{old_attr_name}}{The current name of the attribute.}
      \item{\code{new_attr_name}}{The new name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_rename"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_rename}{}}}
\subsection{\code{H5File$attr_rename()}}{
  This function implements the HDF5-API function H5Arename. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_rename(old_attr_name, new_attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{old_attr_name}}{The current name of the attribute.}
      \item{\code{new_attr_name}}{The new name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_delete"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_delete}{}}}
\subsection{\code{H5File$attr_delete()}}{
  This function implements the HDF5-API function H5Adelete. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_delete(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_delete_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_delete_by_name}{}}}
\subsection{\code{H5File$attr_delete_by_name()}}{
  This function implements the HDF5-API function H5Adelete_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_delete_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_delete_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_delete_by_idx}{}}}
\subsection{\code{H5File$attr_delete_by_idx()}}{
  This function implements the HDF5-API function H5Adelete_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_delete_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_info_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_info_by_name}{}}}
\subsection{\code{H5File$attr_info_by_name()}}{
  This function implements the HDF5-API function H5Aget_info_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_info_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_info_by_idx}{}}}
\subsection{\code{H5File$attr_info_by_idx()}}{
  This function implements the HDF5-API function H5Aget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_info_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_name_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_name_by_idx}{}}}
\subsection{\code{H5File$attr_name_by_idx()}}{
  This function implements the HDF5-API function H5Aget_name_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_name_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-attr_get_number"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-attr_get_number}{}}}
\subsection{\code{H5File$attr_get_number()}}{
  This function implements the HDF5-API function H5Aget_num_attrs. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$attr_get_number()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-create_reference"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-create_reference}{}}}
\subsection{\code{H5File$create_reference()}}{
  This function implements the HDF5-API function H5Rcreate. If \code{space=NULL} then a \code{H5R_OBJECT} reference is created, otherwise a \code{H5R_DATASET_REGION} reference Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_r.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$create_reference(name = ".", space = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-flush"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-flush}{}}}
\subsection{\code{H5File$flush()}}{
  This function implements the HDF5-API function H5Fflush. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$flush(scope = h5const$H5F_SCOPE_GLOBAL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{scope}}{The scope of the flush; one of the \code{H5F_SCOPE_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-get_filename"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-get_filename}{}}}
\subsection{\code{H5File$get_filename()}}{
  This function implements the HDF5-API function H5Fget_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$get_filename()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-initialize}{}}}
\subsection{\code{H5File$new()}}{
  Opens or creates a new HDF5 File
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$new(
  filename = NULL,
  mode = c("a", "r", "r+", "w", "w-", "x"),
  file_create_pl = h5const$H5P_DEFAULT,
  file_access_pl = h5const$H5P_DEFAULT,
  id = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{filename}}{Name of the file}
      \item{\code{mode}}{How to open it. \code{a} creates a new file or opens an existing one for read/write. \code{r} opens an existing file for reading, \code{r+} opens an existing file for read/write. \code{w} creates a file, truncating any existing ones and \code{w-}/\code{x} are synonyms, creating a file and failing if it already exists.}
      \item{\code{file_create_pl}}{The file creation property list. See \code{\link{H5P_FILE_CREATE}}.}
      \item{\code{file_access_pl}}{The file access property list. See \code{\link{H5P_FILE_ACCESS}}.}
      \item{\code{id}}{An HDF5 id; for internal use only.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-get_obj_count"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-get_obj_count}{}}}
\subsection{\code{H5File$get_obj_count()}}{
  This function implements the HDF5-API function H5Fget_obj_count. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$get_obj_count(types = h5const$H5F_OBJ_ALL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{types}}{The types of object to count; one or more of the \code{H5F_OBJ_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-get_obj_ids"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-get_obj_ids}{}}}
\subsection{\code{H5File$get_obj_ids()}}{
  This function implements the HDF5-API function H5Fget_obj_ids. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$get_obj_ids(types = h5const$H5F_OBJ_ALL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{types}}{The types of object to count; one or more of the \code{H5F_OBJ_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-get_filesize"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-get_filesize}{}}}
\subsection{\code{H5File$get_filesize()}}{
  This function implements the HDF5-API function H5Fget_filesize. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$get_filesize()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-file_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-file_info}{}}}
\subsection{\code{H5File$file_info()}}{
  This function implements the HDF5-API function H5Fget_info2. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details. Please note that the returned information differs if HDF5 Version 1.8.16 or HDF5 Version >= 1.10.0 is being used
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$file_info()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-get_intent"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-get_intent}{}}}
\subsection{\code{H5File$get_intent()}}{
  This function implements the HDF5-API function H5Fget_intent. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$get_intent()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-close_all"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-close_all}{}}}
\subsection{\code{H5File$close_all()}}{
  Closes the file, flushes it and also closes all open objects that are still open in it. This is the recommended way of closing any file. If not all objects in a file are closed, the file remains open and cannot be re-opened the regular way.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$close_all(close_self = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{close_self}}{Logical; should the object itself be closed as well.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5File-print"></a>}}
\if{latex}{\out{\hypertarget{method-H5File-print}{}}}
\subsection{\code{H5File$print()}}{
  Prints information for the file
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5File$print(..., max.attributes = 10, max.listing = 10)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
      \item{\code{max.attributes}}{Maximum number of attribute names to print}
      \item{\code{max.listing}}{Maximum number of ls-items to print}
    }
    \if{html}{\out{</div>}}
  }
}

}
