% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5Group.R
\docType{class}
\name{H5Group}
\alias{H5Group}
\alias{H5Group-class}
\title{Class for representing HDF5 groups}
\value{
Object of class \code{\link{H5Group}}.
}
\description{
HDF5-Groups are essentially equivalent to directories in a file system. Inside the groups, other groups or datasets can
be created. For the most parts, groups behave like files, so please also look at the documentation of \code{H5File}.
}
\examples{
fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
group <- file$create_group("testgroup")
group$print()
group$close()
file$close_all()
}
\author{
Holger Hoefling
}
\section{Super class}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{H5Group}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{names}}{Returns the names of the items in the group or at the root of the file}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5Group-open}{\code{H5Group$open()}}
    \item \href{#method-H5Group-open_by_idx}{\code{H5Group$open_by_idx()}}
    \item \href{#method-H5Group-ls}{\code{H5Group$ls()}}
    \item \href{#method-H5Group-exists}{\code{H5Group$exists()}}
    \item \href{#method-H5Group-path_valid}{\code{H5Group$path_valid()}}
    \item \href{#method-H5Group-link}{\code{H5Group$link()}}
    \item \href{#method-H5Group-obj_copy_to}{\code{H5Group$obj_copy_to()}}
    \item \href{#method-H5Group-obj_copy_from}{\code{H5Group$obj_copy_from()}}
    \item \href{#method-H5Group-group_info}{\code{H5Group$group_info()}}
    \item \href{#method-H5Group-group_info_by_name}{\code{H5Group$group_info_by_name()}}
    \item \href{#method-H5Group-group_info_by_idx}{\code{H5Group$group_info_by_idx()}}
    \item \href{#method-H5Group-create_group}{\code{H5Group$create_group()}}
    \item \href{#method-H5Group-create_dataset}{\code{H5Group$create_dataset()}}
    \item \href{#method-H5Group-commit}{\code{H5Group$commit()}}
    \item \href{#method-H5Group-link_create_hard}{\code{H5Group$link_create_hard()}}
    \item \href{#method-H5Group-link_create_soft}{\code{H5Group$link_create_soft()}}
    \item \href{#method-H5Group-link_create_external}{\code{H5Group$link_create_external()}}
    \item \href{#method-H5Group-link_exists}{\code{H5Group$link_exists()}}
    \item \href{#method-H5Group-link_move_from}{\code{H5Group$link_move_from()}}
    \item \href{#method-H5Group-link_move_to}{\code{H5Group$link_move_to()}}
    \item \href{#method-H5Group-link_copy_from}{\code{H5Group$link_copy_from()}}
    \item \href{#method-H5Group-link_copy_to}{\code{H5Group$link_copy_to()}}
    \item \href{#method-H5Group-link_delete}{\code{H5Group$link_delete()}}
    \item \href{#method-H5Group-link_delete_by_idx}{\code{H5Group$link_delete_by_idx()}}
    \item \href{#method-H5Group-link_info}{\code{H5Group$link_info()}}
    \item \href{#method-H5Group-link_info_by_idx}{\code{H5Group$link_info_by_idx()}}
    \item \href{#method-H5Group-link_value}{\code{H5Group$link_value()}}
    \item \href{#method-H5Group-link_value_by_idx}{\code{H5Group$link_value_by_idx()}}
    \item \href{#method-H5Group-link_name_by_idx}{\code{H5Group$link_name_by_idx()}}
    \item \href{#method-H5Group-mount}{\code{H5Group$mount()}}
    \item \href{#method-H5Group-unmount}{\code{H5Group$unmount()}}
    \item \href{#method-H5Group-get_obj_name}{\code{H5Group$get_obj_name()}}
    \item \href{#method-H5Group-create_attr}{\code{H5Group$create_attr()}}
    \item \href{#method-H5Group-attr_open}{\code{H5Group$attr_open()}}
    \item \href{#method-H5Group-create_attr_by_name}{\code{H5Group$create_attr_by_name()}}
    \item \href{#method-H5Group-attr_open_by_name}{\code{H5Group$attr_open_by_name()}}
    \item \href{#method-H5Group-attr_open_by_idx}{\code{H5Group$attr_open_by_idx()}}
    \item \href{#method-H5Group-attr_exists_by_name}{\code{H5Group$attr_exists_by_name()}}
    \item \href{#method-H5Group-attr_exists}{\code{H5Group$attr_exists()}}
    \item \href{#method-H5Group-attr_rename_by_name}{\code{H5Group$attr_rename_by_name()}}
    \item \href{#method-H5Group-attr_rename}{\code{H5Group$attr_rename()}}
    \item \href{#method-H5Group-attr_delete}{\code{H5Group$attr_delete()}}
    \item \href{#method-H5Group-attr_delete_by_name}{\code{H5Group$attr_delete_by_name()}}
    \item \href{#method-H5Group-attr_delete_by_idx}{\code{H5Group$attr_delete_by_idx()}}
    \item \href{#method-H5Group-attr_info_by_name}{\code{H5Group$attr_info_by_name()}}
    \item \href{#method-H5Group-attr_info_by_idx}{\code{H5Group$attr_info_by_idx()}}
    \item \href{#method-H5Group-attr_name_by_idx}{\code{H5Group$attr_name_by_idx()}}
    \item \href{#method-H5Group-attr_get_number}{\code{H5Group$attr_get_number()}}
    \item \href{#method-H5Group-create_reference}{\code{H5Group$create_reference()}}
    \item \href{#method-H5Group-flush}{\code{H5Group$flush()}}
    \item \href{#method-H5Group-get_filename}{\code{H5Group$get_filename()}}
    \item \href{#method-H5Group-print}{\code{H5Group$print()}}
    \item \href{#method-H5Group-obj_info_by_idx}{\code{H5Group$obj_info_by_idx()}}
    \item \href{#method-H5Group-obj_info_by_name}{\code{H5Group$obj_info_by_name()}}
    \item \href{#method-H5Group-obj_info}{\code{H5Group$obj_info()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="initialize"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-initialize'><code>H5RefClass$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-open"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-open}{}}}
\subsection{\code{H5Group$open()}}{
  Opens groups, datasets or types using the appropriate HDF5-API functions. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for datasets, \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for types and \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for general objects.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$open(
  name,
  link_access_pl = h5const$H5P_DEFAULT,
  dataset_access_pl = h5const$H5P_DEFAULT,
  type_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
      \item{\code{dataset_access_pl}}{The dataset access property list. See \code{\link{H5P_DATASET_ACCESS}}.}
      \item{\code{type_access_pl}}{The datatype access property list.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-open_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-open_by_idx}{}}}
\subsection{\code{H5Group$open_by_idx()}}{
  This function implements the HDF5-API function H5Oopen_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$open_by_idx(
  n,
  group_name = ".",
  index_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-ls"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-ls}{}}}
\subsection{\code{H5Group$ls()}}{
  Returns the contents of a file or group as a data.frame.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$ls(
  recursive = FALSE,
  detailed = FALSE,
  index_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_INC,
  link_access_pl = h5const$H5P_DEFAULT,
  dataset_access_pl = h5const$H5P_DEFAULT,
  type_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{recursive}}{Logical; should the listing descend into subgroups.}
      \item{\code{detailed}}{Logical; should a more detailed listing be returned.}
      \item{\code{index_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
      \item{\code{dataset_access_pl}}{The dataset access property list. See \code{\link{H5P_DATASET_ACCESS}}.}
      \item{\code{type_access_pl}}{The datatype access property list.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-exists"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-exists}{}}}
\subsection{\code{H5Group$exists()}}{
  This function implements the HDF5-API function H5Lexists. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$exists(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-path_valid"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-path_valid}{}}}
\subsection{\code{H5Group$path_valid()}}{
  This function implements the HDF5-API function H5LTpath_valid. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$path_valid(path, check_object_valid = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{path}}{The path of the object inside the file.}
      \item{\code{check_object_valid}}{Logical; should the object be checked for validity first.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link}{}}}
\subsection{\code{H5Group$link()}}{
  This function implements the HDF5-API function H5Olink. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link(
  obj,
  new_link_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{obj}}{The HDF5 object to use.}
      \item{\code{new_link_name}}{The new name of the link.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-obj_copy_to"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-obj_copy_to}{}}}
\subsection{\code{H5Group$obj_copy_to()}}{
  This function implements the HDF5-API function H5Ocopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$obj_copy_to(
  dst_loc,
  dst_name,
  src_name,
  object_copy_pl = h5const$H5P_DEFAULT,
  link_create_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dst_loc}}{The destination location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{object_copy_pl}}{The object copy property list. See \code{\link{H5P_OBJECT_COPY}}.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-obj_copy_from"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-obj_copy_from}{}}}
\subsection{\code{H5Group$obj_copy_from()}}{
  This function implements the HDF5-API function H5Ocopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$obj_copy_from(
  src_loc,
  src_name,
  dst_name,
  object_copy_pl = h5const$H5P_DEFAULT,
  link_create_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{src_loc}}{The source location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{object_copy_pl}}{The object copy property list. See \code{\link{H5P_OBJECT_COPY}}.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-group_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-group_info}{}}}
\subsection{\code{H5Group$group_info()}}{
  This function implements the HDF5-API function H5Gget_info. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$group_info()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-group_info_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-group_info_by_name}{}}}
\subsection{\code{H5Group$group_info_by_name()}}{
  This function implements the HDF5-API function H5Gget_info_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$group_info_by_name(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-group_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-group_info_by_idx}{}}}
\subsection{\code{H5Group$group_info_by_idx()}}{
  This function implements the HDF5-API function H5Gget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$group_info_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-create_group"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-create_group}{}}}
\subsection{\code{H5Group$create_group()}}{
  This function implements the HDF5-API function H5Gcreate2 and H5Gcreate_anon (if name is NULL). Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for regular groups and \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_g.html} for anonymous groups for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$create_group(
  name,
  link_create_pl = h5const$H5P_DEFAULT,
  group_create_pl = h5const$H5P_DEFAULT,
  group_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{group_create_pl}}{The group creation property list.}
      \item{\code{group_access_pl}}{The group access property list.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-create_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-create_dataset}{}}}
\subsection{\code{H5Group$create_dataset()}}{
  This function is the main interface to create a new dataset. Its parameters allow for customization of the default behavior, i.e. in order to get a specific datatype, a certain chunk size or dataset dimensionality. Also note that this function implements the HDF5-API function H5Dcreate2 and H5Dcreate_anon (if name is NULL). Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for regular groups and \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_d.html} for anonymous groups for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$create_dataset(
  name,
  robj = NULL,
  dtype = NULL,
  space = NULL,
  dims = NULL,
  chunk_dims = "auto",
  gzip_level = 4,
  link_create_pl = h5const$H5P_DEFAULT,
  dataset_create_pl = h5const$H5P_DEFAULT,
  dataset_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the new dataset. If missing, an anonymous dataset is created}
      \item{\code{robj}}{An R-object to take as a template for creating the dataset. Either \code{robj} or both \code{dtype} and \code{space} have to be provided}
      \item{\code{dtype}}{The datatype to use for the creation of the object. Can be null if \code{robj} is given.}
      \item{\code{space}}{The space to use for the object creation. Can be null if \code{robj} is given. Otherwise an object of type \code{H5S} which specifies the dimensions of the dataset.}
      \item{\code{dims}}{Dimension of the new dataset; used if \code{space} is \code{NULL}. overwrite the dimension guessed from \code{robj} if \code{robj} is given.}
      \item{\code{chunk_dims}}{Size of the chunk. Has to have the same length as the dataset dimension. If \code{"auto"} then the size of each chunk is estimated so that each chunk is roughly as large in bytes as the value in the \code{hdf5r.chunk_size} option. See also \code{\link{guess_chunks}} for a more detailed explanation. If set to \code{NULL}, then no chunking is used, unless explicitly set in \code{dataset_create_pl}.}
      \item{\code{gzip_level}}{Only if \code{chunk_dims} is not null. If given, then the \code{dataset_create_pl} is set to enable zipping at the level given here. If set to NULL, then gzip is not set (but could be set otherwise in \code{dataset_create_pl}}
      \item{\code{link_create_pl}}{Link creation property list. See \code{\link{H5P_LINK_CREATE}}}
      \item{\code{dataset_create_pl}}{Dataset creation property list. See \code{\link{H5P_DATASET_CREATE}}}
      \item{\code{dataset_access_pl}}{Dataset access property list. See \code{\link{H5P_DATASET_ACCESS}}}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-commit"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-commit}{}}}
\subsection{\code{H5Group$commit()}}{
  This function implements the HDF5-API function H5Tcommit2. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$commit(
  name,
  dtype,
  link_create_pl = h5const$H5P_DEFAULT,
  type_create_pl = h5const$H5P_DEFAULT,
  type_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{type_create_pl}}{The datatype creation property list.}
      \item{\code{type_access_pl}}{The datatype access property list.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_create_hard"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_create_hard}{}}}
\subsection{\code{H5Group$link_create_hard()}}{
  This function implements the HDF5-API function H5Lcreate_hard. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_create_hard(
  obj_loc,
  obj_name,
  link_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{obj_loc}}{The location of the object; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_name}}{The name of the link.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_create_soft"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_create_soft}{}}}
\subsection{\code{H5Group$link_create_soft()}}{
  This function implements the HDF5-API function H5Lcreate_soft. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_create_soft(
  target_path,
  link_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_path}}{The path of the target object.}
      \item{\code{link_name}}{The name of the link.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_create_external"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_create_external}{}}}
\subsection{\code{H5Group$link_create_external()}}{
  This function implements the HDF5-API function H5Lcreate_external. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_create_external(
  target_filename,
  target_obj_name,
  link_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_filename}}{The name of the file that holds the target object.}
      \item{\code{target_obj_name}}{The name of the target object.}
      \item{\code{link_name}}{The name of the link.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_exists"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_exists}{}}}
\subsection{\code{H5Group$link_exists()}}{
  This function implements the HDF5-API function H5Lexists. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_exists(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_move_from"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_move_from}{}}}
\subsection{\code{H5Group$link_move_from()}}{
  This function implements the HDF5-API function H5Lmove. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_move_from(
  src_loc,
  src_name,
  dst_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{src_loc}}{The source location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_move_to"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_move_to}{}}}
\subsection{\code{H5Group$link_move_to()}}{
  This function implements the HDF5-API function H5Lmove. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_move_to(
  dst_loc,
  dst_name,
  src_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dst_loc}}{The destination location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_copy_from"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_copy_from}{}}}
\subsection{\code{H5Group$link_copy_from()}}{
  This function implements the HDF5-API function H5Lcopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_copy_from(
  src_loc,
  src_name,
  dst_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{src_loc}}{The source location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_copy_to"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_copy_to}{}}}
\subsection{\code{H5Group$link_copy_to()}}{
  This function implements the HDF5-API function H5Lcopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_copy_to(
  dst_loc,
  dst_name,
  src_name,
  link_create_pl = h5const$H5P_DEFAULT,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dst_loc}}{The destination location; an object of class \code{\link{H5File}} or \code{\link{H5Group}}.}
      \item{\code{dst_name}}{The name of the destination object.}
      \item{\code{src_name}}{The name of the source object.}
      \item{\code{link_create_pl}}{The link creation property list. See \code{\link{H5P_LINK_CREATE}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_delete"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_delete}{}}}
\subsection{\code{H5Group$link_delete()}}{
  This function implements the HDF5-API function H5Ldelete. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_delete(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_delete_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_delete_by_idx}{}}}
\subsection{\code{H5Group$link_delete_by_idx()}}{
  This function implements the HDF5-API function H5Ldelete_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_delete_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_info}{}}}
\subsection{\code{H5Group$link_info()}}{
  This function implements the HDF5-API function H5Lget_info. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_info(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_info_by_idx}{}}}
\subsection{\code{H5Group$link_info_by_idx()}}{
  This function implements the HDF5-API function H5Lget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_info_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_value"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_value}{}}}
\subsection{\code{H5Group$link_value()}}{
  This function implements the HDF5-API function H5Lget_val. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_value(name, link_access_pl = h5const$H5P_DEFAULT)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_value_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_value_by_idx}{}}}
\subsection{\code{H5Group$link_value_by_idx()}}{
  This function implements the HDF5-API function H5Lget_val_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_value_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-link_name_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-link_name_by_idx}{}}}
\subsection{\code{H5Group$link_name_by_idx()}}{
  This function implements the HDF5-API function H5Lget_name_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_l.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$link_name_by_idx(
  n,
  group_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-mount"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-mount}{}}}
\subsection{\code{H5Group$mount()}}{
  This function implements the HDF5-API function H5Fmount. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$mount(name, child)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{child}}{The object to mount; an object of class \code{\link{H5File}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-unmount"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-unmount}{}}}
\subsection{\code{H5Group$unmount()}}{
  This function implements the HDF5-API function H5Funmount. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$unmount(name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-get_obj_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-get_obj_name}{}}}
\subsection{\code{H5Group$get_obj_name()}}{
  This function implements the HDF5-API function H5Iget_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$get_obj_name()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-create_attr"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-create_attr}{}}}
\subsection{\code{H5Group$create_attr()}}{
  This function implements the HDF5-API function H5Acreate2. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$create_attr(attr_name, robj = NULL, dtype = NULL, space = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{robj}}{An R object to use as a template for the datatype and dataspace.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_open"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_open}{}}}
\subsection{\code{H5Group$attr_open()}}{
  This function implements the HDF5-API function H5Aopen. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_open(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-create_attr_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-create_attr_by_name}{}}}
\subsection{\code{H5Group$create_attr_by_name()}}{
  This function implements the HDF5-API function H5Acreate_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$create_attr_by_name(
  attr_name,
  obj_name,
  robj = NULL,
  dtype = NULL,
  space = NULL,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{robj}}{An R object to use as a template for the datatype and dataspace.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_open_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_open_by_name}{}}}
\subsection{\code{H5Group$attr_open_by_name()}}{
  This function implements the HDF5-API function H5Aopen_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_open_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_open_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_open_by_idx}{}}}
\subsection{\code{H5Group$attr_open_by_idx()}}{
  This function implements the HDF5-API function H5Aopen_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_open_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_exists_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_exists_by_name}{}}}
\subsection{\code{H5Group$attr_exists_by_name()}}{
  This function implements the HDF5-API function H5Aexists_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_exists_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_exists"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_exists}{}}}
\subsection{\code{H5Group$attr_exists()}}{
  This function implements the HDF5-API function H5Aexists. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_exists(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_rename_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_rename_by_name}{}}}
\subsection{\code{H5Group$attr_rename_by_name()}}{
  This function implements the HDF5-API function H5Arename_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_rename_by_name(
  old_attr_name,
  new_attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{old_attr_name}}{The current name of the attribute.}
      \item{\code{new_attr_name}}{The new name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_rename"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_rename}{}}}
\subsection{\code{H5Group$attr_rename()}}{
  This function implements the HDF5-API function H5Arename. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_rename(old_attr_name, new_attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{old_attr_name}}{The current name of the attribute.}
      \item{\code{new_attr_name}}{The new name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_delete"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_delete}{}}}
\subsection{\code{H5Group$attr_delete()}}{
  This function implements the HDF5-API function H5Adelete. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_delete(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_delete_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_delete_by_name}{}}}
\subsection{\code{H5Group$attr_delete_by_name()}}{
  This function implements the HDF5-API function H5Adelete_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_delete_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_delete_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_delete_by_idx}{}}}
\subsection{\code{H5Group$attr_delete_by_idx()}}{
  This function implements the HDF5-API function H5Adelete_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_delete_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_info_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_info_by_name}{}}}
\subsection{\code{H5Group$attr_info_by_name()}}{
  This function implements the HDF5-API function H5Aget_info_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_info_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_info_by_idx}{}}}
\subsection{\code{H5Group$attr_info_by_idx()}}{
  This function implements the HDF5-API function H5Aget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_info_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_name_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_name_by_idx}{}}}
\subsection{\code{H5Group$attr_name_by_idx()}}{
  This function implements the HDF5-API function H5Aget_name_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_name_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-attr_get_number"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-attr_get_number}{}}}
\subsection{\code{H5Group$attr_get_number()}}{
  This function implements the HDF5-API function H5Aget_num_attrs. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$attr_get_number()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-create_reference"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-create_reference}{}}}
\subsection{\code{H5Group$create_reference()}}{
  This function implements the HDF5-API function H5Rcreate. If \code{space=NULL} then a \code{H5R_OBJECT} reference is created, otherwise a \code{H5R_DATASET_REGION} reference Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_r.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$create_reference(name = ".", space = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-flush"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-flush}{}}}
\subsection{\code{H5Group$flush()}}{
  This function implements the HDF5-API function H5Fflush. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$flush(scope = h5const$H5F_SCOPE_GLOBAL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{scope}}{The scope of the flush; one of the \code{H5F_SCOPE_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-get_filename"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-get_filename}{}}}
\subsection{\code{H5Group$get_filename()}}{
  This function implements the HDF5-API function H5Fget_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$get_filename()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-print"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-print}{}}}
\subsection{\code{H5Group$print()}}{
  Prints information for the group
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$print(..., max.attributes = 10, max.listing = 10)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
      \item{\code{max.attributes}}{Maximum number of attribute names to print}
      \item{\code{max.listing}}{Maximum number of ls-items to print}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-obj_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-obj_info_by_idx}{}}}
\subsection{\code{H5Group$obj_info_by_idx()}}{
  This function implements the HDF5-API function H5Oget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$obj_info_by_idx(
  n,
  group_name = ".",
  index_field = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  remove_internal_use_only = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{group_name}}{The name of the group, relative to the object the method is called on.}
      \item{\code{index_field}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{remove_internal_use_only}}{Logical; should fields that are only of internal use be dropped from the result.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-obj_info_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-obj_info_by_name}{}}}
\subsection{\code{H5Group$obj_info_by_name()}}{
  This function implements the HDF5-API function H5Oget_info_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$obj_info_by_name(object_name, remove_internal_use_only = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{object_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{remove_internal_use_only}}{Logical; should fields that are only of internal use be dropped from the result.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5Group-obj_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5Group-obj_info}{}}}
\subsection{\code{H5Group$obj_info()}}{
  This function implements the HDF5-API function H5Oget_info. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5Group$obj_info(remove_internal_use_only = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{remove_internal_use_only}}{Logical; should fields that are only of internal use be dropped from the result.}
    }
    \if{html}{\out{</div>}}
  }
}

}
