% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_ATTRIBUTE_CREATE-class}
\alias{H5P_ATTRIBUTE_CREATE-class}
\alias{H5P_ATTRIBUTE_CREATE}
\title{Class for HDF5 property list for attribute creation}
\value{
Object of class \code{\link[=H5P_ATTRIBUTE_CREATE]{H5P_ATTRIBUTE_CREATE}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\section{Methods}{

\describe{
\item{\code{new(id = NULL)}}{

Create a new class of type \code{\link{H5P_ATTRIBUTE_CREATE}}

\strong{Parameters}
\describe{
\item{id}{Internal use only}
}}

\item{\code{set_char_encoding(encoding = h5const$H5T_CSET_UTF8)}}{

This function implements the HDF5-API function H5Pset_char_encoding. Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_SET_CHAR_ENCODING} for details.}

\item{\code{get_char_encoding()}}{

This function implements the HDF5-API function H5Pget_char_encoding. Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_GET_CHAR_ENCODING} for details.}
}}

\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
