% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_DATASET_ACCESS-class}
\alias{H5P_DATASET_ACCESS-class}
\alias{H5P_DATASET_ACCESS}
\title{Class for HDF5 property list for dataset access}
\value{
Object of class \code{\link[=H5P_DATASET_ACCESS]{H5P_DATASET_ACCESS}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\section{Methods}{

\describe{
\item{\code{set_chunk_cache(rdcc_nslots = -1, rdcc_nbytes = -1,
  rdcc_w0 = -1)}}{

This function implements the HDF5-API function H5Pset_chunk_cache.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_SET_CHUNK_CACHE} for details.}

\item{\code{get_chunk_cache()}}{

This function implements the HDF5-API function H5Pget_chunk_cache.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_GET_CHUNK_CACHE} for details.}
}}

\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
