% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_FILE_ACCESS-class}
\alias{H5P_FILE_ACCESS-class}
\alias{H5P_FILE_ACCESS}
\title{Class for HDF5 property list for file creation}
\value{
Object of class \code{\link[=H5P_FILE_ACCESS]{H5P_FILE_ACCESS}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\section{Methods}{

\describe{
\item{\code{new(id = NULL)}}{

Create a new class of type \code{\link{H5P_FILE_ACCESS}}

\strong{Parameters}
\describe{
\item{id}{Internal use only}
}}

\item{\code{set_cache(rdcc_nslots = 521, rdcc_nbytes = 2^20, rdcc_w0 = 0.75)}}{

This function implements the HDF5-API function H5Pset_cache.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_SET_CACHE} for details.}

\item{\code{get_cache()}}{

This function implements the HDF5-API function H5Pget_cache.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/HP5_GET_CACHE} for details.}
}}

\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
