% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_LINK_ACCESS-class}
\alias{H5P_LINK_ACCESS-class}
\alias{H5P_LINK_ACCESS}
\title{Class for HDF5 property list for link access}
\value{
Object of class \code{\link[=H5P_LINK_ACCESS]{H5P_LINK_ACCESS}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\section{Methods}{

\describe{
\item{\code{new(id = NULL)}}{

Create a new class of type \code{\link{H5P_LINK_ACCESS}}

\strong{Parameters}
\describe{
\item{id}{Internal use only}
}}

\item{\code{set_nlinks(nlinks)}}{

This function implements the HDF5-API function H5Pset_nlinks.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_SET_NLINKS} for details.}

\item{\code{get_nlinks()}}{

This function implements the HDF5-API function H5Pget_nlinks.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_GET_NLINKS} for details.}

\item{\code{set_elink_prefix(elink_prefix)}}{

This function implements the HDF5-API function H5Pset_elink_prefix.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_SET_ELINK_PREFIX} for details.}

\item{\code{get_elink_prefix()}}{

This function implements the HDF5-API function H5Pget_elink_prefix.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_GET_ELINK_PREFIX} for details.}

\item{\code{set_elink_acc_flags(elink_acc_flags = h5const$H5F_ACC_RDWR)}}{

This function implements the HDF5-API function H5Pset_elink_acc_flags.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_SET_ELINK_ACC_FLAGS} for details.}

\item{\code{get_elink_acc_flags()}}{

This function implements the HDF5-API function H5Pget_elink_acc_flags.
Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_GET_ELINK_ACC_FLAGS} for details.}
}}

\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
