% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5P.R
\docType{class}
\name{H5P_OBJECT_CREATE-class}
\alias{H5P_OBJECT_CREATE-class}
\alias{H5P_OBJECT_CREATE}
\title{Class for HDF5 property list for object creation}
\value{
Object of class \code{\link[=H5P_OBJECT_CREATE]{H5P_OBJECT_CREATE}}.
}
\description{
It inherits all functions of the \code{\link{H5P}}.
}
\section{Methods}{

\describe{
\item{\code{new(id = NULL)}}{

Create a new class of type \code{\link{H5P_OBJECT_CREATE}}

\strong{Parameters}
\describe{
\item{id}{Internal use only}
}}

\item{\code{set_obj_track_times(track_times = TRUE)}}{

This function implements the HDF5-API function H5Pset_obj_track_times. Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_SET_OBJ_TRACK_TIMES} for details.}

\item{\code{get_obj_track_times()}}{

This function implements the HDF5-API function H5Pget_obj_track_times. Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_GET_OBJ_TRACK_TIMES} for details.}

\item{\code{set_attr_phase_change(max_compact, min_dense)}}{

This function implements the HDF5-API function H5Pset_attr_phase_change. Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_SET_ATTR_PHASE_CHANGE} for details.}

\item{\code{get_attr_phase_change()}}{

This function implements the HDF5-API function H5Pget_attr_phase_change. Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_GET_ATTR_PHASE_CHANGE} for details.}

\item{\code{set_attr_creation_order(crt_order_flags = 0)}}{

This function implements the HDF5-API function H5Pset_attr_creation_order. Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_SET_ATTR_CREATION_ORDER} for details.}

\item{\code{get_attr_creation_order()}}{

This function implements the HDF5-API function H5Pget_attr_creation_order. Please see the documentation at \url{https://portal.hdfgroup.org/display/HDF5/H5P_GET_ATTR_CREATION_ORDER} for details.}
}}

\seealso{
\code{\link[=H5P]{H5P}}
}
\author{
Holger Hoefling
}
