% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5S.R
\docType{class}
\name{H5S-class}
\alias{H5S-class}
\alias{H5S}
\title{Class for representing HDF5 spaces}
\value{
Object of class \code{\link{H5S}}.
}
\description{
This class represents an HDF5 space-id. It inherits all functions of the
\code{\link{H5RefClass-class}}.
}
\section{Methods}{

\describe{
\item{\code{new(type = c("simple", "scalar", "null"), dims = NULL,
  maxdims = dims, decode_buf = NULL, id = NULL)}}{

Create a new HDF5-space. This can be done by either specifying a space with appropriate dimensions or by
decoding a character string that represents an encoded space

\strong{Parameters}
\describe{
\item{type}{Either a \code{simple} space, for which \code{dims} and \code{maxdims} have to be given
or a \code{scalar} or \code{null} space. See the HDF5 user guide on spaces to explain the differences.}
\item{dims}{The dimension of the space in case it is of type \code{simple}}
\item{maxdims}{The maximal dimensions of the space}
\item{decode_buf}{The character string that holds the encoded representation of a space}
\item{id}{An existing HDF5 id; internal use only}
}}

\item{\code{copy()}}{

This function implements the HDF5-API function H5Scopy. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-Copy} for details.}

\item{\code{encode()}}{

This function implements the HDF5-API function H5Sencode. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-Encode} for details.}

\item{\code{is_simple()}}{

This function implements the HDF5-API function H5Sis_simple. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-IsSimple} for details.}

\item{\code{get_simple_extent_ndims()}}{

This function implements the HDF5-API function H5Sget_simple_extent_ndims. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSimpleExtentNdims} for details.}

\item{\code{offset_simple(offset)}}{

This function implements the HDF5-API function H5Soffset_simple. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-OffsetSimple} for details.}

\item{\code{get_simple_extent_dims()}}{

This function implements the HDF5-API function H5Sget_simple_extent_dims. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSimpleExtentDims} for details.}

\item{\code{get_simple_extent_npoints()}}{

This function implements the HDF5-API function H5Sget_simple_extent_npoints. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSimpleExtentNpoints} for details.}

\item{\code{get_simple_extent_type()}}{

This function implements the HDF5-API function H5Sget_simple_extent_type. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSimpleExtentType} for details.}

\item{\code{extent_copy(h5s_source)}}{

This function implements the HDF5-API function H5Sextent_copy. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-ExtentCopy} for details.}

\item{\code{extent_equal(h5s_cmp)}}{

This function implements the HDF5-API function H5Sextent_equal. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-ExtentEqual} for details.}

\item{\code{set_extent_simple(dims, maxdims)}}{

This function implements the HDF5-API function H5Sset_extent_simple. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-SetExtentSimple} for details.}

\item{\code{set_extent_none()}}{

This function implements the HDF5-API function H5Sset_extent_none. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-SetExtentNone} for details.}

\item{\code{get_select_type()}}{

This function implements the HDF5-API function H5Sget_select_type. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSelectType} for details.}

\item{\code{get_select_npoints()}}{

This function implements the HDF5-API function H5Sget_select_npoints. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSelectNpoints} for details.}

\item{\code{get_select_hyper_nblocks()}}{

This function implements the HDF5-API function H5Sget_select_hyper_nblocks. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSelectHyperNblocks} for details.}

\item{\code{get_select_hyper_blocklist(startblock = 0,
  numblocks = (self$get_select_hyper_nblocks() - startblock))}}{

This function implements the HDF5-API function H5Sget_select_hyper_blocklist. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSelectHyperBlocklist} for details.}

\item{\code{get_select_elem_npoints()}}{

This function implements the HDF5-API function H5Sget_select_elem_npoints. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSelectElemNpoints} for details.}

\item{\code{get_select_elem_pointlist(startpoint = 0,
  numpoints = (self$get_select_elem_npoints() - startpoint))}}{

This function implements the HDF5-API function H5Sget_select_elem_pointlist. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSelectElemPointlist} for details.}

\item{\code{get_select_bounds()}}{

This function implements the HDF5-API function H5Sget_select_bounds. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSelectBounds} for details.}

\item{\code{select_all()}}{

This function implements the HDF5-API function H5Sselect_all. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-SelectAll} for details.}

\item{\code{select_none()}}{

This function implements the HDF5-API function H5Sselect_none. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-SelectNone} for details.}

\item{\code{select_valid()}}{

This function implements the HDF5-API function H5Sselect_valid. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-SelectValid} for details.}

\item{\code{select_elements(coord, op = h5const$H5S_SELECT_SET, byrow = TRUE)}}{

This function implements the HDF5-API function H5Sselect_elements. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-SelectElements} for details.}

\item{\code{select_hyperslab(start, count, stride = NULL, block = NULL,
  op = h5const$H5S_SELECT_SET)}}{

This function implements the HDF5-API function H5Sselect_hyperslab. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-SelectHyperslab} for details.}

\item{\code{subset(args, op = h5const$H5S_SELECT_SET, envir = parent.frame())}}{

Subsetting the space. This is mainly intended as a helper function for the '[' function, but
can also be used on its own.

\strong{Parameters}
\describe{
\item{args}{The indices for each dimension to subset given as a list. This makes this easier to use as a programmatic API.
For interactive use we recomment the use of the \code{[} operator.}
\item{op}{The operator to use. Same as for the other HDF5 space selection functions. One of the elements shown in
\code{h5const$H5S_seloper_t}}
\item{envir}{The environment in which to evaluate \code{args}}
}}

\item{\code{print(...)}}{

Prints information for the group

\strong{Parameters}
\describe{
\item{...}{ignored}
}}

\item{\code{dims()}}{

Get the dimensions of the space. Return NULL if the space is not simple (i.e. NULL-space) or a length-0 integer if it is a scalar}

\item{\code{maxdims()}}{

Get the maximal dimensions of the space.Return NULL if the space is not simple (i.e. NULL-space) or a length-0 integer if it is a scalar}

\item{\code{rank()}}{

This function implements the HDF5-API function H5Sget_simple_extent_ndims. Please see the documentation at \url{https://www.hdfgroup.org/HDF5/doc/RM/RM_H5S.html#Dataspace-GetSimpleExtentNdims} for details.}
}}

\author{
Holger Hoefling
}
