% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes_H5T.R
\docType{class}
\name{H5T}
\alias{H5T}
\alias{H5T-class}
\title{Class for HDF5 datatypes.}
\value{
Object of class \code{\link{H5T}}.
}
\description{
This is the base class for all datatypes, but most have a specialised class.
This class represents an HDF5 datatype. It inherits all functions of the
\code{\link{H5RefClass}}. It is also the base class for many other classes well, specifically
\describe{
\item{Integer}{\code{\link{H5T_INTEGER}}}
\item{Bitfield}{\code{\link{H5T_BITFIELD}} (currently identical to the integer class)}
\item{Float}{\code{\link{H5T_FLOAT}}}
\item{Enum}{\code{\link{H5T_ENUM}}}
\item{Compound}{\code{\link{H5T_COMPOUND}}}
\item{String}{\code{\link{H5T_STRING}}}
\item{Complex}{\code{\link{H5T_COMPLEX}}}
\item{Array}{\code{\link{H5T_ARRAY}}}
\item{Variable Length}{\code{\link{H5T_VLEN}}}
}
}
\examples{
my_int <- h5types$H5T_NATIVE_INT
my_int$to_text()
my_int$get_size()

# Show how to commit a datatype
fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
my_int$is_committed()
file$commit("my_int", my_int)
my_int$is_committed()

# can now also add attributes
h5attr(my_int, "test") <- "A string"
h5attributes(my_int)

file$close_all()
file.remove(fname)
}
\author{
Holger Hoefling
}
\section{Super class}{
\code{\link[hdf5r:H5RefClass]{H5RefClass}} -> \code{H5T}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-H5T-get_obj_name}{\code{H5T$get_obj_name()}}
    \item \href{#method-H5T-create_attr}{\code{H5T$create_attr()}}
    \item \href{#method-H5T-attr_open}{\code{H5T$attr_open()}}
    \item \href{#method-H5T-create_attr_by_name}{\code{H5T$create_attr_by_name()}}
    \item \href{#method-H5T-attr_open_by_name}{\code{H5T$attr_open_by_name()}}
    \item \href{#method-H5T-attr_open_by_idx}{\code{H5T$attr_open_by_idx()}}
    \item \href{#method-H5T-attr_exists_by_name}{\code{H5T$attr_exists_by_name()}}
    \item \href{#method-H5T-attr_exists}{\code{H5T$attr_exists()}}
    \item \href{#method-H5T-attr_rename_by_name}{\code{H5T$attr_rename_by_name()}}
    \item \href{#method-H5T-attr_rename}{\code{H5T$attr_rename()}}
    \item \href{#method-H5T-attr_delete}{\code{H5T$attr_delete()}}
    \item \href{#method-H5T-attr_delete_by_name}{\code{H5T$attr_delete_by_name()}}
    \item \href{#method-H5T-attr_delete_by_idx}{\code{H5T$attr_delete_by_idx()}}
    \item \href{#method-H5T-attr_info_by_name}{\code{H5T$attr_info_by_name()}}
    \item \href{#method-H5T-attr_info_by_idx}{\code{H5T$attr_info_by_idx()}}
    \item \href{#method-H5T-attr_name_by_idx}{\code{H5T$attr_name_by_idx()}}
    \item \href{#method-H5T-attr_get_number}{\code{H5T$attr_get_number()}}
    \item \href{#method-H5T-create_reference}{\code{H5T$create_reference()}}
    \item \href{#method-H5T-initialize}{\code{H5T$new()}}
    \item \href{#method-H5T-get_class}{\code{H5T$get_class()}}
    \item \href{#method-H5T-get_size}{\code{H5T$get_size()}}
    \item \href{#method-H5T-set_size}{\code{H5T$set_size()}}
    \item \href{#method-H5T-set_precision}{\code{H5T$set_precision()}}
    \item \href{#method-H5T-get_precision}{\code{H5T$get_precision()}}
    \item \href{#method-H5T-set_order}{\code{H5T$set_order()}}
    \item \href{#method-H5T-get_order}{\code{H5T$get_order()}}
    \item \href{#method-H5T-set_offset}{\code{H5T$set_offset()}}
    \item \href{#method-H5T-get_offset}{\code{H5T$get_offset()}}
    \item \href{#method-H5T-set_pad}{\code{H5T$set_pad()}}
    \item \href{#method-H5T-get_pad}{\code{H5T$get_pad()}}
    \item \href{#method-H5T-copy}{\code{H5T$copy()}}
    \item \href{#method-H5T-is_committed}{\code{H5T$is_committed()}}
    \item \href{#method-H5T-equal}{\code{H5T$equal()}}
    \item \href{#method-H5T-is_vlen}{\code{H5T$is_vlen()}}
    \item \href{#method-H5T-detect_class}{\code{H5T$detect_class()}}
    \item \href{#method-H5T-get_native_type}{\code{H5T$get_native_type()}}
    \item \href{#method-H5T-get_create_plist}{\code{H5T$get_create_plist()}}
    \item \href{#method-H5T-to_text}{\code{H5T$to_text()}}
    \item \href{#method-H5T-print}{\code{H5T$print()}}
    \item \href{#method-H5T-obj_info}{\code{H5T$obj_info()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>H5RefClass$close()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>H5RefClass$dec_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>H5RefClass$get_file_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>H5RefClass$get_obj_type()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>H5RefClass$get_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>H5RefClass$inc_ref()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>H5RefClass$methods()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-get_obj_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-get_obj_name}{}}}
\subsection{\code{H5T$get_obj_name()}}{
  This function implements the HDF5-API function H5Iget_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$get_obj_name()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-create_attr"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-create_attr}{}}}
\subsection{\code{H5T$create_attr()}}{
  This function implements the HDF5-API function H5Acreate2. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$create_attr(attr_name, robj = NULL, dtype = NULL, space = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{robj}}{An R object to use as a template for the datatype and dataspace.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_open"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_open}{}}}
\subsection{\code{H5T$attr_open()}}{
  This function implements the HDF5-API function H5Aopen. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_open(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-create_attr_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-create_attr_by_name}{}}}
\subsection{\code{H5T$create_attr_by_name()}}{
  This function implements the HDF5-API function H5Acreate_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$create_attr_by_name(
  attr_name,
  obj_name,
  robj = NULL,
  dtype = NULL,
  space = NULL,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{robj}}{An R object to use as a template for the datatype and dataspace.}
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_open_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_open_by_name}{}}}
\subsection{\code{H5T$attr_open_by_name()}}{
  This function implements the HDF5-API function H5Aopen_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_open_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_open_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_open_by_idx}{}}}
\subsection{\code{H5T$attr_open_by_idx()}}{
  This function implements the HDF5-API function H5Aopen_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_open_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_exists_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_exists_by_name}{}}}
\subsection{\code{H5T$attr_exists_by_name()}}{
  This function implements the HDF5-API function H5Aexists_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_exists_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_exists"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_exists}{}}}
\subsection{\code{H5T$attr_exists()}}{
  This function implements the HDF5-API function H5Aexists. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_exists(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_rename_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_rename_by_name}{}}}
\subsection{\code{H5T$attr_rename_by_name()}}{
  This function implements the HDF5-API function H5Arename_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_rename_by_name(
  old_attr_name,
  new_attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{old_attr_name}}{The current name of the attribute.}
      \item{\code{new_attr_name}}{The new name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_rename"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_rename}{}}}
\subsection{\code{H5T$attr_rename()}}{
  This function implements the HDF5-API function H5Arename. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_rename(old_attr_name, new_attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{old_attr_name}}{The current name of the attribute.}
      \item{\code{new_attr_name}}{The new name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_delete"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_delete}{}}}
\subsection{\code{H5T$attr_delete()}}{
  This function implements the HDF5-API function H5Adelete. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_delete(attr_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_delete_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_delete_by_name}{}}}
\subsection{\code{H5T$attr_delete_by_name()}}{
  This function implements the HDF5-API function H5Adelete_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_delete_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_delete_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_delete_by_idx}{}}}
\subsection{\code{H5T$attr_delete_by_idx()}}{
  This function implements the HDF5-API function H5Adelete_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_delete_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_info_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_info_by_name}{}}}
\subsection{\code{H5T$attr_info_by_name()}}{
  This function implements the HDF5-API function H5Aget_info_by_name. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_info_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{attr_name}}{The name of the attribute.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_info_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_info_by_idx}{}}}
\subsection{\code{H5T$attr_info_by_idx()}}{
  This function implements the HDF5-API function H5Aget_info_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_info_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_name_by_idx"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_name_by_idx}{}}}
\subsection{\code{H5T$attr_name_by_idx()}}{
  This function implements the HDF5-API function H5Aget_name_by_idx. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_name_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{The index of the item to access; zero-based.}
      \item{\code{obj_name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{idx_type}}{The index to iterate over; one of the \code{H5_INDEX_*} values in \code{\link{h5const}}.}
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
      \item{\code{link_access_pl}}{The link access property list. See \code{\link{H5P_LINK_ACCESS}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-attr_get_number"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-attr_get_number}{}}}
\subsection{\code{H5T$attr_get_number()}}{
  This function implements the HDF5-API function H5Aget_num_attrs. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$attr_get_number()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-create_reference"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-create_reference}{}}}
\subsection{\code{H5T$create_reference()}}{
  This function implements the HDF5-API function H5Rcreate. If \code{space=NULL} then a \code{H5R_OBJECT} reference is created, otherwise a \code{H5R_DATASET_REGION} reference Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_r.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$create_reference(name = ".", space = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of the object, relative to the object the method is called on.}
      \item{\code{space}}{The dataspace to use; an object of class \code{\link{H5S}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-initialize}{}}}
\subsection{\code{H5T$new()}}{
  Internal use only
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$new(id)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{An HDF5 id; for internal use only.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-get_class"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-get_class}{}}}
\subsection{\code{H5T$get_class()}}{
  This function implements the HDF5-API function H5Tget_class. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$get_class()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-get_size"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-get_size}{}}}
\subsection{\code{H5T$get_size()}}{
  This function implements the HDF5-API function H5Tget_size. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$get_size(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-set_size"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-set_size}{}}}
\subsection{\code{H5T$set_size()}}{
  This function implements the HDF5-API function H5Tset_size. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$set_size(size)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{size}}{The size of the object in bytes; use \code{Inf} for variable-length strings.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-set_precision"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-set_precision}{}}}
\subsection{\code{H5T$set_precision()}}{
  This function implements the HDF5-API function H5Tset_precision. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$set_precision(precision)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{precision}}{The precision of the datatype, in bits.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-get_precision"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-get_precision}{}}}
\subsection{\code{H5T$get_precision()}}{
  This function implements the HDF5-API function H5Tget_precision. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$get_precision()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-set_order"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-set_order}{}}}
\subsection{\code{H5T$set_order()}}{
  This function implements the HDF5-API function H5Tset_order. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$set_order(order)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{order}}{The order in which to iterate; one of the \code{H5_ITER_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-get_order"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-get_order}{}}}
\subsection{\code{H5T$get_order()}}{
  This function implements the HDF5-API function H5Tget_order. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$get_order()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-set_offset"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-set_offset}{}}}
\subsection{\code{H5T$set_offset()}}{
  This function implements the HDF5-API function H5Tset_offset. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$set_offset(offset)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{offset}}{The offset to apply; see the corresponding HDF5 function for how it is interpreted.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-get_offset"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-get_offset}{}}}
\subsection{\code{H5T$get_offset()}}{
  This function implements the HDF5-API function H5Tget_offset. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$get_offset()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-set_pad"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-set_pad}{}}}
\subsection{\code{H5T$set_pad()}}{
  This function implements the HDF5-API function H5Tset_pad. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$set_pad(pad)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pad}}{The padding to use for unused bits; one of the \code{H5T_PAD_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-get_pad"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-get_pad}{}}}
\subsection{\code{H5T$get_pad()}}{
  This function implements the HDF5-API function H5Tget_pad. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$get_pad()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-copy"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-copy}{}}}
\subsection{\code{H5T$copy()}}{
  This function implements the HDF5-API function H5Tcopy. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$copy()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-is_committed"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-is_committed}{}}}
\subsection{\code{H5T$is_committed()}}{
  This function implements the HDF5-API function H5Tcommitted. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$is_committed()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-equal"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-equal}{}}}
\subsection{\code{H5T$equal()}}{
  This function implements the HDF5-API function H5Tequal. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$equal(dtype)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dtype}}{The datatype to use; an object of class \code{\link{H5T}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-is_vlen"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-is_vlen}{}}}
\subsection{\code{H5T$is_vlen()}}{
  This function detects if the underlying type is H5T_VLEN or a variable length string. This is used to know if after reading a dataset, memory has to be freed
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$is_vlen()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-detect_class"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-detect_class}{}}}
\subsection{\code{H5T$detect_class()}}{
  This function implements the HDF5-API function H5Tdetect_class. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$detect_class(dtype_class)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dtype_class}}{The datatype class to look for; one of the \code{H5T_*} class values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-get_native_type"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-get_native_type}{}}}
\subsection{\code{H5T$get_native_type()}}{
  This function implements the HDF5-API function H5Tget_native_type. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$get_native_type(direction = h5const$H5T_DIR_ASCEND)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{direction}}{The direction in which to search for a native type; one of the \code{H5T_DIR_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-get_create_plist"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-get_create_plist}{}}}
\subsection{\code{H5T$get_create_plist()}}{
  This function implements the HDF5-API function H5Tget_create_plist. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$get_create_plist()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-to_text"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-to_text}{}}}
\subsection{\code{H5T$to_text()}}{
  This function implements the HDF5-API function H5LTdtype_to_text.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$to_text(lang_type = h5const$H5LT_DDL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{lang_type}}{The language to produce the text representation in; one of the \code{H5LT_LANG_*} values in \code{\link{h5const}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-print"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-print}{}}}
\subsection{\code{H5T$print()}}{
  Prints information for the group
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H5T-obj_info"></a>}}
\if{latex}{\out{\hypertarget{method-H5T-obj_info}{}}}
\subsection{\code{H5T$obj_info()}}{
  This function implements the HDF5-API function H5Oget_info. Please see the documentation at \url{https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html} for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{H5T$obj_info(remove_internal_use_only = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{remove_internal_use_only}}{Logical; should fields that are only of internal use be dropped from the result.}
    }
    \if{html}{\out{</div>}}
  }
}

}
