% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5r.R
\docType{package}
\name{hdf5r-package}
\alias{hdf5r-package}
\title{hdf5r: A package to provide an interface to hdf5 from R}
\description{
A package that allows to interface with the HDF5 C-library. Provides access to most of its
functionality from inside R using R6 classes. For more details please see the README at
the github page \url{https://github.com/hhoeflin/hdf5r}.
}
