% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5r.R
\docType{package}
\name{hdf5r}
\alias{hdf5r}
\alias{hdf5r-package}
\title{hdf5r: A package to provide an interface to hdf5 from R}
\description{
hdf5r: A package to provide an interface to hdf5 from R
}
\section{Overview of classes}{

}

