% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5wrapper.R
\name{list-groups-datasets}
\alias{list-groups-datasets}
\alias{list.groups}
\alias{list.datasets}
\title{List Groups and Datasets in object}
\usage{
list.groups(object, path = "/", recursive = TRUE, ...)

list.datasets(object, path = "/", recursive = TRUE, ...)
}
\arguments{
\item{object}{\code{CommonFG}; Object implementing the CommonFG Interface (e.g. \code{\link{H5File}}, \code{\link{H5Group}}).}

\item{path}{character; Path named to be used for iteration.}

\item{recursive}{logical; Specify if object should be traversed recursively.}

\item{...}{Additional Parameters passed to \code{$ls()}}
}
\value{
\code{\link{character}}
}
\description{
List all Group (\code{\link{H5Group}}) and Dataset (\code{\link{H5D}}) 
names in the current object. This function is part of the \strong{h5} wrapper classes and
uses \code{$ls()} to retrieve group names.
}
